Ouroboros.Consensus.Shelley.Eras

Eras based on the Shelley ledger

type AllegraEra = ShelleyMAEra 'Allegra

data AlonzoEra c

The Alonzo era

data BabbageEra c

The Babbage era

type MaryEra c = ShelleyMAEra 'Mary c

data ShelleyEra c

Eras instantiated with standard crypto

type StandardAllegra = AllegraEra StandardCrypto

The Allegra era with standard crypto

type StandardAlonzo = AlonzoEra StandardCrypto

The Alonzo era with standard crypto

type StandardBabbage = BabbageEra StandardCrypto

The Babbage era with standard crypto

type StandardMary = MaryEra StandardCrypto

The Mary era with standard crypto

type StandardShelley = ShelleyEra StandardCrypto

The Shelley era with standard crypto

Shelley-based era

class (ShelleyBasedEra era, GetLedgerView era, State (EraRule "PPUP" era) ~ PPUPState era, Default (State (EraRule "PPUP" era)), HasField "_maxBBSize" (PParams era) Natural, HasField "_maxBHSize" (PParams era) Natural, HasField "_maxTxSize" (PParams era) Natural, HasField "_a0" (PParams era) NonNegativeInterval, HasField "_nOpt" (PParams era) Natural, HasField "_rho" (PParams era) UnitInterval, HasField "_tau" (PParams era) UnitInterval, ValidateScript era, FromCBOR (PParams era), ToCBOR (PParams era), HasField "_protocolVersion" (PParamsDelta era) (StrictMaybe ProtVer), FromCBOR (PParamsDelta era), AdditionalGenesisConfig era ~ TranslationContext era, ToCBORGroup (TxSeq era), NoThunks (TranslationContext era), FromCBOR (Annotator (Witnesses era)), ToCBOR (Witnesses era), Eq (TxSeq era), Show (TxSeq era), FromCBOR (Annotator (TxSeq era))) => ShelleyBasedEra era where

The ledger already defines ShelleyBasedEra as the top-level constraint on an era, however, consensus often needs some more functionality than the ledger currently provides.

Either the functionality shouldn't or can't live in the ledger, in which case it can be part and remain part of ShelleyBasedEra. Or, the functionality should live in the ledger, but hasn't yet been added to the ledger, or it hasn't yet been propagated to this repository, in which case it can be added to this class until that is the case.

By having the same name as the class defined in ledger, we can, if this class becomes redundant, switch to the ledger-defined one without having to update all the code using it. We can just export the right one from this module.

TODO Currently we include some constraints on the update state which are needed to determine the hard fork point. In the future this should be replaced with an appropriate API - see https://github.com/input-output-hk/ouroboros-network/issues/2890

Methods

shelleyBasedEraName :: proxy era -> Text

Return the name of the Shelley-based era, e.g., Shelley, Allegra, etc.

applyShelleyBasedTx :: Globals -> LedgerEnv era -> MempoolState era -> WhetherToIntervene -> Tx era -> Except (ApplyTxError era) (MempoolState era, Validated (Tx era))

newtype WrapTx era

Wrapper for partially applying the Tx type family

For generality, Consensus uses that type family as eg the index of TranslateEra. We thus need to partially apply it.

cardano-ledger-specs also declares such a newtype, but currently it's only defined in the Alonzo translation module, which seems somewhat inappropriate to use for previous eras. Also, we use a Wrap prefix in Consensus. Hence this minor mediating definition. TODO I'm not even fully persuading myself with this justification.

Constructors

WrapTx

Fields

unwrapTx :: Tx era

Type synonyms for convenience

type EraCrypto era = Crypto era

The Crypto type family conflicts with the Crypto class. To avoid having to import one or both of them qualified, define EraCrypto as an alias of the former: /return the crypto used by this era/.

Re-exports

data StandardCrypto

The same crypto used on the net

Exceptions

data UnexpectedAlonzoLedgerErrors

The ledger responded with Alonzo errors we were not expecting