Ouroboros.Consensus.Shelley.Genesis

Documentation

data ShelleyGenesisStaking c

Genesis Shelley staking configuration.

This allows us to configure some initial stake pools and delegation to them, in order to test Praos in a static configuration, without requiring on-chain registration and delegation.

For simplicity, pools defined in the genesis staking do not pay deposits for their registration.

Constructors

ShelleyGenesisStaking

Fields

sgsPools ∷ !(Map (KeyHash StakePool c) (PoolParams c))

Pools to register

The key in this map is the hash of the public key of the _pool_. This need not correspond to any payment or staking key, but must correspond to the cold key held by TPraosIsCoreNode.

sgsStake ∷ !(Map (KeyHash Staking c) (KeyHash StakePool c))

Stake-holding key hash credentials and the pools to delegate that stake to. We require the raw staking key hash in order to:

Avoid pointer addresses, which would be tricky when there's no slot or transaction to point to.

Avoid script credentials.

emptyGenesisStaking ∷ ShelleyGenesisStaking c

Empty genesis staking

data ShelleyGenesis c

Shelley genesis information

Note that this is needed only for a pure Shelley network, hence it being defined here rather than in its own module. In mainnet, Shelley will transition naturally from Byron, and thus will never have its own genesis information.

Constructors

ShelleyGenesis

Fields

sgStartTime ∷ !SystemStart

sgNetworkMagic ∷ !NetworkMagic

sgProtocolMagicId ∷ !ProtocolMagicId

sgActiveSlotsCoeff ∷ !Double

sgSecurityParam ∷ !SecurityParam

sgEpochLength ∷ !EpochSize

sgSlotsPerKESPeriod ∷ !Word64

sgMaxKESEvolutions ∷ !Word64

sgSlotLength ∷ !SlotLength

sgUpdateQuorum ∷ !Word64

sgMaxMajorPV ∷ !Natural

sgMaxLovelaceSupply ∷ !Word64

sgProtocolParams ∷ !PParams

sgGenDelegs ∷ !(Map (KeyHash Genesis c) (KeyHash GenesisDelegate c))

sgInitialFunds ∷ !(Map (Addr c) Coin)

sgStaking ∷ !(ShelleyGenesisStaking c)

sgActiveSlotCoeff ∷ ShelleyGenesis c → ActiveSlotCoeff