Ouroboros.Consensus.Shelley.Ledger

Documentation

data family Query blk ∷ Type → Type

Different queries supported by the ledger, indexed by the result type.

data family GenTx blk

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data family TxId tx

A generalized transaction, GenTx, identifier.

data family LedgerState blk

Ledger state associated with a block

data family BlockConfig blk

Static configuration required to work with this type of blocks

data family CodecConfig blk

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data family StorageConfig blk

Config needed for the NodeInitStorage class. Defined here to avoid circular dependencies.

data family Header blk

class HasHeader (Header blk) ⇒ GetHeader blk where

Methods

getHeader ∷ blk → Header blk

blockMatchesHeader ∷ Header blk → blk → Bool

Check whether the header is the header of the block.

For example, by checking whether the hash of the body stored in the header matches that of the block.

headerIsEBB ∷ Header blk → Maybe EpochNo

When the given header is the header of an Epoch Boundary Block, returns its epoch number.

data family NestedCtxt_ blk ∷ (Type → Type) → Type → Type

Context identifying what kind of block we have

In almost all places we will use NestedCtxt rather than NestedCtxt_.

data family Ticked st

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

data ApplyTxError era

Constructors

ApplyTxError [PredicateFailure (EraRule "LEDGERS" era)]

class (ShelleyBasedEra era, State (EraRule "PPUP" era) ~ PPUPState era, Default (State (EraRule "PPUP" era)), HasField "_maxBHSize" (PParams era) Natural, HasField "_maxTxSize" (PParams era) Natural, HasField "_a0" (PParams era) Rational, HasField "_nOpt" (PParams era) Natural, HasField "_rho" (PParams era) UnitInterval, HasField "_tau" (PParams era) UnitInterval, HasField "_protocolVersion" (PParamsDelta era) (StrictMaybe ProtVer), AdditionalGenesisConfig era ~ (), FromCBOR (PParams era), FromCBOR (PParamsDelta era), ToCBOR (PParams era), Witnesses era ~ WitnessSet era) ⇒ ShelleyBasedEra era

The ledger already defines ShelleyBasedEra as the top-level constraint on an era, however, consensus often needs some more functionality than the ledger currently provides.

Either the functionality shouldn't or can't live in the ledger, in which case it can be part and remain part of ShelleyBasedEra. Or, the functionality should live in the ledger, but hasn't yet been added to the ledger, or it hasn't yet been propagated to this repository, in which case it can be added to this class until that is the case.

By having the same name as the class defined in ledger, we can, if this class becomes redundant, switch to the ledger-defined one without having to update all the code using it. We can just export the right one from this module.

TODO Currently we include some constraints on the update state which are needed to determine the hard fork point. In the future this should be replaced with an appropriate API - see https://github.com/input-output-hk/ouroboros-network/issues/2890

TODO Currently we include the constraint SL.AdditionalGenesisConfig era ~ (). When we fork to Alonzo we will need additional genesis config information.

Minimal complete definition

shelleyBasedEraName

data ShelleyBlock era

Newtype wrapper to avoid orphan instances

The phantom type parameter is there to record the additional information we need to work with this block. Most of the code here does not care, but we may need different additional information when running the chain.

Constructors

ShelleyBlock

Fields

shelleyBlockRaw ∷ !(Block era)

shelleyBlockHeaderHash ∷ !(ShelleyHash (EraCrypto era))

newtype ShelleyHash c

Constructors

ShelleyHash

Fields

unShelleyHash ∷ HashHeader c

mkShelleyBlock ∷ ShelleyBasedEra era ⇒ Block era → ShelleyBlock era

mkShelleyHeader ∷ ShelleyBasedEra era ⇒ BHeader (EraCrypto era) → Header (ShelleyBlock era)

fromShelleyPrevHash ∷ PrevHash (EraCrypto era) → ChainHash (ShelleyBlock era)

From cardano-ledger-specs to ouroboros-consensus

toShelleyPrevHash ∷ ChainHash (Header (ShelleyBlock era)) → PrevHash (EraCrypto era)

From ouroboros-consensus to cardano-ledger-specs

encodeShelleyBlock ∷ ShelleyBasedEra era ⇒ ShelleyBlock era → Encoding

decodeShelleyBlock ∷ ShelleyBasedEra era ⇒ Decoder s (ByteString → ShelleyBlock era)

shelleyBinaryBlockInfo ∷ ShelleyBasedEra era ⇒ ShelleyBlock era → BinaryBlockInfo

encodeShelleyHeader ∷ ShelleyBasedEra era ⇒ Header (ShelleyBlock era) → Encoding

decodeShelleyHeader ∷ ShelleyBasedEra era ⇒ Decoder s (ByteString → Header (ShelleyBlock era))

data ShelleyNodeToClientVersion

Constructors

ShelleyNodeToClientVersion1

ShelleyNodeToClientVersion2
New queries introduced

ShelleyNodeToClientVersion3
New query introduced

data ShelleyNodeToNodeVersion

Constructors

ShelleyNodeToNodeVersion1

verifyHeaderIntegrity

Arguments

∷ ShelleyBasedEra era

⇒ Word64
tpraosSlotsPerKESPeriod

→ Header (ShelleyBlock era)

→ Bool

Verify whether a header is not corrupted

verifyBlockIntegrity

Arguments

∷ ShelleyBasedEra era

⇒ Word64
tpraosSlotsPerKESPeriod

→ ShelleyBlock era

→ Bool

Verifies whether the block is not corrupted by checking its signature and witnesses

data CompactGenesis era

Compact variant of ShelleyGenesis with some fields erased that are only used on start-up and that should not be kept in memory forever.

Concretely:

The sgInitialFunds field is erased. It is only used to set up the initial UTxO in tests and testnets.

The sgStaking field is erased. It is only used to register initial stake pools in tests and benchmarks.

mkShelleyBlockConfig ∷ ShelleyBasedEra era ⇒ ProtVer → ShelleyGenesis era → [VKey 'BlockIssuer (EraCrypto era)] → BlockConfig (ShelleyBlock era)

compactGenesis ∷ ShelleyGenesis era → CompactGenesis era

Compacts the given ShelleyGenesis.

newtype ShelleyTransition

Information required to determine the hard fork point from Shelley to the next ledger

Constructors

ShelleyTransitionInfo

Fields

shelleyAfterVoting ∷ Word32

The number of blocks in this epoch past the voting deadline

We record this to make sure that we can tell the HFC about hard forks if and only if we are certain:

Blocks that came in within an epoch after the 4k/f voting deadline are not relevant (10kf - 2 * 3kf).

Since there are slots between blocks, we are probably only sure that there will be no more relevant block when we have seen the first block after the deadline.

If we count how many blocks we have seen post deadline, and we have reached k of them, we know that that last pre-deadline block won't be rolled back anymore.

At this point we can look at the ledger state and see which proposals we accepted in the voting period, if any, and notify the HFC is one of them indicates a transition.

data ShelleyTip era

Constructors

ShelleyTip

Fields

shelleyTipSlotNo ∷ !SlotNo

shelleyTipBlockNo ∷ !BlockNo

shelleyTipHash ∷ !(HeaderHash (ShelleyBlock era))

data ShelleyLedgerConfig era

Constructors

ShelleyLedgerConfig

Fields

shelleyLedgerCompactGenesis ∷ !(CompactGenesis era)

shelleyLedgerGlobals ∷ !Globals

Derived from shelleyLedgerGenesis but we store a cached version because it used very often.

newtype ShelleyLedgerError era

Constructors

BBodyError (BlockTransitionError era)

shelleyLedgerGenesis ∷ ShelleyLedgerConfig era → ShelleyGenesis era

shelleyEraParams ∷ ShelleyGenesis era → EraParams

shelleyEraParamsNeverHardForks ∷ ShelleyGenesis era → EraParams

Separate variant of shelleyEraParams to be used for a Shelley-only chain.

mkShelleyLedgerConfig ∷ ShelleyGenesis era → EpochInfo Identity → MaxMajorProtVer → ShelleyLedgerConfig era

shelleyTipToPoint ∷ WithOrigin (ShelleyTip era) → Point (ShelleyBlock era)

shelleyLedgerTipPoint ∷ LedgerState (ShelleyBlock era) → Point (ShelleyBlock era)

getPParams ∷ NewEpochState era → PParams era

encodeShelleyAnnTip ∷ ShelleyBasedEra era ⇒ AnnTip (ShelleyBlock era) → Encoding

decodeShelleyAnnTip ∷ ShelleyBasedEra era ⇒ Decoder s (AnnTip (ShelleyBlock era))

encodeShelleyHeaderState ∷ ShelleyBasedEra era ⇒ HeaderState (ShelleyBlock era) → Encoding

encodeShelleyLedgerState ∷ ShelleyBasedEra era ⇒ LedgerState (ShelleyBlock era) → Encoding

decodeShelleyLedgerState ∷ ∀ era s. ShelleyBasedEra era ⇒ Decoder s (LedgerState (ShelleyBlock era))

newtype NonMyopicMemberRewards c

Constructors

NonMyopicMemberRewards

Fields

unNonMyopicMemberRewards ∷ Map (Either Coin (Credential 'Staking c)) (Map (KeyHash 'StakePool c) Coin)

querySupportedVersion ∷ Query (ShelleyBlock era) result → ShelleyNodeToClientVersion → Bool

Is the given query supported by the given ShelleyNodeToClientVersion?

encodeShelleyQuery ∷ ShelleyBasedEra era ⇒ Query (ShelleyBlock era) result → Encoding

decodeShelleyQuery ∷ ShelleyBasedEra era ⇒ Decoder s (SomeSecond Query (ShelleyBlock era))

encodeShelleyResult ∷ ShelleyBasedEra era ⇒ Query (ShelleyBlock era) result → result → Encoding

decodeShelleyResult ∷ ShelleyBasedEra era ⇒ Query (ShelleyBlock era) result → ∀ s. Decoder s result

fixedBlockBodyOverhead ∷ Num a ⇒ a

txInBlockSize is used to estimate how many transactions we can grab from the Mempool to put into the block we are going to forge without exceeding the maximum block body size according to the ledger. If we exceed that limit, we will have forged a block that is invalid according to the ledger. We ourselves won't even adopt it, causing us to lose our slot, something we must try to avoid.

For this reason it is better to overestimate the size of a transaction than to underestimate. The only downside is that we maybe could have put one (or more?) transactions extra in that block.

As the sum of the serialised transaction sizes is not equal to the size of the serialised block body (TxSeq) consisting of those transactions (see cardano-node#1545 for an example), we account for some extra overhead per transaction as a safety margin.

Also see perTxOverhead.

perTxOverhead ∷ Num a ⇒ a

See fixedBlockBodyOverhead.

mkShelleyTx ∷ ∀ era. ShelleyBasedEra era ⇒ Tx era → GenTx (ShelleyBlock era)

forgeShelleyBlock

Arguments

∷ ∀ m era. (ShelleyBasedEra era, Monad m)

⇒ HotKey (EraCrypto era) m

→ TPraosCanBeLeader (EraCrypto era)

→ TopLevelConfig (ShelleyBlock era)

→ BlockNo
Current block number

→ SlotNo
Current slot number

→ TickedLedgerState (ShelleyBlock era)
Current ledger

→ [GenTx (ShelleyBlock era)]
Txs to add in the block

→ TPraosIsLeader (EraCrypto era)
Leader proof

→ m (ShelleyBlock era)