Ouroboros.Consensus.Shelley.Ledger.Block

Documentation

newtype ShelleyHash era

Constructors

ShelleyHash

Fields

unShelleyHash ∷ HashHeader era

data ShelleyBlock era

Newtype wrapper to avoid orphan instances

The phantom type parameter is there to record the additional information we need to work with this block. Most of the code here does not care, but we may need different additional information when running the chain.

Constructors

ShelleyBlock

Fields

shelleyBlockRaw ∷ !(Block era)

shelleyBlockHeaderHash ∷ !(ShelleyHash era)

mkShelleyBlock ∷ Era era ⇒ Block era → ShelleyBlock era

class HasHeader (Header blk) ⇒ GetHeader blk where

Methods

getHeader ∷ blk → Header blk

blockMatchesHeader ∷ Header blk → blk → Bool

Check whether the header is the header of the block.

For example, by checking whether the hash of the body stored in the header matches that of the block.

headerIsEBB ∷ Header blk → Maybe EpochNo

When the given header is the header of an Epoch Boundary Block, returns its epoch number.

data family Header blk ∷ Type

mkShelleyHeader ∷ Era era ⇒ BHeader era → Header (ShelleyBlock era)

data family NestedCtxt_ blk (a ∷ Type → Type) b ∷ Type

Context identifying what kind of block we have

In almost all places we will use NestedCtxt rather than NestedCtxt_.

Serialisation

encodeShelleyBlock ∷ Era era ⇒ ShelleyBlock era → Encoding

decodeShelleyBlock ∷ Era era ⇒ Decoder s (ByteString → ShelleyBlock era)

shelleyBinaryBlockInfo ∷ Era era ⇒ ShelleyBlock era → BinaryBlockInfo

encodeShelleyHeader ∷ Era era ⇒ Header (ShelleyBlock era) → Encoding

decodeShelleyHeader ∷ Era era ⇒ Decoder s (ByteString → Header (ShelleyBlock era))

Conversion

fromShelleyPrevHash ∷ PrevHash era → ChainHash (ShelleyBlock era)

From cardano-ledger-specs to ouroboros-consensus

toShelleyPrevHash ∷ ChainHash (Header (ShelleyBlock era)) → PrevHash era

From ouroboros-consensus to cardano-ledger-specs

Re-exported

class (Crypto (Crypto e), Typeable e) ⇒ Era e