Ouroboros.Consensus.Shelley.Ledger.Config

Documentation

data family BlockConfig blk

Static configuration required to work with this type of blocks

data family CodecConfig blk

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data family StorageConfig blk

Config needed for the NodeInitStorage class. Defined here to avoid circular dependencies.

compactGenesis :: ShelleyGenesis era -> CompactGenesis era

Compacts the given ShelleyGenesis.

getCompactGenesis :: CompactGenesis era -> ShelleyGenesis era

mkShelleyBlockConfig :: ShelleyBasedEra era => ProtVer -> ShelleyGenesis era -> [VKey 'BlockIssuer (EraCrypto era)] -> BlockConfig (ShelleyBlock proto era)

opaque

data CompactGenesis era

Compact variant of ShelleyGenesis with some fields erased that are only used on start-up and that should not be kept in memory forever.

Concretely:

The sgInitialFunds field is erased. It is only used to set up the initial UTxO in tests and testnets.

The sgStaking field is erased. It is only used to register initial stake pools in tests and benchmarks.

Orphan instances

ShelleyBasedEra era => Show (BlockConfig (ShelleyBlock proto era))

Instance details

Methods

showsPrec :: Int -> BlockConfig (ShelleyBlock proto era) -> ShowS

show :: BlockConfig (ShelleyBlock proto era) -> String

showList :: [BlockConfig (ShelleyBlock proto era)] -> ShowS

Generic (BlockConfig (ShelleyBlock proto era))

Instance details

Associated Types

type Rep (BlockConfig (ShelleyBlock proto era)) :: Type -> Type

Methods

from :: BlockConfig (ShelleyBlock proto era) -> Rep (BlockConfig (ShelleyBlock proto era)) x

to :: Rep (BlockConfig (ShelleyBlock proto era)) x -> BlockConfig (ShelleyBlock proto era)

Generic (CodecConfig (ShelleyBlock proto era))

Instance details

Associated Types

type Rep (CodecConfig (ShelleyBlock proto era)) :: Type -> Type

Methods

from :: CodecConfig (ShelleyBlock proto era) -> Rep (CodecConfig (ShelleyBlock proto era)) x

to :: Rep (CodecConfig (ShelleyBlock proto era)) x -> CodecConfig (ShelleyBlock proto era)

Generic (StorageConfig (ShelleyBlock proto era))

Instance details

Associated Types

type Rep (StorageConfig (ShelleyBlock proto era)) :: Type -> Type

Methods

from :: StorageConfig (ShelleyBlock proto era) -> Rep (StorageConfig (ShelleyBlock proto era)) x

to :: Rep (StorageConfig (ShelleyBlock proto era)) x -> StorageConfig (ShelleyBlock proto era)

ShelleyBasedEra era => NoThunks (BlockConfig (ShelleyBlock proto era))

Instance details

Methods

noThunks :: Context -> BlockConfig (ShelleyBlock proto era) -> IO (Maybe ThunkInfo)

wNoThunks :: Context -> BlockConfig (ShelleyBlock proto era) -> IO (Maybe ThunkInfo)

showTypeOf :: Proxy (BlockConfig (ShelleyBlock proto era)) -> String

NoThunks (CodecConfig (ShelleyBlock proto era))

Instance details

Methods

noThunks :: Context -> CodecConfig (ShelleyBlock proto era) -> IO (Maybe ThunkInfo)

wNoThunks :: Context -> CodecConfig (ShelleyBlock proto era) -> IO (Maybe ThunkInfo)

showTypeOf :: Proxy (CodecConfig (ShelleyBlock proto era)) -> String

NoThunks (StorageConfig (ShelleyBlock proto era))

Instance details

Methods

noThunks :: Context -> StorageConfig (ShelleyBlock proto era) -> IO (Maybe ThunkInfo)

wNoThunks :: Context -> StorageConfig (ShelleyBlock proto era) -> IO (Maybe ThunkInfo)

showTypeOf :: Proxy (StorageConfig (ShelleyBlock proto era)) -> String