Ouroboros.Consensus.Shelley.Ledger.Forge

ForgeState

data TPraosForgeState c

This is the state containing the private key corresponding to the public key that Praos uses to verify headers. That is why we call it TPraosForgeState instead of ShelleyForgeState.

Constructors

TPraosForgeState

Fields

tpraosHotKey ∷ !(HotKey c)

The online KES key used to sign blocks

Forging

forgeShelleyBlock

Arguments

∷ (MonadRandom m, TPraosCrypto c)

⇒ TopLevelConfig (ShelleyBlock c)

→ ForgeState (ShelleyBlock c)

→ BlockNo
Current block number

→ TickedLedgerState (ShelleyBlock c)
Current ledger

→ [GenTx (ShelleyBlock c)]
Txs to add in the block

→ TPraosProof c
Leader proof (IsLeader)

→ m (ShelleyBlock c)

Updating the state

evolveKESKeyIfNecessary

Arguments

∷ (MonadRandom m, TPraosCrypto c)

⇒ Update m (TPraosForgeState c)

→ KESPeriod
Relative KES period (to the start period of the OCert)

→ m ()

Get the KES key from the node state, evolve if its KES period doesn't match the given one.

Orphan instances

TPraosCrypto c ⇒ CanForge (ShelleyBlock c)

Instance details

Associated Types

type ForgeState (ShelleyBlock c) ∷ Type

Methods

forgeBlock ∷ MonadRandom m ⇒ TopLevelConfig (ShelleyBlock c) → ForgeState (ShelleyBlock c) → BlockNo → TickedLedgerState (ShelleyBlock c) → [GenTx (ShelleyBlock c)] → IsLeader (BlockProtocol (ShelleyBlock c)) → m (ShelleyBlock c)