Ouroboros.Consensus.Shelley.Ledger.History

Description

Store the history of the Shelley ledger view in order to allow time-travel to the recent past.

Intended for qualified import

Documentation

data LedgerView crypto

Data required by the Transitional Praos protocol from the Shelley ledger.

data LedgerViewHistory c

Ledger view history

Motivation: the ledger state does not give us information about past LedgerViews, therefore we track past LedgerView as long as they are required, as rollbacks are limited to k blocks.

See History for more details.

empty ∷ LedgerViewHistory c

Empty (genesis) LedgerView history

snapOld

Arguments

∷ Word64
Maximum rollback (k)

→ SlotNo
Slot number of the block that changed the LedgerView

→ LedgerView c
LedgerView before it changed

→ LedgerViewHistory c

→ LedgerViewHistory c

Take a snapshot of the LedgerView

find ∷ ∀ c. WithOrigin SlotNo → LedgerViewHistory c → Maybe (LedgerView c)

Serialisation

encodeLedgerViewHistory ∷ Crypto c ⇒ LedgerViewHistory c → Encoding

decodeLedgerViewHistory ∷ Crypto c ⇒ Decoder s (LedgerViewHistory c)