Ouroboros.Consensus.Shelley.Ledger.Mempool

Description

Shelley mempool integration

Documentation

data family GenTx blk

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

newtype ApplyTxError era

Constructors

ApplyTxError [PredicateFailure (EraRule "LEDGER" era)]

data family TxId tx

A generalized transaction, GenTx, identifier.

data family Validated x

" Validated " transaction or block

The ledger defines how to validate transactions and blocks. It's possible the type before and after validation may be distinct (eg Alonzo transactions), which originally motivated this family.

We also gain the related benefit that certain interface functions, such as those that reapply blocks, can have a more precise type now. TODO

Similarly, the Node-to-Client mini protocols can explicitly indicate that the client trusts the blocks from the local server, by having the server send Validated blocks to the client. TODO

Note that validation has different implications for a transaction than for a block. In particular, a validated transaction can be " reapplied " to different ledger states, whereas a validated block must only be " reapplied " to the exact same ledger state (eg as part of rebuilding from an on-disk ledger snapshot).

Since the ledger defines validation, see the ledger details for concrete examples of what determines the validity (wrt to a LedgerState) of a transaction and/or block. Example properties include: a transaction's claimed inputs exist and are still unspent, a block carries a sufficient cryptographic signature, etc.

data WithTop a

Add a unique top element to a lattice.

TODO This should be relocated to `cardano-base:Data.Measure'.

Constructors

NotTop a

Top

fixedBlockBodyOverhead :: Num a => a

txInBlockSize is used to estimate how many transactions we can grab from the Mempool to put into the block we are going to forge without exceeding the maximum block body size according to the ledger. If we exceed that limit, we will have forged a block that is invalid according to the ledger. We ourselves won't even adopt it, causing us to lose our slot, something we must try to avoid.

For this reason it is better to overestimate the size of a transaction than to underestimate. The only downside is that we maybe could have put one (or more?) transactions extra in that block.

As the sum of the serialised transaction sizes is not equal to the size of the serialised block body (TxSeq) consisting of those transactions (see cardano-node#1545 for an example), we account for some extra overhead per transaction as a safety margin.

Also see perTxOverhead.

mkShelleyTx :: forall era proto. ShelleyBasedEra era => Tx era -> GenTx (ShelleyBlock proto era)

mkShelleyValidatedTx :: forall era proto. ShelleyBasedEra era => Validated (Tx era) -> Validated (GenTx (ShelleyBlock proto era))

perTxOverhead :: Num a => a

See fixedBlockBodyOverhead.

Exported for tests

data AlonzoMeasure

Constructors

AlonzoMeasure

Fields

byteSize :: !ByteSize

exUnits :: !(ExUnits' (WithTop Natural))

fromExUnits :: ExUnits -> ExUnits' (WithTop Natural)

Orphan instances

(Typeable era, Typeable proto) => ShowProxy (GenTx (ShelleyBlock proto era) :: Type)

Instance details

Methods

showProxy :: Proxy (GenTx (ShelleyBlock proto era)) -> String

(Typeable era, Typeable proto) => ShowProxy (TxId (GenTx (ShelleyBlock proto era)) :: Type)

Instance details

Methods

showProxy :: Proxy (TxId (GenTx (ShelleyBlock proto era))) -> String

(Typeable era, Typeable proto) => ShowProxy (Validated (GenTx (ShelleyBlock proto era)) :: Type)

Instance details

Methods

showProxy :: Proxy (Validated (GenTx (ShelleyBlock proto era))) -> String

ShelleyBasedEra era => Eq (GenTx (ShelleyBlock proto era))

Instance details

Methods

(==) :: GenTx (ShelleyBlock proto era) -> GenTx (ShelleyBlock proto era) -> Bool

(/=) :: GenTx (ShelleyBlock proto era) -> GenTx (ShelleyBlock proto era) -> Bool

Eq (TxId (GenTx (ShelleyBlock proto era)))

Instance details

Methods

(==) :: TxId (GenTx (ShelleyBlock proto era)) -> TxId (GenTx (ShelleyBlock proto era)) -> Bool

(/=) :: TxId (GenTx (ShelleyBlock proto era)) -> TxId (GenTx (ShelleyBlock proto era)) -> Bool

ShelleyBasedEra era => Eq (Validated (GenTx (ShelleyBlock proto era)))

Instance details

Methods

(==) :: Validated (GenTx (ShelleyBlock proto era)) -> Validated (GenTx (ShelleyBlock proto era)) -> Bool

(/=) :: Validated (GenTx (ShelleyBlock proto era)) -> Validated (GenTx (ShelleyBlock proto era)) -> Bool

Ord (TxId (GenTx (ShelleyBlock proto era)))

Instance details

Methods

compare :: TxId (GenTx (ShelleyBlock proto era)) -> TxId (GenTx (ShelleyBlock proto era)) -> Ordering

(<) :: TxId (GenTx (ShelleyBlock proto era)) -> TxId (GenTx (ShelleyBlock proto era)) -> Bool

(<=) :: TxId (GenTx (ShelleyBlock proto era)) -> TxId (GenTx (ShelleyBlock proto era)) -> Bool

(>) :: TxId (GenTx (ShelleyBlock proto era)) -> TxId (GenTx (ShelleyBlock proto era)) -> Bool

(>=) :: TxId (GenTx (ShelleyBlock proto era)) -> TxId (GenTx (ShelleyBlock proto era)) -> Bool

max :: TxId (GenTx (ShelleyBlock proto era)) -> TxId (GenTx (ShelleyBlock proto era)) -> TxId (GenTx (ShelleyBlock proto era))

min :: TxId (GenTx (ShelleyBlock proto era)) -> TxId (GenTx (ShelleyBlock proto era)) -> TxId (GenTx (ShelleyBlock proto era))

ShelleyBasedEra era => Show (GenTx (ShelleyBlock proto era))

Instance details

Methods

showsPrec :: Int -> GenTx (ShelleyBlock proto era) -> ShowS

show :: GenTx (ShelleyBlock proto era) -> String

showList :: [GenTx (ShelleyBlock proto era)] -> ShowS

Show (GenTxId (ShelleyBlock proto era))

Instance details

Methods

showsPrec :: Int -> GenTxId (ShelleyBlock proto era) -> ShowS

show :: GenTxId (ShelleyBlock proto era) -> String

showList :: [GenTxId (ShelleyBlock proto era)] -> ShowS

ShelleyBasedEra era => Show (Validated (GenTx (ShelleyBlock proto era)))

Instance details

Methods

showsPrec :: Int -> Validated (GenTx (ShelleyBlock proto era)) -> ShowS

show :: Validated (GenTx (ShelleyBlock proto era)) -> String

showList :: [Validated (GenTx (ShelleyBlock proto era))] -> ShowS

Generic (GenTx (ShelleyBlock proto era))

Instance details

Associated Types

type Rep (GenTx (ShelleyBlock proto era)) :: Type -> Type

Methods

from :: GenTx (ShelleyBlock proto era) -> Rep (GenTx (ShelleyBlock proto era)) x

to :: Rep (GenTx (ShelleyBlock proto era)) x -> GenTx (ShelleyBlock proto era)

Generic (Validated (GenTx (ShelleyBlock proto era)))

Instance details

Associated Types

type Rep (Validated (GenTx (ShelleyBlock proto era))) :: Type -> Type

Methods

from :: Validated (GenTx (ShelleyBlock proto era)) -> Rep (Validated (GenTx (ShelleyBlock proto era))) x

to :: Rep (Validated (GenTx (ShelleyBlock proto era))) x -> Validated (GenTx (ShelleyBlock proto era))

ShelleyCompatible proto era => ToCBOR (GenTx (ShelleyBlock proto era))

Instance details

Methods

toCBOR :: GenTx (ShelleyBlock proto era) -> Encoding

encodedSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy (GenTx (ShelleyBlock proto era)) -> Size

encodedListSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy [GenTx (ShelleyBlock proto era)] -> Size

(Crypto (EraCrypto era), Typeable era, Typeable proto) => ToCBOR (TxId (GenTx (ShelleyBlock proto era)))

Instance details

Methods

toCBOR :: TxId (GenTx (ShelleyBlock proto era)) -> Encoding

encodedSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy (TxId (GenTx (ShelleyBlock proto era))) -> Size

encodedListSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy [TxId (GenTx (ShelleyBlock proto era))] -> Size

ShelleyCompatible proto era => FromCBOR (GenTx (ShelleyBlock proto era))

Instance details

Methods

fromCBOR :: Decoder s (GenTx (ShelleyBlock proto era))

label :: Proxy (GenTx (ShelleyBlock proto era)) -> Text

(Crypto (EraCrypto era), Typeable era, Typeable proto) => FromCBOR (TxId (GenTx (ShelleyBlock proto era)))

Instance details

Methods

fromCBOR :: Decoder s (TxId (GenTx (ShelleyBlock proto era)))

label :: Proxy (TxId (GenTx (ShelleyBlock proto era))) -> Text

ShelleyBasedEra era => NoThunks (GenTx (ShelleyBlock proto era))

Instance details

Methods

noThunks :: Context -> GenTx (ShelleyBlock proto era) -> IO (Maybe ThunkInfo)

wNoThunks :: Context -> GenTx (ShelleyBlock proto era) -> IO (Maybe ThunkInfo)

showTypeOf :: Proxy (GenTx (ShelleyBlock proto era)) -> String

NoThunks (TxId (GenTx (ShelleyBlock proto era)))

Instance details

Methods

noThunks :: Context -> TxId (GenTx (ShelleyBlock proto era)) -> IO (Maybe ThunkInfo)

wNoThunks :: Context -> TxId (GenTx (ShelleyBlock proto era)) -> IO (Maybe ThunkInfo)

showTypeOf :: Proxy (TxId (GenTx (ShelleyBlock proto era))) -> String

ShelleyBasedEra era => NoThunks (Validated (GenTx (ShelleyBlock proto era)))

Instance details

Methods

noThunks :: Context -> Validated (GenTx (ShelleyBlock proto era)) -> IO (Maybe ThunkInfo)

wNoThunks :: Context -> Validated (GenTx (ShelleyBlock proto era)) -> IO (Maybe ThunkInfo)

showTypeOf :: Proxy (Validated (GenTx (ShelleyBlock proto era))) -> String

ShelleyBasedEra era => HasTxId (GenTx (ShelleyBlock proto era))

Instance details

Methods

txId :: GenTx (ShelleyBlock proto era) -> TxId (GenTx (ShelleyBlock proto era))

ShelleyBasedEra era => Condense (GenTx (ShelleyBlock proto era))

Instance details

Methods

condense :: GenTx (ShelleyBlock proto era) -> String

Condense (GenTxId (ShelleyBlock proto era))

Instance details

Methods

condense :: GenTxId (ShelleyBlock proto era) -> String

Typeable era => ShowProxy (ApplyTxError era :: Type)

Instance details

Methods

showProxy :: Proxy (ApplyTxError era) -> String

ShelleyCompatible p (BabbageEra c) => TxLimits (ShelleyBlock p (BabbageEra c))

Instance details

Associated Types

type TxMeasure (ShelleyBlock p (BabbageEra c))

Methods

txMeasure :: Validated (GenTx (ShelleyBlock p (BabbageEra c))) -> TxMeasure (ShelleyBlock p (BabbageEra c))

txsBlockCapacity :: Ticked (LedgerState (ShelleyBlock p (BabbageEra c))) -> TxMeasure (ShelleyBlock p (BabbageEra c))

ShelleyCompatible p (AlonzoEra c) => TxLimits (ShelleyBlock p (AlonzoEra c))

Instance details

Associated Types

type TxMeasure (ShelleyBlock p (AlonzoEra c))

Methods

txMeasure :: Validated (GenTx (ShelleyBlock p (AlonzoEra c))) -> TxMeasure (ShelleyBlock p (AlonzoEra c))

txsBlockCapacity :: Ticked (LedgerState (ShelleyBlock p (AlonzoEra c))) -> TxMeasure (ShelleyBlock p (AlonzoEra c))

ShelleyCompatible p (MaryEra c) => TxLimits (ShelleyBlock p (MaryEra c))

Instance details

Associated Types

type TxMeasure (ShelleyBlock p (MaryEra c))

Methods

txMeasure :: Validated (GenTx (ShelleyBlock p (MaryEra c))) -> TxMeasure (ShelleyBlock p (MaryEra c))

txsBlockCapacity :: Ticked (LedgerState (ShelleyBlock p (MaryEra c))) -> TxMeasure (ShelleyBlock p (MaryEra c))

ShelleyCompatible p (AllegraEra c) => TxLimits (ShelleyBlock p (AllegraEra c))

Instance details

Associated Types

type TxMeasure (ShelleyBlock p (AllegraEra c))

Methods

txMeasure :: Validated (GenTx (ShelleyBlock p (AllegraEra c))) -> TxMeasure (ShelleyBlock p (AllegraEra c))

txsBlockCapacity :: Ticked (LedgerState (ShelleyBlock p (AllegraEra c))) -> TxMeasure (ShelleyBlock p (AllegraEra c))

ShelleyCompatible p (ShelleyEra c) => TxLimits (ShelleyBlock p (ShelleyEra c))

Instance details

Associated Types

type TxMeasure (ShelleyBlock p (ShelleyEra c))

Methods

txMeasure :: Validated (GenTx (ShelleyBlock p (ShelleyEra c))) -> TxMeasure (ShelleyBlock p (ShelleyEra c))

txsBlockCapacity :: Ticked (LedgerState (ShelleyBlock p (ShelleyEra c))) -> TxMeasure (ShelleyBlock p (ShelleyEra c))

ShelleyCompatible proto era => LedgerSupportsMempool (ShelleyBlock proto era)

Instance details

Methods

txInvariant :: GenTx (ShelleyBlock proto era) -> Bool

applyTx :: LedgerConfig (ShelleyBlock proto era) -> WhetherToIntervene -> SlotNo -> GenTx (ShelleyBlock proto era) -> TickedLedgerState (ShelleyBlock proto era) -> Except (ApplyTxErr (ShelleyBlock proto era)) (TickedLedgerState (ShelleyBlock proto era), Validated (GenTx (ShelleyBlock proto era)))

reapplyTx :: LedgerConfig (ShelleyBlock proto era) -> SlotNo -> Validated (GenTx (ShelleyBlock proto era)) -> TickedLedgerState (ShelleyBlock proto era) -> Except (ApplyTxErr (ShelleyBlock proto era)) (TickedLedgerState (ShelleyBlock proto era))

txsMaxBytes :: TickedLedgerState (ShelleyBlock proto era) -> Word32

txInBlockSize :: GenTx (ShelleyBlock proto era) -> Word32

txForgetValidated :: Validated (GenTx (ShelleyBlock proto era)) -> GenTx (ShelleyBlock proto era)

ShelleyBasedEra era => HasTxs (ShelleyBlock proto era)

Instance details

Methods

extractTxs :: ShelleyBlock proto era -> [GenTx (ShelleyBlock proto era)]