Ouroboros.Consensus.Shelley.Ledger.Query

Documentation

data family BlockQuery blk :: Type -> Type

Different queries supported by the ledger, indexed by the result type.

newtype NonMyopicMemberRewards c

Constructors

NonMyopicMemberRewards

Fields

unNonMyopicMemberRewards :: Map (Either Coin (Credential 'Staking c)) (Map (KeyHash 'StakePool c) Coin)

querySupportedVersion :: BlockQuery (ShelleyBlock proto era) result -> ShelleyNodeToClientVersion -> Bool

Is the given query supported by the given ShelleyNodeToClientVersion?

Serialisation

decodeShelleyQuery :: ShelleyBasedEra era => Decoder s (SomeSecond BlockQuery (ShelleyBlock proto era))

decodeShelleyResult :: ShelleyCompatible proto era => BlockQuery (ShelleyBlock proto era) result -> forall s. Decoder s result

encodeShelleyQuery :: ShelleyBasedEra era => BlockQuery (ShelleyBlock proto era) result -> Encoding

encodeShelleyResult :: ShelleyCompatible proto era => BlockQuery (ShelleyBlock proto era) result -> result -> Encoding

Orphan instances

SameDepIndex (BlockQuery (ShelleyBlock proto era))

Instance details

Methods

sameDepIndex :: BlockQuery (ShelleyBlock proto era) a -> BlockQuery (ShelleyBlock proto era) b -> Maybe (a :~: b)

ShelleyCompatible proto era => ShowQuery (BlockQuery (ShelleyBlock proto era))

Instance details

Methods

showResult :: BlockQuery (ShelleyBlock proto era) result -> result -> String

Eq (BlockQuery (ShelleyBlock proto era) result)

Instance details

Methods

(==) :: BlockQuery (ShelleyBlock proto era) result -> BlockQuery (ShelleyBlock proto era) result -> Bool

(/=) :: BlockQuery (ShelleyBlock proto era) result -> BlockQuery (ShelleyBlock proto era) result -> Bool

Show (BlockQuery (ShelleyBlock proto era) result)

Instance details

Methods

showsPrec :: Int -> BlockQuery (ShelleyBlock proto era) result -> ShowS

show :: BlockQuery (ShelleyBlock proto era) result -> String

showList :: [BlockQuery (ShelleyBlock proto era) result] -> ShowS

ShelleyCompatible proto era => QueryLedger (ShelleyBlock proto era)

Instance details

Methods

answerBlockQuery :: ExtLedgerCfg (ShelleyBlock proto era) -> BlockQuery (ShelleyBlock proto era) result -> ExtLedgerState (ShelleyBlock proto era) -> result

(Typeable era, Typeable proto) => ShowProxy (BlockQuery (ShelleyBlock proto era) :: Type -> Type)

Instance details

Methods

showProxy :: Proxy (BlockQuery (ShelleyBlock proto era)) -> String