Ouroboros.Consensus.Shelley.Node

Documentation

protocolInfoShelley

Arguments

∷ (IOLike m, TPraosCrypto era)

⇒ ShelleyGenesis era

→ Nonce
The initial nonce, typically derived from the hash of Genesis config JSON file.

→ MaxMajorProtVer

→ ProtVer

→ Maybe (TPraosLeaderCredentials era)

→ ProtocolInfo m (ShelleyBlock era)

protocolClientInfoShelley ∷ ProtocolClientInfo (ShelleyBlock era)

data ShelleyGenesis era

Shelley genesis information

Note that this is needed only for a pure Shelley network, hence it being defined here rather than in its own module. In mainnet, Shelley will transition naturally from Byron, and thus will never have its own genesis information.

Constructors

ShelleyGenesis

Fields

sgSystemStart ∷ !UTCTime

sgNetworkMagic ∷ !Word32

sgNetworkId ∷ !Network

sgActiveSlotsCoeff ∷ !Rational

sgSecurityParam ∷ !Word64

sgEpochLength ∷ !EpochSize

sgSlotsPerKESPeriod ∷ !Word64

sgMaxKESEvolutions ∷ !Word64

sgSlotLength ∷ !NominalDiffTime

sgUpdateQuorum ∷ !Word64

sgMaxLovelaceSupply ∷ !Word64

sgProtocolParams ∷ !(PParams era)

sgGenDelegs ∷ !(Map (KeyHash Genesis era) (GenDelegPair era))

sgInitialFunds ∷ !(Map (Addr era) Coin)

sgStaking ∷ !(ShelleyGenesisStaking era)

data ShelleyGenesisStaking era

Genesis Shelley staking configuration.

This allows us to configure some initial stake pools and delegation to them, in order to test Praos in a static configuration, without requiring on-chain registration and delegation.

For simplicity, pools defined in the genesis staking do not pay deposits for their registration.

Constructors

ShelleyGenesisStaking

Fields

sgsPools ∷ !(Map (KeyHash StakePool era) (PoolParams era))

Pools to register

The key in this map is the hash of the public key of the _pool_. This need not correspond to any payment or staking key, but must correspond to the cold key held by TPraosIsCoreNode.

sgsStake ∷ !(Map (KeyHash Staking era) (KeyHash StakePool era))

Stake-holding key hash credentials and the pools to delegate that stake to. We require the raw staking key hash in order to:

Avoid pointer addresses, which would be tricky when there's no slot or transaction to point to.

Avoid script credentials.

data TPraosLeaderCredentials era

Constructors

TPraosLeaderCredentials

Fields

tpraosLeaderCredentialsInitSignKey ∷ SignKeyKES era

The unevolved signing KES key (at evolution 0).

Note that this is not inside TPraosCanBeLeader since it gets evolved automatically, whereas TPraosCanBeLeader does not change.

tpraosLeaderCredentialsCanBeLeader ∷ TPraosCanBeLeader era

shelleyBlockForging ∷ ∀ m era. (TPraosCrypto era, IOLike m) ⇒ TPraosParams → TPraosLeaderCredentials era → m (BlockForging m (ShelleyBlock era))

tpraosBlockIssuerVKey ∷ Maybe (TPraosLeaderCredentials era) → BlockIssuerVKey era

data ProtVer

data Nonce

Evolving nonce type.

Constructors

Nonce !(Hash Blake2b_256 Nonce)

NeutralNonce
Identity element

newtype MaxMajorProtVer

The maximum major protocol version.

Must be at least the current major protocol version. For Cardano mainnet, the Shelley era has major protocol verison 2.

Constructors

MaxMajorProtVer

Fields

getMaxMajorProtVer ∷ Natural

emptyGenesisStaking ∷ ShelleyGenesisStaking era

Empty genesis staking

validateGenesis ∷ TPraosCrypto era ⇒ ShelleyGenesis era → Either String ()

Check the validity of the genesis config. To be used in conjunction with assertWithMsg.

Orphan instances

TPraosCrypto era ⇒ RunNode (ShelleyBlock era)

Instance details

Methods

nodeBlockFetchSize ∷ Header (ShelleyBlock era) → SizeInBytes

nodeImmutableDbChunkInfo ∷ TopLevelConfig (ShelleyBlock era) → ChunkInfo

nodeCheckIntegrity ∷ TopLevelConfig (ShelleyBlock era) → ShelleyBlock era → Bool

nodeInitChainDB ∷ IOLike m ⇒ TopLevelConfig (ShelleyBlock era) → InitChainDB m (ShelleyBlock era) → m ()

ConfigSupportsNode (ShelleyBlock era)

Instance details

Methods

getSystemStart ∷ BlockConfig (ShelleyBlock era) → SystemStart

getNetworkMagic ∷ BlockConfig (ShelleyBlock era) → NetworkMagic

InspectLedger (ShelleyBlock era)

Instance details

Associated Types

type LedgerWarning (ShelleyBlock era) ∷ Type

type LedgerUpdate (ShelleyBlock era) ∷ Type

Methods

inspectLedger ∷ TopLevelConfig (ShelleyBlock era) → LedgerState (ShelleyBlock era) → LedgerState (ShelleyBlock era) → [LedgerEvent (ShelleyBlock era)]