Ouroboros.Consensus.Shelley.Node.Common

Description

Node configuration common to all (era, protocol) combinations deriving from Shelley.

Documentation

data ProtocolParamsShelleyBased era

Parameters common to all Shelley-based ledgers.

When running a chain with multiple Shelley-based eras, in addition to the per-era protocol parameters, one value of ProtocolParamsShelleyBased will be needed, which is shared among all Shelley-based eras.

The era parameter determines from which era the genesis config will be used.

Constructors

ProtocolParamsShelleyBased

Fields

shelleyBasedGenesis ∷ ShelleyGenesis (EraCrypto era)

shelleyBasedInitialNonce ∷ Nonce

The initial nonce, typically derived from the hash of Genesis config JSON file.

WARNING: chains using different values of this parameter will be mutually incompatible.

shelleyBasedLeaderCredentials ∷ [ShelleyLeaderCredentials (EraCrypto era)]

class (ShelleyCompatible proto era, TxLimits (ShelleyBlock proto era), EraCrypto era ~ c) ⇒ ShelleyEraWithCrypto c proto era

Needed in '*SharedBlockForging' because we can't partially apply equality constraints.

data ShelleyLeaderCredentials c

Constructors

ShelleyLeaderCredentials

Fields

shelleyLeaderCredentialsInitSignKey ∷ SignKeyKES c

The unevolved signing KES key (at evolution 0).

Note that this is not inside ShelleyCanBeLeader since it gets evolved automatically, whereas ShelleyCanBeLeader does not change.

shelleyLeaderCredentialsCanBeLeader ∷ PraosCanBeLeader c

shelleyLeaderCredentialsLabel ∷ Text

Identifier for this set of credentials.

Useful when the node is running with multiple sets of credentials.

shelleyBlockIssuerVKey ∷ ShelleyLeaderCredentials c → VKey 'BlockIssuer c

Orphan instances

ShelleyCompatible proto era ⇒ NodeInitStorage (ShelleyBlock proto era)

Instance details

Methods

nodeImmutableDbChunkInfo ∷ StorageConfig (ShelleyBlock proto era) → ChunkInfo

nodeCheckIntegrity ∷ StorageConfig (ShelleyBlock proto era) → ShelleyBlock proto era → Bool

nodeInitChainDB ∷ IOLike m ⇒ StorageConfig (ShelleyBlock proto era) → InitChainDB m (ShelleyBlock proto era) → m ()

ConfigSupportsNode (ShelleyBlock proto era)

Instance details

Methods

getSystemStart ∷ BlockConfig (ShelleyBlock proto era) → SystemStart

getNetworkMagic ∷ BlockConfig (ShelleyBlock proto era) → NetworkMagic