Ouroboros.Consensus.Shelley.Node.Serialisation

Orphan instances

Crypto c ⇒ ReconstructNestedCtxt Header (ShelleyBlock c)

Instance details

Methods

reconstructPrefixLen ∷ proxy (Header (ShelleyBlock c)) → Word8

reconstructNestedCtxt ∷ proxy (Header (ShelleyBlock c)) → ShortByteString → SizeInBytes → SomeBlock (NestedCtxt Header) (ShelleyBlock c)

Crypto c ⇒ SerialiseNodeToNodeConstraints (ShelleyBlock c)

Instance details

Crypto c ⇒ SerialiseNodeToClientConstraints (ShelleyBlock c)

Instance details

Crypto c ⇒ SerialiseDiskConstraints (ShelleyBlock c)

Instance details

Crypto c ⇒ LgrDbSerialiseConstraints (ShelleyBlock c)

Instance details

Crypto c ⇒ ImmDbSerialiseConstraints (ShelleyBlock c)

Instance details

Crypto c ⇒ VolDbSerialiseConstraints (ShelleyBlock c)

Instance details

Crypto c ⇒ EncodeDisk (ShelleyBlock c) (AnnTip (ShelleyBlock c))

Instance details

Methods

encodeDisk ∷ CodecConfig (ShelleyBlock c) → AnnTip (ShelleyBlock c) → Encoding

Crypto c ⇒ EncodeDisk (ShelleyBlock c) (TPraosState c)
ConsensusState (BlockProtocol (ShelleyBlock c))

Instance details

Methods

encodeDisk ∷ CodecConfig (ShelleyBlock c) → TPraosState c → Encoding

Crypto c ⇒ EncodeDisk (ShelleyBlock c) (LedgerState (ShelleyBlock c))

Instance details

Methods

encodeDisk ∷ CodecConfig (ShelleyBlock c) → LedgerState (ShelleyBlock c) → Encoding

Crypto c ⇒ EncodeDisk (ShelleyBlock c) (Header (ShelleyBlock c))

Instance details

Methods

encodeDisk ∷ CodecConfig (ShelleyBlock c) → Header (ShelleyBlock c) → Encoding

Crypto c ⇒ EncodeDisk (ShelleyBlock c) (ShelleyBlock c)

Instance details

Methods

encodeDisk ∷ CodecConfig (ShelleyBlock c) → ShelleyBlock c → Encoding

Crypto c ⇒ DecodeDisk (ShelleyBlock c) (AnnTip (ShelleyBlock c))

Instance details

Methods

decodeDisk ∷ CodecConfig (ShelleyBlock c) → ∀ s. Decoder s (AnnTip (ShelleyBlock c))

Crypto c ⇒ DecodeDisk (ShelleyBlock c) (TPraosState c)
ConsensusState (BlockProtocol (ShelleyBlock c))

Instance details

Methods

decodeDisk ∷ CodecConfig (ShelleyBlock c) → ∀ s. Decoder s (TPraosState c)

Crypto c ⇒ DecodeDisk (ShelleyBlock c) (LedgerState (ShelleyBlock c))

Instance details

Methods

decodeDisk ∷ CodecConfig (ShelleyBlock c) → ∀ s. Decoder s (LedgerState (ShelleyBlock c))

Crypto c ⇒ EncodeDiskDepIx (NestedCtxt Header) (ShelleyBlock c)

Instance details

Methods

encodeDiskDepIx ∷ CodecConfig (ShelleyBlock c) → SomeBlock (NestedCtxt Header) (ShelleyBlock c) → Encoding

Crypto c ⇒ EncodeDiskDep (NestedCtxt Header) (ShelleyBlock c)

Instance details

Methods

encodeDiskDep ∷ CodecConfig (ShelleyBlock c) → NestedCtxt Header (ShelleyBlock c) a → a → Encoding

Crypto c ⇒ DecodeDiskDepIx (NestedCtxt Header) (ShelleyBlock c)

Instance details

Methods

decodeDiskDepIx ∷ CodecConfig (ShelleyBlock c) → Decoder s (SomeBlock (NestedCtxt Header) (ShelleyBlock c))

Crypto c ⇒ DecodeDiskDep (NestedCtxt Header) (ShelleyBlock c)

Instance details

Methods

decodeDiskDep ∷ CodecConfig (ShelleyBlock c) → NestedCtxt Header (ShelleyBlock c) a → ∀ s. Decoder s (ByteString → a)

Crypto c ⇒ SerialiseNodeToNode (ShelleyBlock c) (GenTxId (ShelleyBlock c))

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock c) → NodeToNodeVersion (ShelleyBlock c) → GenTxId (ShelleyBlock c) → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock c) → NodeToNodeVersion (ShelleyBlock c) → ∀ s. Decoder s (GenTxId (ShelleyBlock c))

Crypto c ⇒ SerialiseNodeToNode (ShelleyBlock c) (GenTx (ShelleyBlock c))
The To/FromCBOR instances defined in cardano-ledger-specs use CBOR-in-CBOR to get the annotation.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock c) → NodeToNodeVersion (ShelleyBlock c) → GenTx (ShelleyBlock c) → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock c) → NodeToNodeVersion (ShelleyBlock c) → ∀ s. Decoder s (GenTx (ShelleyBlock c))

SerialiseNodeToNode (ShelleyBlock c) (SerialisedHeader (ShelleyBlock c))
We use CBOR-in-CBOR

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock c) → NodeToNodeVersion (ShelleyBlock c) → SerialisedHeader (ShelleyBlock c) → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock c) → NodeToNodeVersion (ShelleyBlock c) → ∀ s. Decoder s (SerialisedHeader (ShelleyBlock c))

Crypto c ⇒ SerialiseNodeToNode (ShelleyBlock c) (Header (ShelleyBlock c))
CBOR-in-CBOR to be compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock c) → NodeToNodeVersion (ShelleyBlock c) → Header (ShelleyBlock c) → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock c) → NodeToNodeVersion (ShelleyBlock c) → ∀ s. Decoder s (Header (ShelleyBlock c))

SerialiseNodeToNode (ShelleyBlock c) (Serialised (ShelleyBlock c))
Serialised uses CBOR-in-CBOR by default.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock c) → NodeToNodeVersion (ShelleyBlock c) → Serialised (ShelleyBlock c) → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock c) → NodeToNodeVersion (ShelleyBlock c) → ∀ s. Decoder s (Serialised (ShelleyBlock c))

Crypto c ⇒ SerialiseNodeToNode (ShelleyBlock c) (ShelleyBlock c)
CBOR-in-CBOR for the annotation. This also makes it compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock c) → NodeToNodeVersion (ShelleyBlock c) → ShelleyBlock c → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock c) → NodeToNodeVersion (ShelleyBlock c) → ∀ s. Decoder s (ShelleyBlock c)

Crypto c ⇒ SerialiseNodeToNode (ShelleyBlock c) (ShelleyHash c)
HeaderHash '(ShelleyBlock c)'

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock c) → NodeToNodeVersion (ShelleyBlock c) → ShelleyHash c → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock c) → NodeToNodeVersion (ShelleyBlock c) → ∀ s. Decoder s (ShelleyHash c)

Crypto c ⇒ SerialiseNodeToClient (ShelleyBlock c) (ApplyTxError c)
ApplyTxErr '(ShelleyBlock c)'

Instance details

Methods

encodeNodeToClient ∷ CodecConfig (ShelleyBlock c) → NodeToClientVersion (ShelleyBlock c) → ApplyTxError c → Encoding

decodeNodeToClient ∷ CodecConfig (ShelleyBlock c) → NodeToClientVersion (ShelleyBlock c) → ∀ s. Decoder s (ApplyTxError c)

Crypto c ⇒ SerialiseNodeToClient (ShelleyBlock c) (GenTx (ShelleyBlock c))
Uses CBOR-in-CBOR in the To/FromCBOR instances to get the annotation.

Instance details

Methods

encodeNodeToClient ∷ CodecConfig (ShelleyBlock c) → NodeToClientVersion (ShelleyBlock c) → GenTx (ShelleyBlock c) → Encoding

decodeNodeToClient ∷ CodecConfig (ShelleyBlock c) → NodeToClientVersion (ShelleyBlock c) → ∀ s. Decoder s (GenTx (ShelleyBlock c))

SerialiseNodeToClient (ShelleyBlock c) (Serialised (ShelleyBlock c))
Serialised uses CBOR-in-CBOR by default.

Instance details

Methods

encodeNodeToClient ∷ CodecConfig (ShelleyBlock c) → NodeToClientVersion (ShelleyBlock c) → Serialised (ShelleyBlock c) → Encoding

decodeNodeToClient ∷ CodecConfig (ShelleyBlock c) → NodeToClientVersion (ShelleyBlock c) → ∀ s. Decoder s (Serialised (ShelleyBlock c))

Crypto c ⇒ SerialiseNodeToClient (ShelleyBlock c) (ShelleyBlock c)
CBOR-in-CBOR for the annotation. This also makes it compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToClient ∷ CodecConfig (ShelleyBlock c) → NodeToClientVersion (ShelleyBlock c) → ShelleyBlock c → Encoding

decodeNodeToClient ∷ CodecConfig (ShelleyBlock c) → NodeToClientVersion (ShelleyBlock c) → ∀ s. Decoder s (ShelleyBlock c)

Crypto c ⇒ SerialiseNodeToClient (ShelleyBlock c) (ShelleyHash c)
HeaderHash '(ShelleyBlock c)'

Instance details

Methods

encodeNodeToClient ∷ CodecConfig (ShelleyBlock c) → NodeToClientVersion (ShelleyBlock c) → ShelleyHash c → Encoding

decodeNodeToClient ∷ CodecConfig (ShelleyBlock c) → NodeToClientVersion (ShelleyBlock c) → ∀ s. Decoder s (ShelleyHash c)

Crypto c ⇒ SerialiseResult (ShelleyBlock c) (Query (ShelleyBlock c))

Instance details

Methods

encodeResult ∷ CodecConfig (ShelleyBlock c) → NodeToClientVersion (ShelleyBlock c) → Query (ShelleyBlock c) result → result → Encoding

decodeResult ∷ CodecConfig (ShelleyBlock c) → NodeToClientVersion (ShelleyBlock c) → Query (ShelleyBlock c) result → ∀ s. Decoder s result

Crypto c ⇒ DecodeDisk (ShelleyBlock c) (ByteString → Header (ShelleyBlock c))

Instance details

Methods

decodeDisk ∷ CodecConfig (ShelleyBlock c) → ∀ s. Decoder s (ByteString → Header (ShelleyBlock c))

Crypto c ⇒ DecodeDisk (ShelleyBlock c) (ByteString → ShelleyBlock c)

Instance details

Methods

decodeDisk ∷ CodecConfig (ShelleyBlock c) → ∀ s. Decoder s (ByteString → ShelleyBlock c)

Crypto c ⇒ SerialiseNodeToClient (ShelleyBlock c) (Some (Query (ShelleyBlock c)))

Instance details

Methods

encodeNodeToClient ∷ CodecConfig (ShelleyBlock c) → NodeToClientVersion (ShelleyBlock c) → Some (Query (ShelleyBlock c)) → Encoding

decodeNodeToClient ∷ CodecConfig (ShelleyBlock c) → NodeToClientVersion (ShelleyBlock c) → ∀ s. Decoder s (Some (Query (ShelleyBlock c)))