Ouroboros.Consensus.Shelley.Node.TPraos

Documentation

newtype MaxMajorProtVer

The maximum major protocol version.

Must be at least the current major protocol version. For Cardano mainnet, the Shelley era has major protocol verison 2.

Constructors

MaxMajorProtVer

Fields

getMaxMajorProtVer :: Natural

data ProtocolParamsAllegra c

Parameters needed to run Allegra

Constructors

ProtocolParamsAllegra

Fields

allegraProtVer :: ProtVer

allegraMaxTxCapacityOverrides :: Overrides (ShelleyBlock (TPraos c) (AllegraEra c))

data ProtocolParamsAlonzo c

Parameters needed to run Alonzo

Constructors

ProtocolParamsAlonzo

Fields

alonzoProtVer :: ProtVer

alonzoMaxTxCapacityOverrides :: Overrides (ShelleyBlock (TPraos c) (AlonzoEra c))

data ProtocolParamsMary c

Parameters needed to run Mary

Constructors

ProtocolParamsMary

Fields

maryProtVer :: ProtVer

maryMaxTxCapacityOverrides :: Overrides (ShelleyBlock (TPraos c) (MaryEra c))

data ProtocolParamsShelley c

Parameters needed to run Shelley

Constructors

ProtocolParamsShelley

Fields

shelleyProtVer :: ProtVer

shelleyMaxTxCapacityOverrides :: Overrides (ShelleyBlock (TPraos c) (ShelleyEra c))

data ProtocolParamsShelleyBased era

Parameters common to all Shelley-based ledgers.

When running a chain with multiple Shelley-based eras, in addition to the per-era protocol parameters, one value of ProtocolParamsShelleyBased will be needed, which is shared among all Shelley-based eras.

The era parameter determines from which era the genesis config will be used.

Constructors

ProtocolParamsShelleyBased

Fields

shelleyBasedGenesis :: ShelleyGenesis era

shelleyBasedInitialNonce :: Nonce

The initial nonce, typically derived from the hash of Genesis config JSON file.

WARNING: chains using different values of this parameter will be mutually incompatible.

shelleyBasedLeaderCredentials :: [ShelleyLeaderCredentials (EraCrypto era)]

data Nonce

Evolving nonce type.

Constructors

Nonce !(Hash Blake2b_256 Nonce)

NeutralNonce
Identity element

data ProtVer

Constructors

ProtVer

Fields

pvMajor :: !Natural

pvMinor :: !Natural

data ShelleyGenesis era

Shelley genesis information

Note that this is needed only for a pure Shelley network, hence it being defined here rather than in its own module. In mainnet, Shelley will transition naturally from Byron, and thus will never have its own genesis information.

Constructors

ShelleyGenesis

Fields

sgSystemStart :: !UTCTime

sgNetworkMagic :: !Word32

sgNetworkId :: !Network

sgActiveSlotsCoeff :: !PositiveUnitInterval

sgSecurityParam :: !Word64

sgEpochLength :: !EpochSize

sgSlotsPerKESPeriod :: !Word64

sgMaxKESEvolutions :: !Word64

sgSlotLength :: !NominalDiffTime

sgUpdateQuorum :: !Word64

sgMaxLovelaceSupply :: !Word64

sgProtocolParams :: !(PParams era)

sgGenDelegs :: !(Map (KeyHash 'Genesis (Crypto era)) (GenDelegPair (Crypto era)))

sgInitialFunds :: !(Map (Addr (Crypto era)) Coin)

sgStaking :: !(ShelleyGenesisStaking (Crypto era))

data ShelleyGenesisStaking crypto

Genesis Shelley staking configuration.

This allows us to configure some initial stake pools and delegation to them, in order to test Praos in a static configuration, without requiring on-chain registration and delegation.

For simplicity, pools defined in the genesis staking do not pay deposits for their registration.

Constructors

ShelleyGenesisStaking

Fields

sgsPools :: !(Map (KeyHash 'StakePool crypto) (PoolParams crypto))

Pools to register

The key in this map is the hash of the public key of the _pool_. This need not correspond to any payment or staking key, but must correspond to the cold key held by TPraosIsCoreNode.

sgsStake :: !(Map (KeyHash 'Staking crypto) (KeyHash 'StakePool crypto))

Stake-holding key hash credentials and the pools to delegate that stake to. We require the raw staking key hash in order to:

Avoid pointer addresses, which would be tricky when there's no slot or transaction to point to.

Avoid script credentials.

emptyGenesisStaking :: ShelleyGenesisStaking crypto

Empty genesis staking

data ShelleyLeaderCredentials c

Constructors

ShelleyLeaderCredentials

Fields

shelleyLeaderCredentialsInitSignKey :: SignKeyKES c

The unevolved signing KES key (at evolution 0).

Note that this is not inside ShelleyCanBeLeader since it gets evolved automatically, whereas ShelleyCanBeLeader does not change.

shelleyLeaderCredentialsCanBeLeader :: PraosCanBeLeader c

shelleyLeaderCredentialsLabel :: Text

Identifier for this set of credentials.

Useful when the node is running with multiple sets of credentials.

protocolInfoShelley :: forall m c. (IOLike m, PraosCrypto c, ShelleyCompatible (TPraos c) (ShelleyEra c), TxLimits (ShelleyBlock (TPraos c) (ShelleyEra c))) => ProtocolParamsShelleyBased (ShelleyEra c) -> ProtocolParamsShelley c -> ProtocolInfo m (ShelleyBlock (TPraos c) (ShelleyEra c))

protocolInfoTPraosShelleyBased :: forall m era c. (IOLike m, PraosCrypto c, ShelleyCompatible (TPraos c) era, TxLimits (ShelleyBlock (TPraos c) era), c ~ EraCrypto era) => ProtocolParamsShelleyBased era -> TranslationContext era -> ProtVer -> Overrides (ShelleyBlock (TPraos c) era) -> ProtocolInfo m (ShelleyBlock (TPraos c) era)

registerGenesisStaking :: forall era. ShelleyBasedEra era => ShelleyGenesisStaking (EraCrypto era) -> NewEpochState era -> NewEpochState era

Register the initial staking information in the NewEpochState.

HERE BE DRAGONS! This function is intended to help in testing.

In production, the genesis should not contain any initial staking.

Any existing staking information is overridden, but the UTxO is left untouched.

TODO adapt and reuse registerGenesisStaking from cardano-ledger-specs.

registerInitialFunds :: forall era. (ShelleyBasedEra era, HasCallStack) => Map (Addr (EraCrypto era)) Coin -> NewEpochState era -> NewEpochState era

Register the initial funds in the NewEpochState.

HERE BE DRAGONS! This function is intended to help in testing.

In production, the genesis should not contain any initial funds.

The given funds are added to the existing UTxO.

PRECONDITION: the given funds must not be part of the existing UTxO. > forall (addr, _) in initialFunds. > Map.notElem (SL.initialFundsPseudoTxIn addr) existingUTxO

PROPERTY: > genesisUTxO genesis > == genesisUTxO' (sgInitialFunds genesis) > == extractUTxO (registerInitialFunds (sgInitialFunds genesis) > NewEpochState)

TODO move to cardano-ledger-specs.

shelleyBlockForging :: forall m era c. (ShelleyCompatible (TPraos c) era, PraosCrypto c, c ~ EraCrypto era, TxLimits (ShelleyBlock (TPraos c) era), IOLike m) => TPraosParams -> Overrides (ShelleyBlock (TPraos c) era) -> ShelleyLeaderCredentials (EraCrypto era) -> m (BlockForging m (ShelleyBlock (TPraos c) era))

Create a BlockForging record for a single era.

In case the same credentials should be shared across multiple Shelley-based eras, use shelleySharedBlockForging.

shelleySharedBlockForging :: forall m c era. (PraosCrypto c, ShelleyEraWithCrypto c (TPraos c) era, IOLike m) => HotKey c m -> (SlotNo -> KESPeriod) -> ShelleyLeaderCredentials c -> Overrides (ShelleyBlock (TPraos c) era) -> BlockForging m (ShelleyBlock (TPraos c) era)

Create a BlockForging record safely using a given Hotkey.

The name of the era (separated by a _) will be appended to each forgeLabel.

validateGenesis :: ShelleyBasedEra era => ShelleyGenesis era -> Either String ()

Check the validity of the genesis config. To be used in conjunction with assertWithMsg.