Ouroboros.Consensus.Shelley.Protocol

Description

Transitional Praos.

Transitional praos allows for the overlaying of Praos with an overlay schedule determining slots to be produced by BFT

Documentation

newtype MaxMajorProtVer

The maximum major protocol version.

Must be at least the current major protocol version. For Cardano mainnet, the Shelley era has major protocol verison 2.

Constructors

MaxMajorProtVer

Fields

getMaxMajorProtVer ∷ Natural

data SelfIssued

Separate type instead of Bool for the custom Ord instance + documentation.

Constructors

SelfIssued
A block we produced ourself

NotSelfIssued
A block produced by another node

data TPraos c

data TPraosCanBeLeader c

Constructors

TPraosCanBeLeader

Fields

tpraosCanBeLeaderOpCert ∷ !(OCert c)

Certificate delegating rights from the stake pool cold key (or genesis stakeholder delegate cold key) to the online KES key.

tpraosCanBeLeaderColdVerKey ∷ !(VKey 'BlockIssuer c)

Stake pool cold key or genesis stakeholder delegate cold key.

tpraosCanBeLeaderSignKeyVRF ∷ !(SignKeyVRF c)

data TPraosChainSelectView c

View of the ledger tip for chain selection.

We order between chains as follows:

By chain length, with longer chains always preferred.

If the tip of each chain has the same slot number, we prefer the one tip that we produced ourselves.

If the tip of each chain was issued by the same agent, then we prefer the chain whose tip has the highest ocert issue number.

By the leader value of the chain tip, with lower values preferred.

Constructors

TPraosChainSelectView

Fields

csvChainLength ∷ BlockNo

csvSlotNo ∷ SlotNo

csvSelfIssued ∷ SelfIssued

csvIssuer ∷ VKey 'BlockIssuer c

csvIssueNo ∷ Word64

csvLeaderVRF ∷ OutputVRF (VRF c)

data TPraosFields c toSign

Constructors

TPraosFields

Fields

tpraosSignature ∷ SignedKES c toSign

tpraosToSign ∷ toSign

data TPraosIsLeader c

Assembled proof that the issuer has the right to issue a block in the selected slot.

Constructors

TPraosIsLeader

Fields

tpraosIsLeaderEta ∷ CertifiedVRF c Nonce

tpraosIsLeaderProof ∷ CertifiedVRF c Natural

tpraosIsLeaderGenVRFHash ∷ Maybe (Hash c (VerKeyVRF c))

When in the overlay schedule (otherwise Nothing), return the hash of the VRF verification key in the overlay schedule

data TPraosParams

TPraos parameters that are node independent

Constructors

TPraosParams

Fields

tpraosSlotsPerKESPeriod ∷ !Word64

See slotsPerKESPeriod.

tpraosLeaderF ∷ !ActiveSlotCoeff

Active slots coefficient. This parameter represents the proportion of slots in which blocks should be issued. This can be interpreted as the probability that a party holding all the stake will be elected as leader for a given slot.

tpraosSecurityParam ∷ !SecurityParam

See securityParameter.

tpraosMaxKESEvo ∷ !Word64

Maximum number of KES iterations, see maxKESEvo.

tpraosQuorum ∷ !Word64

Quorum for update system votes and MIR certificates, see quorum.

tpraosMaxMajorPV ∷ !MaxMajorProtVer

All blocks invalid after this protocol version, see maxMajorPV.

tpraosMaxLovelaceSupply ∷ !Word64

Maximum number of lovelace in the system, see maxLovelaceSupply.

tpraosNetworkId ∷ !Network

Testnet or mainnet?

tpraosInitialNonce ∷ !Nonce

Initial nonce used for the TPraos protocol state. Typically this is derived from the hash of the Shelley genesis config JSON file, but different values may be used for testing purposes.

NOTE: this is only used when translating the Byron ChainDepState to the Shelley ChainDepState, at which point we'll need access to the initial nonce at runtime. TODO #2326.

tpraosSystemStart ∷ !SystemStart

The system start, as projected from the chain's genesis block.

data TPraosState c

Transitional Praos consensus state.

In addition to the ChainDepState provided by the ledger, we track the slot number of the last applied header.

Constructors

TPraosState

Fields

tpraosStateLastSlot ∷ !(WithOrigin SlotNo)

tpraosStateChainDepState ∷ !(ChainDepState c)

data TPraosToSign c

Fields arising from transitional praos execution which must be included in the block signature.

Constructors

TPraosToSign

Fields

tpraosToSignIssuerVK ∷ VKey 'BlockIssuer c

Verification key for the issuer of this block.

Note that unlike in Classic/BFT where we have a key for the genesis delegate on whose behalf we are issuing this block, this key corresponds to the stake pool/core node actually forging the block.

tpraosToSignVrfVK ∷ VerKeyVRF c

tpraosToSignEta ∷ CertifiedVRF c Nonce

Verifiable result containing the updated nonce value.

tpraosToSignLeader ∷ CertifiedVRF c Natural

Verifiable proof of the leader value, used to determine whether the node has the right to issue a block in this slot.

We include a value here even for blocks forged under the BFT schedule. It is not required that such a value be verifiable (though by default it will be verifiably correct, but unused.)

tpraosToSignOCert ∷ OCert c

Lightweight delegation certificate mapping the cold (DSIGN) key to the online KES key.

type TPraosValidateView c = BHeader c

Because we are using the executable spec, rather than implementing the protocol directly here, we have a fixed header type rather than an abstraction. So our validate view is fixed to this.

forgeTPraosFields ∷ (PraosCrypto c, KESignable c toSign, Monad m) ⇒ HotKey c m → CanBeLeader (TPraos c) → IsLeader (TPraos c) → (TPraosToSign c → toSign) → m (TPraosFields c toSign)

mkShelleyGlobals ∷ ConsensusConfig (TPraos c) → Globals

mkTPraosParams

Arguments

∷ MaxMajorProtVer

→ Nonce
Initial nonce

→ ShelleyGenesis era

→ TPraosParams

Crypto

class (Crypto c, DSignable c (OCertSignable c), DSignable c (Hash c EraIndependentTxBody), KESignable c (BHBody c), VRFSignable c Seed) ⇒ PraosCrypto c

data StandardCrypto

CannotForge

data TPraosCannotForge c

Expresses that, whilst we believe ourselves to be a leader for this slot, we are nonetheless unable to forge a block.

Constructors

TPraosCannotForgeKeyNotUsableYet
The KES key in our operational certificate can't be used because the current (wall clock) period is before the start period of the key. current KES period.

Note: the opposite case, i.e., the wall clock period being after the end period of the key, is caught when trying to update the key in updateForgeState.

Fields

!KESPeriod

Current KES period according to the wallclock slot, i.e., the KES period in which we want to use the key.

!KESPeriod

Start KES period of the KES key.

TPraosCannotForgeWrongVRF !(Hash c (VerKeyVRF c)) !(Hash c (VerKeyVRF c))
We are a genesis delegate, but our VRF key (second argument) does not match the registered key for that delegate (first argument).

tpraosCheckCanForge

Arguments

∷ ConsensusConfig (TPraos c)

→ Hash c (VerKeyVRF c)
Precomputed hash of the VRF verification key

→ SlotNo

→ IsLeader (TPraos c)

→ KESInfo

→ Either (TPraosCannotForge c) ()

Type instances

data family ConsensusConfig p

Static configuration required to run the consensus protocol

Every method in the ConsensusProtocol class takes the consensus configuration as a parameter, so having this as a data family rather than a type family resolves most ambiguity.

Defined out of the class so that protocols can define this type without having to define the entire protocol at the same time (or indeed in the same module).

data family Ticked st

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.