Ouroboros.Consensus.Shelley.Protocol.HotKey

Description

Hot key

Intended for qualified import

KES Info

type KESEvolution = Period

We call the relative periods that a KES key is valid its evolution, to avoid confusion with absolute periods.

data KESInfo

Constructors

KESInfo

Fields

kesStartPeriod ∷ !KESPeriod

kesEndPeriod ∷ !KESPeriod

Currently derived from TPraosParams: > kesEndPeriod = kesStartPeriod + tpraosMaxKESEvo

kesEvolution ∷ !KESEvolution

Current evolution or relative period.

Invariant: > kesStartPeriod + kesEvolution in [kesStartPeriod, kesEndPeriod)

kesAbsolutePeriod ∷ KESInfo → KESPeriod

Return the absolute KES period

KES Status

data KESStatus

Constructors

BeforeKESStart
The given period is before the start period of the KES key.

Fields

KESPeriod

Given period

KESPeriod

Start period of the KES key

InKESRange KESEvolution
Relative period or evolution corresponding to the given absolute period

AfterKESEnd
The given period is after the end period of the KES key.

Fields

KESPeriod

Given period

KESPeriod

End period of the KES key

kesStatus ∷ KESInfo → KESPeriod → KESStatus

Return the evolution of the given KES period, when it falls within the range of the HotKey ([hkStart, hkEnd)).

Note that the upper bound is exclusive, the spec says: > c0 <= kesPeriod s < c0 + MaxKESEvo

Hot Key

data KESEvolutionError

Failed to evolve the KES key.

Constructors

KESCouldNotEvolve KESInfo KESPeriod
The KES key could not be evolved to the target period.

KESKeyAlreadyPoisoned KESInfo KESPeriod
Target period outside the range of the current KES key. Typically the current KES period according to the wallclock slot.

type KESEvolutionInfo = UpdateInfo KESInfo KESInfo KESEvolutionError

Result of evolving the KES key.

data HotKey era m

API to interact with the key.

Constructors

HotKey

Fields

evolve ∷ KESPeriod → m KESEvolutionInfo

Evolve the KES signing key to the given absolute KES period.

When the key cannot evolve anymore, we poison it.

getInfo ∷ m KESInfo

Return KESInfo of the signing key.

isPoisoned ∷ m Bool

Return True when the signing key is poisoned because it expired.

sign_ ∷ ∀ toSign. (KESignable era toSign, HasCallStack) ⇒ toSign → m (SignedKES era toSign)

Sign the given toSign with the current signing key.

PRECONDITION: the key is not poisoned.

POSTCONDITION: the signature is in normal form.

sign ∷ (KESignable era toSign, HasCallStack) ⇒ HotKey era m → toSign → m (SignedKES era toSign)

mkHotKey

Arguments

∷ (Era era, IOLike m)

⇒ SignKeyKES era

→ KESPeriod
Start period

→ Word64
Max KES evolutions

→ m (HotKey era m)