Ouroboros.Consensus.Shelley.Protocol.Util

Description

Assorted utility functions for Shelley protocol integration.

In particular, various things we need for integration with the delegation package from cardano-ledger-specs.

Documentation

isNewEpoch

Arguments

∷ EpochInfo Identity

→ SlotNo
Slot we want to check

→ WithOrigin SlotNo
Slot we are comparing a new epoch against

→ Bool

Verify whether a slot represents a change to a new epoch with regard to some other slot.

prtclStateHash ∷ State (PRTCL c) → WithOrigin (HashHeader c)

prtclStateSlot ∷ State (PRTCL c) → WithOrigin SlotNo

prtclStateEta0 ∷ State (PRTCL c) → Nonce