Ouroboros.Consensus.Shelley.ShelleyBased

Injection from Shelley-based eras into the Cardano eras

class consensusModeEra ~ ShelleyBlock shelleyEra ⇒ InjectShelley shelleyEra consensusModeEra

Witness the relation between consensus mode (e.g. Cardano) eras and the Shelley-based eras.

injectShelleyNP ∷ AllZip InjectShelley shelleyEras consensusModeEras ⇒ (∀ shelleyEra consensusModeEra. InjectShelley shelleyEra consensusModeEra ⇒ f shelleyEra → g consensusModeEra) → NP f shelleyEras → NP g consensusModeEras

injectShelleyOptNP ∷ AllZip InjectShelley shelleyEras consensusModeEras ⇒ (∀ shelleyEra consensusModeEra. InjectShelley shelleyEra consensusModeEra ⇒ f shelleyEra → g consensusModeEra) → OptNP empty f shelleyEras → OptNP empty g consensusModeEras

Transform Shelley-based types

class EraCrypto era ~ c ⇒ HasCrypto c era

Witness the relation between the crypto used by a Shelley-based era.

Can be partially applied while an equality constraint cannot.

type EraCrypto era = Crypto era

The Crypto type family conflicts with the Crypto class. To avoid having to import one or both of them qualified, define EraCrypto as an alias of the former: /return the crypto used by this era/.