Ouroboros.Consensus.Shelley.Eras

Eras based on the Shelley ledger

data ShelleyEra c

type AllegraEra = ShelleyMAEra 'Allegra

type MaryEra = ShelleyMAEra 'Mary

Eras instantiated with standard crypto

type StandardShelley = ShelleyEra StandardCrypto

The Shelley era with standard crypto

type StandardAllegra = AllegraEra StandardCrypto

The Allegra era with standard crypto

type StandardMary = MaryEra StandardCrypto

The Mary era with standard crypto

Shelley-based era

class (PraosCrypto (Crypto era), ShelleyBased era, GetLedgerView era, ApplyBlock era, ApplyTx era) ⇒ ShelleyBasedEra era

Type synonyms for convenience

type EraCrypto era = Crypto era

The Crypto type family conflicts with the Crypto class. To avoid having to import one or both of them qualified, define EraCrypto as an alias of the former: /return the crypto used by this era/.

Re-exports

data StandardCrypto