Ouroboros.Consensus.Shelley.Ledger

Documentation

data family BlockQuery blk ∷ Type → Type

Different queries supported by the ledger, indexed by the result type.

data family GenTx blk

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data family TxId tx

A generalized transaction, GenTx, identifier.

data family Validated x

" Validated " transaction or block

The ledger defines how to validate transactions and blocks. It's possible the type before and after validation may be distinct (eg Alonzo transactions), which originally motivated this family.

We also gain the related benefit that certain interface functions, such as those that reapply blocks, can have a more precise type now. TODO

Similarly, the Node-to-Client mini protocols can explicitly indicate that the client trusts the blocks from the local server, by having the server send Validated blocks to the client. TODO

Note that validation has different implications for a transaction than for a block. In particular, a validated transaction can be " reapplied " to different ledger states, whereas a validated block must only be " reapplied " to the exact same ledger state (eg as part of rebuilding from an on-disk ledger snapshot).

Since the ledger defines validation, see the ledger details for concrete examples of what determines the validity (wrt to a LedgerState) of a transaction and/or block. Example properties include: a transaction's claimed inputs exist and are still unspent, a block carries a sufficient cryptographic signature, etc.

data family LedgerState blk

Ledger state associated with a block

data family BlockConfig blk

Static configuration required to work with this type of blocks

data family CodecConfig blk

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data family StorageConfig blk

Config needed for the NodeInitStorage class. Defined here to avoid circular dependencies.

data family Header blk

class HasHeader (Header blk) ⇒ GetHeader blk where

Methods

getHeader ∷ blk → Header blk

blockMatchesHeader ∷ Header blk → blk → Bool

Check whether the header is the header of the block.

For example, by checking whether the hash of the body stored in the header matches that of the block.

headerIsEBB ∷ Header blk → Maybe EpochNo

When the given header is the header of an Epoch Boundary Block, returns its epoch number.

data family NestedCtxt_ blk ∷ (Type → Type) → Type → Type

Context identifying what kind of block we have

In almost all places we will use NestedCtxt rather than NestedCtxt_.

data family Ticked st

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

newtype ApplyTxError era

Constructors

ApplyTxError [PredicateFailure (EraRule "LEDGER" era)]

class (EraSegWits era, EraGovernance era, ApplyTx era, ApplyBlock era, CanStartFromGenesis era, GetLedgerView era, NoThunks (StashedAVVMAddresses era), EncCBOR (StashedAVVMAddresses era), DecCBOR (StashedAVVMAddresses era), Show (StashedAVVMAddresses era), Eq (StashedAVVMAddresses era), DecCBOR (PredicateFailure (EraRule "LEDGER" era)), EncCBOR (PredicateFailure (EraRule "LEDGER" era)), DecCBOR (PredicateFailure (EraRule "DELEGS" era)), EncCBOR (PredicateFailure (EraRule "DELEGS" era)), DecCBOR (PredicateFailure (EraRule "UTXOW" era)), EncCBOR (PredicateFailure (EraRule "UTXOW" era)), DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody), NoThunks (PredicateFailure (EraRule "BBODY" era)), NoThunks (TranslationContext era)) ⇒ ShelleyBasedEra era

Consensus often needs some more functionality than the ledger currently provides.

Either the functionality shouldn't or can't live in the ledger, in which case it can be part and remain part of ShelleyBasedEra. Or, the functionality should live in the ledger, but hasn't yet been added to the ledger, or it hasn't yet been propagated to this repository, in which case it can be added to this class until that is the case.

If this class becomes redundant, We can move it to ledger and re-export it from here.

TODO Currently we include some constraints on the update state which are needed to determine the hard fork point. In the future this should be replaced with an appropriate API - see https://github.com/input-output-hk/ouroboros-network/issues/2890

Minimal complete definition

shelleyBasedEraName, applyShelleyBasedTx

verifyHeaderIntegrity

Arguments

∷ ProtocolHeaderSupportsKES proto

⇒ Word64
Slots per KES period

→ ShelleyProtocolHeader proto

→ Bool

Verify that the signature on a header is correct and valid.

newtype ShelleyHash crypto

Constructors

ShelleyHash

Fields

unShelleyHash ∷ Hash crypto EraIndependentBlockHeader

data ShelleyBlock proto era

Shelley-based block type.

This block is parametrised over both the (ledger) era and the protocol.

Constructors

ShelleyBlock

Fields

shelleyBlockRaw ∷ !(Block (ShelleyProtocolHeader proto) era)

shelleyBlockHeaderHash ∷ !(ShelleyHash (ProtoCrypto proto))

class (ShelleyBasedEra era, ShelleyProtocol proto, Eq (ShelleyProtocolHeader proto), Show (ShelleyProtocolHeader proto), NoThunks (ShelleyProtocolHeader proto), EncCBOR (ShelleyProtocolHeader proto), DecCBOR (Annotator (ShelleyProtocolHeader proto)), Show (CannotForgeError proto), SelectView proto ~ PraosChainSelectView (EraCrypto era), SignedHeader (ShelleyProtocolHeader proto), DecodeDisk (ShelleyBlock proto era) (ChainDepState proto), EncodeDisk (ShelleyBlock proto era) (ChainDepState proto), EraCrypto era ~ ProtoCrypto proto, HasPartialConsensusConfig proto, DecCBOR (PState (ProtoCrypto proto))) ⇒ ShelleyCompatible proto era

mkShelleyBlock ∷ ShelleyCompatible proto era ⇒ Block (ShelleyProtocolHeader proto) era → ShelleyBlock proto era

mkShelleyHeader ∷ ShelleyCompatible proto era ⇒ ShelleyProtocolHeader proto → Header (ShelleyBlock proto era)

fromShelleyPrevHash ∷ EraCrypto era ~ ProtoCrypto proto ⇒ PrevHash (EraCrypto era) → ChainHash (ShelleyBlock proto era)

From cardano-ledger-specs to ouroboros-consensus

toShelleyPrevHash ∷ EraCrypto era ~ ProtoCrypto proto ⇒ ChainHash (Header (ShelleyBlock proto era)) → PrevHash (EraCrypto era)

From ouroboros-consensus to cardano-ledger-specs

encodeShelleyBlock ∷ ∀ proto era. ShelleyCompatible proto era ⇒ ShelleyBlock proto era → Encoding

decodeShelleyBlock ∷ ∀ proto era. ShelleyCompatible proto era ⇒ ∀ s. Decoder s (ByteString → ShelleyBlock proto era)

shelleyBinaryBlockInfo ∷ ∀ proto era. ShelleyCompatible proto era ⇒ ShelleyBlock proto era → BinaryBlockInfo

encodeShelleyHeader ∷ ∀ proto era. ShelleyCompatible proto era ⇒ Header (ShelleyBlock proto era) → Encoding

decodeShelleyHeader ∷ ∀ proto era. ShelleyCompatible proto era ⇒ ∀ s. Decoder s (ByteString → Header (ShelleyBlock proto era))

data ShelleyNodeToClientVersion

Constructors

ShelleyNodeToClientVersion1

ShelleyNodeToClientVersion2
New queries introduced

ShelleyNodeToClientVersion3
New query introduced

ShelleyNodeToClientVersion4
New queries introduced

ShelleyNodeToClientVersion5
New queries introduced: GetRewardInfoPools

ShelleyNodeToClientVersion6
New queries introduced: GetPoolDistr, GetPoolState, GetStakeSnapshots

data ShelleyNodeToNodeVersion

Constructors

ShelleyNodeToNodeVersion1

verifyBlockIntegrity ∷ ShelleyCompatible proto era ⇒ Word64 → ShelleyBlock proto era → Bool

Verifies whether the block is not corrupted by checking its signature and witnesses.

data CompactGenesis c

Compact variant of ShelleyGenesis with some fields erased that are only used on start-up and that should not be kept in memory forever.

Concretely:

The sgInitialFunds field is erased. It is only used to set up the initial UTxO in tests and testnets.

The sgStaking field is erased. It is only used to register initial stake pools in tests and benchmarks.

mkShelleyBlockConfig ∷ ShelleyBasedEra era ⇒ ProtVer → ShelleyGenesis (EraCrypto era) → [VKey 'BlockIssuer (EraCrypto era)] → BlockConfig (ShelleyBlock proto era)

compactGenesis ∷ ShelleyGenesis c → CompactGenesis c

Compacts the given ShelleyGenesis.

data ShelleyReapplyException

Constructors

∀ era.Show (BlockTransitionError era) ⇒ ShelleyReapplyException (BlockTransitionError era)

data ShelleyLedgerEvent era

All events emitted by the Shelley ledger API

Constructors

ShelleyLedgerEventBBODY (Event (EraRule "BBODY" era))
An event emitted when (re)applying a block

ShelleyLedgerEventTICK (Event (EraRule "TICK" era))
An event emitted during the chain tick

newtype ShelleyTransition

Information required to determine the hard fork point from Shelley to the next ledger

Constructors

ShelleyTransitionInfo

Fields

shelleyAfterVoting ∷ Word32

The number of blocks in this epoch past the voting deadline

We record this to make sure that we can tell the HFC about hard forks if and only if we are certain:

Blocks that came in within an epoch after the 4k/f voting deadline are not relevant (10kf - 2 * 3kf).

Since there are slots between blocks, we are probably only sure that there will be no more relevant block when we have seen the first block after the deadline.

If we count how many blocks we have seen post deadline, and we have reached k of them, we know that that last pre-deadline block won't be rolled back anymore.

At this point we can look at the ledger state and see which proposals we accepted in the voting period, if any, and notify the HFC is one of them indicates a transition.

data ShelleyTip proto era

Constructors

ShelleyTip

Fields

shelleyTipSlotNo ∷ !SlotNo

shelleyTipBlockNo ∷ !BlockNo

shelleyTipHash ∷ !(HeaderHash (ShelleyBlock proto era))

data ShelleyLedgerConfig era

Constructors

ShelleyLedgerConfig

Fields

shelleyLedgerCompactGenesis ∷ !(CompactGenesis (EraCrypto era))

shelleyLedgerGlobals ∷ !Globals

Derived from shelleyLedgerGenesis but we store a cached version because it used very often.

shelleyLedgerTranslationContext ∷ !(TranslationContext era)

newtype ShelleyLedgerError era

Constructors

BBodyError (BlockTransitionError era)

shelleyLedgerGenesis ∷ ShelleyLedgerConfig era → ShelleyGenesis (EraCrypto era)

shelleyEraParams ∷ ShelleyGenesis c → EraParams

shelleyEraParamsNeverHardForks ∷ ShelleyGenesis c → EraParams

Separate variant of shelleyEraParams to be used for a Shelley-only chain.

mkShelleyLedgerConfig ∷ ShelleyGenesis (EraCrypto era) → TranslationContext era → EpochInfo (Except PastHorizonException) → MaxMajorProtVer → ShelleyLedgerConfig era

shelleyTipToPoint ∷ WithOrigin (ShelleyTip proto era) → Point (ShelleyBlock proto era)

castShelleyTip ∷ HeaderHash (ShelleyBlock proto era) ~ HeaderHash (ShelleyBlock proto' era') ⇒ ShelleyTip proto era → ShelleyTip proto' era'

shelleyLedgerTipPoint ∷ LedgerState (ShelleyBlock proto era) → Point (ShelleyBlock proto era)

getPParams ∷ NewEpochState era → PParams era

encodeShelleyAnnTip ∷ ShelleyCompatible proto era ⇒ AnnTip (ShelleyBlock proto era) → Encoding

decodeShelleyAnnTip ∷ ShelleyCompatible proto era ⇒ Decoder s (AnnTip (ShelleyBlock proto era))

encodeShelleyHeaderState ∷ ShelleyCompatible proto era ⇒ HeaderState (ShelleyBlock proto era) → Encoding

encodeShelleyLedgerState ∷ ShelleyCompatible proto era ⇒ LedgerState (ShelleyBlock proto era) → Encoding

decodeShelleyLedgerState ∷ ∀ era proto s. ShelleyCompatible proto era ⇒ Decoder s (LedgerState (ShelleyBlock proto era))

data StakeSnapshots crypto

Constructors

StakeSnapshots

Fields

ssStakeSnapshots ∷ !(Map (KeyHash 'StakePool crypto) (StakeSnapshot crypto))

ssMarkTotal ∷ !Coin

ssSetTotal ∷ !Coin

ssGoTotal ∷ !Coin

data StakeSnapshot crypto

The stake snapshot returns information about the mark, set, go ledger snapshots for a pool, plus the total active stake for each snapshot that can be used in a sigma calculation.

Each snapshot is taken at the end of a different era. The go snapshot is the current one and was taken two epochs earlier, set was taken one epoch ago, and mark was taken immediately before the start of the current epoch.

Constructors

StakeSnapshot

Fields

ssMarkPool ∷ !Coin

ssSetPool ∷ !Coin

ssGoPool ∷ !Coin

newtype NonMyopicMemberRewards c

Constructors

NonMyopicMemberRewards

Fields

unNonMyopicMemberRewards ∷ Map (Either Coin (Credential 'Staking c)) (Map (KeyHash 'StakePool c) Coin)

querySupportedVersion ∷ BlockQuery (ShelleyBlock proto era) result → ShelleyNodeToClientVersion → Bool

Is the given query supported by the given ShelleyNodeToClientVersion?

encodeShelleyQuery ∷ ∀ era proto result. ShelleyBasedEra era ⇒ BlockQuery (ShelleyBlock proto era) result → Encoding

decodeShelleyQuery ∷ ∀ era proto. ShelleyBasedEra era ⇒ ∀ s. Decoder s (SomeSecond BlockQuery (ShelleyBlock proto era))

encodeShelleyResult ∷ ∀ proto era result. ShelleyCompatible proto era ⇒ BlockQuery (ShelleyBlock proto era) result → result → Encoding

decodeShelleyResult ∷ ∀ proto era result. ShelleyCompatible proto era ⇒ BlockQuery (ShelleyBlock proto era) result → ∀ s. Decoder s result

data WithTop a

Add a unique top element to a lattice.

TODO This should be relocated to `cardano-base:Data.Measure'.

Constructors

NotTop a

Top

data AlonzoMeasure

Constructors

AlonzoMeasure

Fields

byteSize ∷ !ByteSize

exUnits ∷ !(ExUnits' (WithTop Natural))

fixedBlockBodyOverhead ∷ Num a ⇒ a

txInBlockSize is used to estimate how many transactions we can grab from the Mempool to put into the block we are going to forge without exceeding the maximum block body size according to the ledger. If we exceed that limit, we will have forged a block that is invalid according to the ledger. We ourselves won't even adopt it, causing us to lose our slot, something we must try to avoid.

For this reason it is better to overestimate the size of a transaction than to underestimate. The only downside is that we maybe could have put one (or more?) transactions extra in that block.

As the sum of the serialised transaction sizes is not equal to the size of the serialised block body (TxSeq) consisting of those transactions (see cardano-node#1545 for an example), we account for some extra overhead per transaction as a safety margin.

Also see perTxOverhead.

perTxOverhead ∷ Num a ⇒ a

See fixedBlockBodyOverhead.

mkShelleyTx ∷ ∀ era proto. ShelleyBasedEra era ⇒ Tx era → GenTx (ShelleyBlock proto era)

mkShelleyValidatedTx ∷ ∀ era proto. ShelleyBasedEra era ⇒ Validated (Tx era) → Validated (GenTx (ShelleyBlock proto era))

fromExUnits ∷ ExUnits → ExUnits' (WithTop Natural)

forgeShelleyBlock

Arguments

∷ ∀ m era proto. (ShelleyCompatible proto era, TxLimits (ShelleyBlock proto era), Monad m)

⇒ HotKey (EraCrypto era) m

→ CanBeLeader proto

→ TopLevelConfig (ShelleyBlock proto era)

→ TxOverrides (ShelleyBlock proto era)
How to override max tx capacity defined by ledger

→ BlockNo
Current block number

→ SlotNo
Current slot number

→ TickedLedgerState (ShelleyBlock proto era)
Current ledger

→ [Validated (GenTx (ShelleyBlock proto era))]
Txs to add in the block

→ IsLeader proto

→ m (ShelleyBlock proto era)