Ouroboros.Consensus.Shelley.Ledger.Block

Documentation

class HasHeader (Header blk) ⇒ GetHeader blk where

Methods

getHeader ∷ blk → Header blk

blockMatchesHeader ∷ Header blk → blk → Bool

Check whether the header is the header of the block.

For example, by checking whether the hash of the body stored in the header matches that of the block.

headerIsEBB ∷ Header blk → Maybe EpochNo

When the given header is the header of an Epoch Boundary Block, returns its epoch number.

data family Header blk

data family NestedCtxt_ blk ∷ (Type → Type) → Type → Type

Context identifying what kind of block we have

In almost all places we will use NestedCtxt rather than NestedCtxt_.

class (EraSegWits era, EraGovernance era, ApplyTx era, ApplyBlock era, CanStartFromGenesis era, GetLedgerView era, NoThunks (StashedAVVMAddresses era), EncCBOR (StashedAVVMAddresses era), DecCBOR (StashedAVVMAddresses era), Show (StashedAVVMAddresses era), Eq (StashedAVVMAddresses era), DecCBOR (PredicateFailure (EraRule "LEDGER" era)), EncCBOR (PredicateFailure (EraRule "LEDGER" era)), DecCBOR (PredicateFailure (EraRule "DELEGS" era)), EncCBOR (PredicateFailure (EraRule "DELEGS" era)), DecCBOR (PredicateFailure (EraRule "UTXOW" era)), EncCBOR (PredicateFailure (EraRule "UTXOW" era)), DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody), NoThunks (PredicateFailure (EraRule "BBODY" era)), NoThunks (TranslationContext era)) ⇒ ShelleyBasedEra era

Consensus often needs some more functionality than the ledger currently provides.

Either the functionality shouldn't or can't live in the ledger, in which case it can be part and remain part of ShelleyBasedEra. Or, the functionality should live in the ledger, but hasn't yet been added to the ledger, or it hasn't yet been propagated to this repository, in which case it can be added to this class until that is the case.

If this class becomes redundant, We can move it to ledger and re-export it from here.

TODO Currently we include some constraints on the update state which are needed to determine the hard fork point. In the future this should be replaced with an appropriate API - see https://github.com/input-output-hk/ouroboros-network/issues/2890

Minimal complete definition

shelleyBasedEraName, applyShelleyBasedTx

data ShelleyBlock proto era

Shelley-based block type.

This block is parametrised over both the (ledger) era and the protocol.

Constructors

ShelleyBlock

Fields

shelleyBlockRaw ∷ !(Block (ShelleyProtocolHeader proto) era)

shelleyBlockHeaderHash ∷ !(ShelleyHash (ProtoCrypto proto))

newtype ShelleyHash crypto

Constructors

ShelleyHash

Fields

unShelleyHash ∷ Hash crypto EraIndependentBlockHeader

Shelley Compatibility

class (ShelleyBasedEra era, ShelleyProtocol proto, Eq (ShelleyProtocolHeader proto), Show (ShelleyProtocolHeader proto), NoThunks (ShelleyProtocolHeader proto), EncCBOR (ShelleyProtocolHeader proto), DecCBOR (Annotator (ShelleyProtocolHeader proto)), Show (CannotForgeError proto), SelectView proto ~ PraosChainSelectView (EraCrypto era), SignedHeader (ShelleyProtocolHeader proto), DecodeDisk (ShelleyBlock proto era) (ChainDepState proto), EncodeDisk (ShelleyBlock proto era) (ChainDepState proto), EraCrypto era ~ ProtoCrypto proto, HasPartialConsensusConfig proto, DecCBOR (PState (ProtoCrypto proto))) ⇒ ShelleyCompatible proto era

mkShelleyBlock ∷ ShelleyCompatible proto era ⇒ Block (ShelleyProtocolHeader proto) era → ShelleyBlock proto era

mkShelleyHeader ∷ ShelleyCompatible proto era ⇒ ShelleyProtocolHeader proto → Header (ShelleyBlock proto era)

Serialisation

decodeShelleyBlock ∷ ∀ proto era. ShelleyCompatible proto era ⇒ ∀ s. Decoder s (ByteString → ShelleyBlock proto era)

decodeShelleyHeader ∷ ∀ proto era. ShelleyCompatible proto era ⇒ ∀ s. Decoder s (ByteString → Header (ShelleyBlock proto era))

encodeShelleyBlock ∷ ∀ proto era. ShelleyCompatible proto era ⇒ ShelleyBlock proto era → Encoding

encodeShelleyHeader ∷ ∀ proto era. ShelleyCompatible proto era ⇒ Header (ShelleyBlock proto era) → Encoding

shelleyBinaryBlockInfo ∷ ∀ proto era. ShelleyCompatible proto era ⇒ ShelleyBlock proto era → BinaryBlockInfo

Conversion

fromShelleyPrevHash ∷ EraCrypto era ~ ProtoCrypto proto ⇒ PrevHash (EraCrypto era) → ChainHash (ShelleyBlock proto era)

From cardano-ledger-specs to ouroboros-consensus

toShelleyPrevHash ∷ EraCrypto era ~ ProtoCrypto proto ⇒ ChainHash (Header (ShelleyBlock proto era)) → PrevHash (EraCrypto era)

From ouroboros-consensus to cardano-ledger-specs