Ouroboros.Consensus.Shelley.Ledger.Block

Documentation

class HasHeader (Header blk) ⇒ GetHeader blk where

Methods

getHeader ∷ blk → Header blk

blockMatchesHeader ∷ Header blk → blk → Bool

Check whether the header is the header of the block.

For example, by checking whether the hash of the body stored in the header matches that of the block.

headerIsEBB ∷ Header blk → Maybe EpochNo

When the given header is the header of an Epoch Boundary Block, returns its epoch number.

data family Header blk

data family NestedCtxt_ blk ∷ (Type → Type) → Type → Type

Context identifying what kind of block we have

In almost all places we will use NestedCtxt rather than NestedCtxt_.

class (ShelleyBasedEra era, PraosCrypto (EraCrypto era), GetLedgerView era, State (EraRule "PPUP" era) ~ PPUPState era, Default (State (EraRule "PPUP" era)), HasField "_maxBHSize" (PParams era) Natural, HasField "_maxTxSize" (PParams era) Natural, HasField "_a0" (PParams era) NonNegativeInterval, HasField "_nOpt" (PParams era) Natural, HasField "_rho" (PParams era) UnitInterval, HasField "_tau" (PParams era) UnitInterval, ValidateScript era, FromCBOR (PParams era), ToCBOR (PParams era), HasField "_protocolVersion" (PParamsDelta era) (StrictMaybe ProtVer), FromCBOR (PParamsDelta era), AdditionalGenesisConfig era ~ TranslationContext era, ToCBORGroup (TxSeq era), NoThunks (TranslationContext era), FromCBOR (Annotator (Witnesses era)), ToCBOR (Witnesses era), Eq (TxSeq era), Show (TxSeq era), FromCBOR (Annotator (TxSeq era))) ⇒ ShelleyBasedEra era

The ledger already defines ShelleyBasedEra as the top-level constraint on an era, however, consensus often needs some more functionality than the ledger currently provides.

Either the functionality shouldn't or can't live in the ledger, in which case it can be part and remain part of ShelleyBasedEra. Or, the functionality should live in the ledger, but hasn't yet been added to the ledger, or it hasn't yet been propagated to this repository, in which case it can be added to this class until that is the case.

By having the same name as the class defined in ledger, we can, if this class becomes redundant, switch to the ledger-defined one without having to update all the code using it. We can just export the right one from this module.

TODO Currently we include some constraints on the update state which are needed to determine the hard fork point. In the future this should be replaced with an appropriate API - see https://github.com/input-output-hk/ouroboros-network/issues/2890

Minimal complete definition

shelleyBasedEraName, applyShelleyBasedTx

data ShelleyBlock era

Newtype wrapper to avoid orphan instances

The phantom type parameter is there to record the additional information we need to work with this block. Most of the code here does not care, but we may need different additional information when running the chain.

Constructors

ShelleyBlock

Fields

shelleyBlockRaw ∷ !(Block (BHeader (EraCrypto era)) era)

shelleyBlockHeaderHash ∷ !(ShelleyHash (EraCrypto era))

newtype ShelleyHash c

Constructors

ShelleyHash

Fields

unShelleyHash ∷ HashHeader c

mkShelleyBlock ∷ ShelleyBasedEra era ⇒ Block (BHeader (EraCrypto era)) era → ShelleyBlock era

mkShelleyHeader ∷ ShelleyBasedEra era ⇒ BHeader (EraCrypto era) → Header (ShelleyBlock era)

Serialisation

decodeShelleyBlock ∷ ShelleyBasedEra era ⇒ Decoder s (ByteString → ShelleyBlock era)

decodeShelleyHeader ∷ ShelleyBasedEra era ⇒ Decoder s (ByteString → Header (ShelleyBlock era))

encodeShelleyBlock ∷ ShelleyBasedEra era ⇒ ShelleyBlock era → Encoding

encodeShelleyHeader ∷ ShelleyBasedEra era ⇒ Header (ShelleyBlock era) → Encoding

shelleyBinaryBlockInfo ∷ ShelleyBasedEra era ⇒ ShelleyBlock era → BinaryBlockInfo

Conversion

fromShelleyPrevHash ∷ PrevHash (EraCrypto era) → ChainHash (ShelleyBlock era)

From cardano-ledger-specs to ouroboros-consensus

toShelleyPrevHash ∷ ChainHash (Header (ShelleyBlock era)) → PrevHash (EraCrypto era)

From ouroboros-consensus to cardano-ledger-specs