Ouroboros.Consensus.Shelley.Ledger.Block

Documentation

newtype ShelleyHash c

Constructors

ShelleyHash

Fields

unShelleyHash ∷ HashHeader c

shelleyHashInfo ∷ ∀ c. Crypto c ⇒ HashInfo (ShelleyHash c)

data ShelleyBlock c

Newtype wrapper to avoid orphan instances

The phantom type parameter is there to record the additional information we need to work with this block. Most of the code here does not care, but we may need different additional information when running the chain.

Constructors

ShelleyBlock

Fields

shelleyBlockRaw ∷ !(Block c)

shelleyBlockHeaderHash ∷ !(ShelleyHash c)

mkShelleyBlock ∷ Crypto c ⇒ Block c → ShelleyBlock c

class GetHeader blk where

Associated Types

data Header blk ∷ Type

Methods

getHeader ∷ blk → Header blk

data family Header blk ∷ Type

mkShelleyHeader ∷ Crypto c ⇒ BHeader c → Header (ShelleyBlock c)

encodeShelleyBlockWithInfo ∷ Crypto c ⇒ ShelleyBlock c → BinaryInfo Encoding

shelleyAddHeaderEnvelope ∷ ByteString → ByteString

When given the raw header bytes extracted from the block (see encodeShelleyBlockWithInfo), we have a bytestring exactly corresponding to the header, no modifications needed.

Conversion

fromShelleyPrevHash ∷ PrevHash c → ChainHash (Header (ShelleyBlock c))

From cardano-ledger-specs to ouroboros-consensus

toShelleyPrevHash ∷ ChainHash (Header (ShelleyBlock c)) → PrevHash c

From ouroboros-consensus to cardano-ledger-specs

Re-exported for convenience

class (HashAlgorithm (HASH c), DSIGNAlgorithm (DSIGN c), KESAlgorithm (KES c), VRFAlgorithm (VRF c), ContextDSIGN (DSIGN c) ~ (), ContextKES (KES c) ~ (), ContextVRF (VRF c) ~ (), Typeable c) ⇒ Crypto c