Ouroboros.Consensus.Shelley.Ledger.Config

Documentation

data family BlockConfig blk

Static configuration required to work with this type of blocks

data family CodecConfig blk

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data family StorageConfig blk

Config needed for the NodeInitStorage class. Defined here to avoid circular dependencies.

compactGenesis ∷ ShelleyGenesis era → CompactGenesis era

Compacts the given ShelleyGenesis.

getCompactGenesis ∷ CompactGenesis era → ShelleyGenesis era

mkShelleyBlockConfig ∷ ShelleyBasedEra era ⇒ ProtVer → ShelleyGenesis era → [VKey 'BlockIssuer (EraCrypto era)] → BlockConfig (ShelleyBlock era)

opaque

data CompactGenesis era

Compact variant of ShelleyGenesis with some fields erased that are only used on start-up and that should not be kept in memory forever.

Concretely:

The sgInitialFunds field is erased. It is only used to set up the initial UTxO in tests and testnets.

The sgStaking field is erased. It is only used to register initial stake pools in tests and benchmarks.

Orphan instances

ShelleyBasedEra era ⇒ Show (BlockConfig (ShelleyBlock era))

Instance details

Methods

showsPrec ∷ Int → BlockConfig (ShelleyBlock era) → ShowS

show ∷ BlockConfig (ShelleyBlock era) → String

showList ∷ [BlockConfig (ShelleyBlock era)] → ShowS

Generic (BlockConfig (ShelleyBlock era))

Instance details

Associated Types

type Rep (BlockConfig (ShelleyBlock era)) ∷ Type → Type

Methods

from ∷ BlockConfig (ShelleyBlock era) → Rep (BlockConfig (ShelleyBlock era)) x

to ∷ Rep (BlockConfig (ShelleyBlock era)) x → BlockConfig (ShelleyBlock era)

Generic (CodecConfig (ShelleyBlock era))

Instance details

Associated Types

type Rep (CodecConfig (ShelleyBlock era)) ∷ Type → Type

Methods

from ∷ CodecConfig (ShelleyBlock era) → Rep (CodecConfig (ShelleyBlock era)) x

to ∷ Rep (CodecConfig (ShelleyBlock era)) x → CodecConfig (ShelleyBlock era)

Generic (StorageConfig (ShelleyBlock era))

Instance details

Associated Types

type Rep (StorageConfig (ShelleyBlock era)) ∷ Type → Type

Methods

from ∷ StorageConfig (ShelleyBlock era) → Rep (StorageConfig (ShelleyBlock era)) x

to ∷ Rep (StorageConfig (ShelleyBlock era)) x → StorageConfig (ShelleyBlock era)

ShelleyBasedEra era ⇒ NoThunks (BlockConfig (ShelleyBlock era))

Instance details

Methods

noThunks ∷ Context → BlockConfig (ShelleyBlock era) → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → BlockConfig (ShelleyBlock era) → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (BlockConfig (ShelleyBlock era)) → String

NoThunks (CodecConfig (ShelleyBlock era))

Instance details

Methods

noThunks ∷ Context → CodecConfig (ShelleyBlock era) → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → CodecConfig (ShelleyBlock era) → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (CodecConfig (ShelleyBlock era)) → String

NoThunks (StorageConfig (ShelleyBlock era))

Instance details

Methods

noThunks ∷ Context → StorageConfig (ShelleyBlock era) → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → StorageConfig (ShelleyBlock era) → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (StorageConfig (ShelleyBlock era)) → String