Ouroboros.Consensus.Shelley.Ledger.Config

Documentation

data family BlockConfig blk ∷ Type

Static configuration required to work with this type of blocks

data family CodecConfig blk ∷ Type

Static configuration required for serialisation and deserialisation of types pertaining to this type of block

Orphan instances

Show (CodecConfig (ShelleyBlock c))

Instance details

Methods

showsPrec ∷ Int → CodecConfig (ShelleyBlock c) → ShowS

show ∷ CodecConfig (ShelleyBlock c) → String

showList ∷ [CodecConfig (ShelleyBlock c)] → ShowS

Show (BlockConfig (ShelleyBlock c))

Instance details

Methods

showsPrec ∷ Int → BlockConfig (ShelleyBlock c) → ShowS

show ∷ BlockConfig (ShelleyBlock c) → String

showList ∷ [BlockConfig (ShelleyBlock c)] → ShowS

Generic (CodecConfig (ShelleyBlock c))

Instance details

Associated Types

type Rep (CodecConfig (ShelleyBlock c)) ∷ Type → Type

Methods

from ∷ CodecConfig (ShelleyBlock c) → Rep (CodecConfig (ShelleyBlock c)) x

to ∷ Rep (CodecConfig (ShelleyBlock c)) x → CodecConfig (ShelleyBlock c)

Generic (BlockConfig (ShelleyBlock c))

Instance details

Associated Types

type Rep (BlockConfig (ShelleyBlock c)) ∷ Type → Type

Methods

from ∷ BlockConfig (ShelleyBlock c) → Rep (BlockConfig (ShelleyBlock c)) x

to ∷ Rep (BlockConfig (ShelleyBlock c)) x → BlockConfig (ShelleyBlock c)

NoUnexpectedThunks (CodecConfig (ShelleyBlock c))

Instance details

Methods

noUnexpectedThunks ∷ [String] → CodecConfig (ShelleyBlock c) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → CodecConfig (ShelleyBlock c) → IO ThunkInfo

showTypeOf ∷ Proxy (CodecConfig (ShelleyBlock c)) → String

NoUnexpectedThunks (BlockConfig (ShelleyBlock c))

Instance details

Methods

noUnexpectedThunks ∷ [String] → BlockConfig (ShelleyBlock c) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → BlockConfig (ShelleyBlock c) → IO ThunkInfo

showTypeOf ∷ Proxy (BlockConfig (ShelleyBlock c)) → String

BlockHasCodecConfig (ShelleyBlock c)

Instance details

Methods

getCodecConfig ∷ BlockConfig (ShelleyBlock c) → CodecConfig (ShelleyBlock c)