Ouroboros.Consensus.Shelley.Ledger.Forge

ForgeState

data TPraosForgeState c

This is the state containing the private key corresponding to the public key that Praos uses to verify headers. That is why we call it TPraosForgeState instead of ShelleyForgeState.

Constructors

TPraosKeyAvailable !(HotKey c)
The online KES key used to sign blocks is available at the given period

TPraosKeyEvolving
The KES key is being evolved by another thread

Any thread that sees this value should back off and retry.

TPraosNoKey
This node is not a core node, it doesn't have the capability to sign blocks.

The ForgeState of such a node will always be ShelleyNoKey.

Forging

forgeShelleyBlock

Arguments

∷ (MonadRandom m, TPraosCrypto c)

⇒ TopLevelConfig (ShelleyBlock c)

→ Update m (ForgeState (ShelleyBlock c))

→ BlockNo
Current block number

→ TickedLedgerState (ShelleyBlock c)
Current ledger

→ [GenTx (ShelleyBlock c)]
Txs to add in the block

→ TPraosProof c
Leader proof (IsLeader)

→ m (ShelleyBlock c)

Orphan instances

TPraosCrypto c ⇒ CanForge (ShelleyBlock c)

Instance details

Associated Types

type ForgeState (ShelleyBlock c) ∷ Type

Methods

forgeBlock ∷ MonadRandom m ⇒ TopLevelConfig (ShelleyBlock c) → Update m (ForgeState (ShelleyBlock c)) → BlockNo → TickedLedgerState (ShelleyBlock c) → [GenTx (ShelleyBlock c)] → IsLeader (BlockProtocol (ShelleyBlock c)) → m (ShelleyBlock c)