Ouroboros.Consensus.Shelley.Ledger.Ledger

Documentation

newtype ShelleyLedgerError era

Constructors

BBodyError (BlockTransitionError era)

class (PraosCrypto (Crypto era), ShelleyBased era, GetLedgerView era, ApplyBlock era, ApplyTx era) ⇒ ShelleyBasedEra era

data ShelleyTip era

Constructors

ShelleyTip

Fields

shelleyTipSlotNo ∷ !SlotNo

shelleyTipBlockNo ∷ !BlockNo

shelleyTipHash ∷ !(HeaderHash (ShelleyBlock era))

shelleyTipToPoint ∷ WithOrigin (ShelleyTip era) → Point (ShelleyBlock era)

shelleyLedgerTipPoint ∷ LedgerState (ShelleyBlock era) → Point (ShelleyBlock era)

newtype ShelleyTransition

Information required to determine the hard fork point from Shelley to the next ledger

Constructors

ShelleyTransitionInfo

Fields

shelleyAfterVoting ∷ Word32

The number of blocks in this epoch past the voting deadline

We record this to make sure that we can tell the HFC about hard forks if and only if we are certain:

Blocks that came in within an epoch after the 4k/f voting deadline are not relevant (10kf - 2 * 3kf).

Since there are slots between blocks, we are probably only sure that there will be no more relevant block when we have seen the first block after the deadline.

If we count how many blocks we have seen post deadline, and we have reached k of them, we know that that last pre-deadline block won't be rolled back anymore.

At this point we can look at the ledger state and see which proposals we accepted in the voting period, if any, and notify the HFC is one of them indicates a transition.

data family LedgerState blk

Ledger state associated with a block

data family Ticked st

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

Ledger config

data ShelleyLedgerConfig era

Constructors

ShelleyLedgerConfig

Fields

shelleyLedgerCompactGenesis ∷ !(CompactGenesis era)

shelleyLedgerGlobals ∷ !Globals

Derived from shelleyLedgerGenesis but we store a cached version because it used very often.

shelleyLedgerGenesis ∷ ShelleyLedgerConfig era → ShelleyGenesis era

mkShelleyLedgerConfig ∷ ShelleyGenesis era → EpochInfo Identity → MaxMajorProtVer → ShelleyLedgerConfig era

shelleyEraParams ∷ SafeBeforeEpoch → ShelleyGenesis era → EraParams

shelleyEraParamsNeverHardForks ∷ ShelleyGenesis era → EraParams

Separate variant of shelleyEraParams to be used for a Shelley-only chain.

Auxiliary

getPParams ∷ NewEpochState era → PParams era

Serialisation

encodeShelleyAnnTip ∷ ShelleyBasedEra era ⇒ AnnTip (ShelleyBlock era) → Encoding

decodeShelleyAnnTip ∷ ShelleyBasedEra era ⇒ Decoder s (AnnTip (ShelleyBlock era))

decodeShelleyLedgerState ∷ ∀ era s. ShelleyBasedEra era ⇒ Decoder s (LedgerState (ShelleyBlock era))

encodeShelleyLedgerState ∷ ShelleyBasedEra era ⇒ LedgerState (ShelleyBlock era) → Encoding

encodeShelleyHeaderState ∷ ShelleyBasedEra era ⇒ HeaderState (ShelleyBlock era) → Encoding

Orphan instances

ShelleyBasedEra era ⇒ Eq (LedgerState (ShelleyBlock era))

Instance details

Methods

(==) ∷ LedgerState (ShelleyBlock era) → LedgerState (ShelleyBlock era) → Bool

(/=) ∷ LedgerState (ShelleyBlock era) → LedgerState (ShelleyBlock era) → Bool

ShelleyBasedEra era ⇒ Show (LedgerState (ShelleyBlock era))

Instance details

Methods

showsPrec ∷ Int → LedgerState (ShelleyBlock era) → ShowS

show ∷ LedgerState (ShelleyBlock era) → String

showList ∷ [LedgerState (ShelleyBlock era)] → ShowS

Generic (LedgerState (ShelleyBlock era))

Instance details

Associated Types

type Rep (LedgerState (ShelleyBlock era)) ∷ Type → Type

Methods

from ∷ LedgerState (ShelleyBlock era) → Rep (LedgerState (ShelleyBlock era)) x

to ∷ Rep (LedgerState (ShelleyBlock era)) x → LedgerState (ShelleyBlock era)

Generic (Ticked (LedgerState (ShelleyBlock era)))

Instance details

Associated Types

type Rep (Ticked (LedgerState (ShelleyBlock era))) ∷ Type → Type

Methods

from ∷ Ticked (LedgerState (ShelleyBlock era)) → Rep (Ticked (LedgerState (ShelleyBlock era))) x

to ∷ Rep (Ticked (LedgerState (ShelleyBlock era))) x → Ticked (LedgerState (ShelleyBlock era))

HasHardForkHistory (ShelleyBlock era)

Instance details

Associated Types

type HardForkIndices (ShelleyBlock era) ∷ [Type]

Methods

hardForkSummary ∷ LedgerConfig (ShelleyBlock era) → LedgerState (ShelleyBlock era) → Summary (HardForkIndices (ShelleyBlock era))

ShelleyBasedEra era ⇒ LedgerSupportsProtocol (ShelleyBlock era)

Instance details

Methods

protocolLedgerView ∷ LedgerConfig (ShelleyBlock era) → Ticked (LedgerState (ShelleyBlock era)) → Ticked (LedgerView (BlockProtocol (ShelleyBlock era)))

ledgerViewForecastAt ∷ LedgerConfig (ShelleyBlock era) → LedgerState (ShelleyBlock era) → Forecast (LedgerView (BlockProtocol (ShelleyBlock era)))

ShelleyBasedEra era ⇒ BasicEnvelopeValidation (ShelleyBlock era)

Instance details

Methods

expectedFirstBlockNo ∷ proxy (ShelleyBlock era) → BlockNo

expectedNextBlockNo ∷ proxy (ShelleyBlock era) → TipInfo (ShelleyBlock era) → TipInfo (ShelleyBlock era) → BlockNo → BlockNo

minimumPossibleSlotNo ∷ Proxy (ShelleyBlock era) → SlotNo

minimumNextSlotNo ∷ proxy (ShelleyBlock era) → TipInfo (ShelleyBlock era) → TipInfo (ShelleyBlock era) → SlotNo → SlotNo

ShelleyBasedEra era ⇒ ValidateEnvelope (ShelleyBlock era)

Instance details

Associated Types

type OtherHeaderEnvelopeError (ShelleyBlock era)

Methods

additionalEnvelopeChecks ∷ TopLevelConfig (ShelleyBlock era) → Ticked (LedgerView (BlockProtocol (ShelleyBlock era))) → Header (ShelleyBlock era) → Except (OtherHeaderEnvelopeError (ShelleyBlock era)) ()

ShelleyBasedEra era ⇒ CommonProtocolParams (ShelleyBlock era)

Instance details

Methods

maxHeaderSize ∷ LedgerState (ShelleyBlock era) → Word32

maxTxSize ∷ LedgerState (ShelleyBlock era) → Word32

ShelleyBasedEra era ⇒ UpdateLedger (ShelleyBlock era)

Instance details

GetTip (LedgerState (ShelleyBlock era))

Instance details

Methods

getTip ∷ LedgerState (ShelleyBlock era) → Point (LedgerState (ShelleyBlock era))

GetTip (Ticked (LedgerState (ShelleyBlock era)))

Instance details

Methods

getTip ∷ Ticked (LedgerState (ShelleyBlock era)) → Point (Ticked (LedgerState (ShelleyBlock era)))

ShelleyBasedEra era ⇒ IsLedger (LedgerState (ShelleyBlock era))

Instance details

Associated Types

type LedgerErr (LedgerState (ShelleyBlock era))

Methods

applyChainTick ∷ LedgerCfg (LedgerState (ShelleyBlock era)) → SlotNo → LedgerState (ShelleyBlock era) → Ticked (LedgerState (ShelleyBlock era))

NoThunks (LedgerState (ShelleyBlock era))

Instance details

Methods

noThunks ∷ Context → LedgerState (ShelleyBlock era) → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → LedgerState (ShelleyBlock era) → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (LedgerState (ShelleyBlock era)) → String

NoThunks (Ticked (LedgerState (ShelleyBlock era)))

Instance details

Methods

noThunks ∷ Context → Ticked (LedgerState (ShelleyBlock era)) → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → Ticked (LedgerState (ShelleyBlock era)) → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (Ticked (LedgerState (ShelleyBlock era))) → String

ShelleyBasedEra era ⇒ ApplyBlock (LedgerState (ShelleyBlock era)) (ShelleyBlock era)

Instance details

Methods

applyLedgerBlock ∷ LedgerCfg (LedgerState (ShelleyBlock era)) → ShelleyBlock era → Ticked (LedgerState (ShelleyBlock era)) → Except (LedgerErr (LedgerState (ShelleyBlock era))) (LedgerState (ShelleyBlock era))

reapplyLedgerBlock ∷ LedgerCfg (LedgerState (ShelleyBlock era)) → ShelleyBlock era → Ticked (LedgerState (ShelleyBlock era)) → LedgerState (ShelleyBlock era)