Ouroboros.Consensus.Shelley.Ledger.Ledger

Documentation

data ShelleyLedgerError era

Constructors

TickError !(TickTransitionError era)

BBodyError !(BlockTransitionError era)

data ShelleyTip era

Constructors

ShelleyTip

Fields

shelleyTipSlotNo ∷ !SlotNo

shelleyTipBlockNo ∷ !BlockNo

shelleyTipHash ∷ !(HeaderHash (ShelleyBlock era))

shelleyTipToPoint ∷ WithOrigin (ShelleyTip era) → Point (ShelleyBlock era)

data ShelleyTransition

Hardfork point from Shelley to the next ledger

Placeholder type until we properly address #2471

Constructors

ShelleyTransitionUnknown
Transition point not yet known

data family LedgerState blk ∷ Type

Ledger state associated with a block

data family Ticked st ∷ Type

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

class (ShowQuery (Query blk), SameDepIndex (Query blk)) ⇒ QueryLedger blk where

Query the ledger state.

Used by the LocalStateQuery protocol to allow clients to query the ledger state.

Methods

answerQuery ∷ LedgerConfig blk → Query blk result → LedgerState blk → result

Answer the given query about the ledger state.

data family Query blk a ∷ Type

Different queries supported by the ledger, indexed by the result type.

newtype NonMyopicMemberRewards era

Constructors

NonMyopicMemberRewards

Fields

unNonMyopicMemberRewards ∷ Map (Either Coin (Credential Staking era)) (Map (KeyHash StakePool era) Coin)

Ledger config

data ShelleyLedgerConfig era

Constructors

ShelleyLedgerConfig

Fields

shelleyLedgerGenesis ∷ !(ShelleyGenesis era)

shelleyLedgerGlobals ∷ !Globals

Derived from shelleyLedgerGenesis but we store a cached version because it used very often.

shelleyLedgerEraParams ∷ !EraParams

Derived from shelleyLedgerGenesis but we store a cached version because it used very often.

mkShelleyLedgerConfig ∷ ShelleyGenesis era → EpochInfo Identity → MaxMajorProtVer → ShelleyLedgerConfig era

shelleyEraParams ∷ ShelleyGenesis era → EraParams

Auxiliary

getPParams ∷ ShelleyState era → PParams era

Serialisation

encodeShelleyAnnTip ∷ Era era ⇒ AnnTip (ShelleyBlock era) → Encoding

decodeShelleyAnnTip ∷ Era era ⇒ Decoder s (AnnTip (ShelleyBlock era))

decodeShelleyLedgerState ∷ ∀ era s. Era era ⇒ Decoder s (LedgerState (ShelleyBlock era))

encodeShelleyLedgerState ∷ Era era ⇒ LedgerState (ShelleyBlock era) → Encoding

encodeShelleyQuery ∷ Era era ⇒ Query (ShelleyBlock era) result → Encoding

decodeShelleyQuery ∷ Era era ⇒ Decoder s (SomeBlock Query (ShelleyBlock era))

encodeShelleyResult ∷ Era era ⇒ Query (ShelleyBlock era) result → result → Encoding

decodeShelleyResult ∷ Era era ⇒ Query (ShelleyBlock era) result → ∀ s. Decoder s result

encodeShelleyHeaderState ∷ Era era ⇒ HeaderState (ShelleyBlock era) → Encoding

Orphan instances

Era era ⇒ Eq (LedgerState (ShelleyBlock era))

Instance details

Methods

(==) ∷ LedgerState (ShelleyBlock era) → LedgerState (ShelleyBlock era) → Bool

(/=) ∷ LedgerState (ShelleyBlock era) → LedgerState (ShelleyBlock era) → Bool

Era era ⇒ Show (LedgerState (ShelleyBlock era))

Instance details

Methods

showsPrec ∷ Int → LedgerState (ShelleyBlock era) → ShowS

show ∷ LedgerState (ShelleyBlock era) → String

showList ∷ [LedgerState (ShelleyBlock era)] → ShowS

Generic (LedgerState (ShelleyBlock era))

Instance details

Associated Types

type Rep (LedgerState (ShelleyBlock era)) ∷ Type → Type

Methods

from ∷ LedgerState (ShelleyBlock era) → Rep (LedgerState (ShelleyBlock era)) x

to ∷ Rep (LedgerState (ShelleyBlock era)) x → LedgerState (ShelleyBlock era)

Generic (Ticked (LedgerState (ShelleyBlock era)))

Instance details

Associated Types

type Rep (Ticked (LedgerState (ShelleyBlock era))) ∷ Type → Type

Methods

from ∷ Ticked (LedgerState (ShelleyBlock era)) → Rep (Ticked (LedgerState (ShelleyBlock era))) x

to ∷ Rep (Ticked (LedgerState (ShelleyBlock era))) x → Ticked (LedgerState (ShelleyBlock era))

NoUnexpectedThunks (LedgerState (ShelleyBlock era))

Instance details

Methods

noUnexpectedThunks ∷ [String] → LedgerState (ShelleyBlock era) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → LedgerState (ShelleyBlock era) → IO ThunkInfo

showTypeOf ∷ Proxy (LedgerState (ShelleyBlock era)) → String

NoUnexpectedThunks (Ticked (LedgerState (ShelleyBlock era)))

Instance details

Methods

noUnexpectedThunks ∷ [String] → Ticked (LedgerState (ShelleyBlock era)) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → Ticked (LedgerState (ShelleyBlock era)) → IO ThunkInfo

showTypeOf ∷ Proxy (Ticked (LedgerState (ShelleyBlock era))) → String

HasHardForkHistory (ShelleyBlock era)

Instance details

Associated Types

type HardForkIndices (ShelleyBlock era) ∷ [Type]

Methods

hardForkSummary ∷ LedgerConfig (ShelleyBlock era) → LedgerState (ShelleyBlock era) → Summary (HardForkIndices (ShelleyBlock era))

TPraosCrypto era ⇒ LedgerSupportsProtocol (ShelleyBlock era)

Instance details

Methods

protocolLedgerView ∷ LedgerConfig (ShelleyBlock era) → Ticked (LedgerState (ShelleyBlock era)) → Ticked (LedgerView (BlockProtocol (ShelleyBlock era)))

ledgerViewForecastAt ∷ LedgerConfig (ShelleyBlock era) → LedgerState (ShelleyBlock era) → Forecast (LedgerView (BlockProtocol (ShelleyBlock era)))

Era era ⇒ BasicEnvelopeValidation (ShelleyBlock era)

Instance details

Methods

expectedFirstBlockNo ∷ proxy (ShelleyBlock era) → BlockNo

expectedNextBlockNo ∷ proxy (ShelleyBlock era) → TipInfo (ShelleyBlock era) → TipInfo (ShelleyBlock era) → BlockNo → BlockNo

minimumPossibleSlotNo ∷ Proxy (ShelleyBlock era) → SlotNo

minimumNextSlotNo ∷ proxy (ShelleyBlock era) → TipInfo (ShelleyBlock era) → TipInfo (ShelleyBlock era) → SlotNo → SlotNo

Era era ⇒ ValidateEnvelope (ShelleyBlock era)

Instance details

Associated Types

type OtherHeaderEnvelopeError (ShelleyBlock era) ∷ Type

Methods

additionalEnvelopeChecks ∷ TopLevelConfig (ShelleyBlock era) → Ticked (LedgerView (BlockProtocol (ShelleyBlock era))) → Header (ShelleyBlock era) → Except (OtherHeaderEnvelopeError (ShelleyBlock era)) ()

TPraosCrypto era ⇒ CommonProtocolParams (ShelleyBlock era)

Instance details

Methods

maxHeaderSize ∷ LedgerState (ShelleyBlock era) → Word32

maxTxSize ∷ LedgerState (ShelleyBlock era) → Word32

TPraosCrypto era ⇒ UpdateLedger (ShelleyBlock era)

Instance details

TPraosCrypto era ⇒ QueryLedger (ShelleyBlock era)

Instance details

Methods

answerQuery ∷ LedgerConfig (ShelleyBlock era) → Query (ShelleyBlock era) result → LedgerState (ShelleyBlock era) → result

GetTip (LedgerState (ShelleyBlock era))

Instance details

Methods

getTip ∷ LedgerState (ShelleyBlock era) → Point (LedgerState (ShelleyBlock era))

GetTip (Ticked (LedgerState (ShelleyBlock era)))

Instance details

Methods

getTip ∷ Ticked (LedgerState (ShelleyBlock era)) → Point (Ticked (LedgerState (ShelleyBlock era)))

Era era ⇒ IsLedger (LedgerState (ShelleyBlock era))

Instance details

Associated Types

type LedgerErr (LedgerState (ShelleyBlock era)) ∷ Type

Methods

applyChainTick ∷ LedgerCfg (LedgerState (ShelleyBlock era)) → SlotNo → LedgerState (ShelleyBlock era) → Ticked (LedgerState (ShelleyBlock era))

SameDepIndex (Query (ShelleyBlock era))

Instance details

Methods

sameDepIndex ∷ Query (ShelleyBlock era) a → Query (ShelleyBlock era) b → Maybe (a :~: b)

Era era ⇒ ShowQuery (Query (ShelleyBlock era))

Instance details

Methods

showResult ∷ Query (ShelleyBlock era) result → result → String

TPraosCrypto era ⇒ ApplyBlock (LedgerState (ShelleyBlock era)) (ShelleyBlock era)

Instance details

Methods

applyLedgerBlock ∷ LedgerCfg (LedgerState (ShelleyBlock era)) → ShelleyBlock era → Ticked (LedgerState (ShelleyBlock era)) → Except (LedgerErr (LedgerState (ShelleyBlock era))) (LedgerState (ShelleyBlock era))

reapplyLedgerBlock ∷ LedgerCfg (LedgerState (ShelleyBlock era)) → ShelleyBlock era → Ticked (LedgerState (ShelleyBlock era)) → LedgerState (ShelleyBlock era)

Eq (Query (ShelleyBlock era) result)

Instance details

Methods

(==) ∷ Query (ShelleyBlock era) result → Query (ShelleyBlock era) result → Bool

(/=) ∷ Query (ShelleyBlock era) result → Query (ShelleyBlock era) result → Bool

Show (Query (ShelleyBlock era) result)

Instance details

Methods

showsPrec ∷ Int → Query (ShelleyBlock era) result → ShowS

show ∷ Query (ShelleyBlock era) result → String

showList ∷ [Query (ShelleyBlock era) result] → ShowS

Typeable era ⇒ ShowProxy (Query (ShelleyBlock era) ∷ Type → Type)

Instance details

Methods

showProxy ∷ Proxy (Query (ShelleyBlock era)) → String