Ouroboros.Consensus.Shelley.Ledger.Ledger

Documentation

data family LedgerState blk

Ledger state associated with a block

class (ShelleyBasedEra era, PraosCrypto (EraCrypto era), GetLedgerView era, State (EraRule "PPUP" era) ~ PPUPState era, Default (State (EraRule "PPUP" era)), HasField "_maxBHSize" (PParams era) Natural, HasField "_maxTxSize" (PParams era) Natural, HasField "_a0" (PParams era) NonNegativeInterval, HasField "_nOpt" (PParams era) Natural, HasField "_rho" (PParams era) UnitInterval, HasField "_tau" (PParams era) UnitInterval, ValidateScript era, FromCBOR (PParams era), ToCBOR (PParams era), HasField "_protocolVersion" (PParamsDelta era) (StrictMaybe ProtVer), FromCBOR (PParamsDelta era), AdditionalGenesisConfig era ~ TranslationContext era, ToCBORGroup (TxSeq era), NoThunks (TranslationContext era), FromCBOR (Annotator (Witnesses era)), ToCBOR (Witnesses era), Eq (TxSeq era), Show (TxSeq era), FromCBOR (Annotator (TxSeq era))) ⇒ ShelleyBasedEra era

The ledger already defines ShelleyBasedEra as the top-level constraint on an era, however, consensus often needs some more functionality than the ledger currently provides.

Either the functionality shouldn't or can't live in the ledger, in which case it can be part and remain part of ShelleyBasedEra. Or, the functionality should live in the ledger, but hasn't yet been added to the ledger, or it hasn't yet been propagated to this repository, in which case it can be added to this class until that is the case.

By having the same name as the class defined in ledger, we can, if this class becomes redundant, switch to the ledger-defined one without having to update all the code using it. We can just export the right one from this module.

TODO Currently we include some constraints on the update state which are needed to determine the hard fork point. In the future this should be replaced with an appropriate API - see https://github.com/input-output-hk/ouroboros-network/issues/2890

Minimal complete definition

shelleyBasedEraName, applyShelleyBasedTx

newtype ShelleyLedgerError era

Constructors

BBodyError (BlockTransitionError era)

data ShelleyTip era

Constructors

ShelleyTip

Fields

shelleyTipSlotNo ∷ !SlotNo

shelleyTipBlockNo ∷ !BlockNo

shelleyTipHash ∷ !(HeaderHash (ShelleyBlock era))

newtype ShelleyTransition

Information required to determine the hard fork point from Shelley to the next ledger

Constructors

ShelleyTransitionInfo

Fields

shelleyAfterVoting ∷ Word32

The number of blocks in this epoch past the voting deadline

We record this to make sure that we can tell the HFC about hard forks if and only if we are certain:

Blocks that came in within an epoch after the 4k/f voting deadline are not relevant (10kf - 2 * 3kf).

Since there are slots between blocks, we are probably only sure that there will be no more relevant block when we have seen the first block after the deadline.

If we count how many blocks we have seen post deadline, and we have reached k of them, we know that that last pre-deadline block won't be rolled back anymore.

At this point we can look at the ledger state and see which proposals we accepted in the voting period, if any, and notify the HFC is one of them indicates a transition.

data family Ticked st

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

shelleyLedgerTipPoint ∷ LedgerState (ShelleyBlock era) → Point (ShelleyBlock era)

shelleyTipToPoint ∷ WithOrigin (ShelleyTip era) → Point (ShelleyBlock era)

Ledger config

data ShelleyLedgerConfig era

Constructors

ShelleyLedgerConfig

Fields

shelleyLedgerCompactGenesis ∷ !(CompactGenesis era)

shelleyLedgerGlobals ∷ !Globals

Derived from shelleyLedgerGenesis but we store a cached version because it used very often.

shelleyLedgerTranslationContext ∷ !(TranslationContext era)

mkShelleyLedgerConfig ∷ ShelleyGenesis era → TranslationContext era → EpochInfo (Except PastHorizonException) → MaxMajorProtVer → ShelleyLedgerConfig era

shelleyEraParams ∷ ShelleyGenesis era → EraParams

shelleyEraParamsNeverHardForks ∷ ShelleyGenesis era → EraParams

Separate variant of shelleyEraParams to be used for a Shelley-only chain.

shelleyLedgerGenesis ∷ ShelleyLedgerConfig era → ShelleyGenesis era

Auxiliary

data ShelleyLedgerEvent era

All events emitted by the Shelley ledger API

Constructors

ShelleyLedgerEventBBODY (Event (EraRule "BBODY" era))
An event emitted when (re)applying a block

ShelleyLedgerEventTICK (Event (EraRule "TICK" era))
An event emitted during the chain tick

data ShelleyReapplyException

Constructors

∀ era.Show (BlockTransitionError era) ⇒ ShelleyReapplyException (BlockTransitionError era)

getPParams ∷ NewEpochState era → PParams era

Serialisation

decodeShelleyAnnTip ∷ ShelleyBasedEra era ⇒ Decoder s (AnnTip (ShelleyBlock era))

decodeShelleyLedgerState ∷ ∀ era s. ShelleyBasedEra era ⇒ Decoder s (LedgerState (ShelleyBlock era))

encodeShelleyAnnTip ∷ ShelleyBasedEra era ⇒ AnnTip (ShelleyBlock era) → Encoding

encodeShelleyHeaderState ∷ ShelleyBasedEra era ⇒ HeaderState (ShelleyBlock era) → Encoding

encodeShelleyLedgerState ∷ ShelleyBasedEra era ⇒ LedgerState (ShelleyBlock era) → Encoding

Orphan instances

ShelleyBasedEra era ⇒ Eq (LedgerState (ShelleyBlock era))

Instance details

Methods

(==) ∷ LedgerState (ShelleyBlock era) → LedgerState (ShelleyBlock era) → Bool

(/=) ∷ LedgerState (ShelleyBlock era) → LedgerState (ShelleyBlock era) → Bool

ShelleyBasedEra era ⇒ Show (LedgerState (ShelleyBlock era))

Instance details

Methods

showsPrec ∷ Int → LedgerState (ShelleyBlock era) → ShowS

show ∷ LedgerState (ShelleyBlock era) → String

showList ∷ [LedgerState (ShelleyBlock era)] → ShowS

Generic (LedgerState (ShelleyBlock era))

Instance details

Associated Types

type Rep (LedgerState (ShelleyBlock era)) ∷ Type → Type

Methods

from ∷ LedgerState (ShelleyBlock era) → Rep (LedgerState (ShelleyBlock era)) x

to ∷ Rep (LedgerState (ShelleyBlock era)) x → LedgerState (ShelleyBlock era)

Generic (Ticked (LedgerState (ShelleyBlock era)))

Instance details

Associated Types

type Rep (Ticked (LedgerState (ShelleyBlock era))) ∷ Type → Type

Methods

from ∷ Ticked (LedgerState (ShelleyBlock era)) → Rep (Ticked (LedgerState (ShelleyBlock era))) x

to ∷ Rep (Ticked (LedgerState (ShelleyBlock era))) x → Ticked (LedgerState (ShelleyBlock era))

HasHardForkHistory (ShelleyBlock era)

Instance details

Associated Types

type HardForkIndices (ShelleyBlock era) ∷ [Type]

Methods

hardForkSummary ∷ LedgerConfig (ShelleyBlock era) → LedgerState (ShelleyBlock era) → Summary (HardForkIndices (ShelleyBlock era))

ShelleyBasedEra era ⇒ LedgerSupportsProtocol (ShelleyBlock era)

Instance details

Methods

protocolLedgerView ∷ LedgerConfig (ShelleyBlock era) → Ticked (LedgerState (ShelleyBlock era)) → Ticked (LedgerView (BlockProtocol (ShelleyBlock era)))

ledgerViewForecastAt ∷ LedgerConfig (ShelleyBlock era) → LedgerState (ShelleyBlock era) → Forecast (LedgerView (BlockProtocol (ShelleyBlock era)))

ShelleyBasedEra era ⇒ BasicEnvelopeValidation (ShelleyBlock era)

Instance details

Methods

expectedFirstBlockNo ∷ proxy (ShelleyBlock era) → BlockNo

expectedNextBlockNo ∷ proxy (ShelleyBlock era) → TipInfo (ShelleyBlock era) → TipInfo (ShelleyBlock era) → BlockNo → BlockNo

minimumPossibleSlotNo ∷ Proxy (ShelleyBlock era) → SlotNo

minimumNextSlotNo ∷ proxy (ShelleyBlock era) → TipInfo (ShelleyBlock era) → TipInfo (ShelleyBlock era) → SlotNo → SlotNo

ShelleyBasedEra era ⇒ ValidateEnvelope (ShelleyBlock era)

Instance details

Associated Types

type OtherHeaderEnvelopeError (ShelleyBlock era)

Methods

additionalEnvelopeChecks ∷ TopLevelConfig (ShelleyBlock era) → Ticked (LedgerView (BlockProtocol (ShelleyBlock era))) → Header (ShelleyBlock era) → Except (OtherHeaderEnvelopeError (ShelleyBlock era)) ()

ShelleyBasedEra era ⇒ CommonProtocolParams (ShelleyBlock era)

Instance details

Methods

maxHeaderSize ∷ LedgerState (ShelleyBlock era) → Word32

maxTxSize ∷ LedgerState (ShelleyBlock era) → Word32

ShelleyBasedEra era ⇒ UpdateLedger (ShelleyBlock era)

Instance details

GetTip (LedgerState (ShelleyBlock era))

Instance details

Methods

getTip ∷ LedgerState (ShelleyBlock era) → Point (LedgerState (ShelleyBlock era))

GetTip (Ticked (LedgerState (ShelleyBlock era)))

Instance details

Methods

getTip ∷ Ticked (LedgerState (ShelleyBlock era)) → Point (Ticked (LedgerState (ShelleyBlock era)))

ShelleyBasedEra era ⇒ IsLedger (LedgerState (ShelleyBlock era))

Instance details

Associated Types

type LedgerErr (LedgerState (ShelleyBlock era))

type AuxLedgerEvent (LedgerState (ShelleyBlock era))

Methods

applyChainTickLedgerResult ∷ LedgerCfg (LedgerState (ShelleyBlock era)) → SlotNo → LedgerState (ShelleyBlock era) → LedgerResult (LedgerState (ShelleyBlock era)) (Ticked (LedgerState (ShelleyBlock era)))

ShelleyBasedEra era ⇒ NoThunks (LedgerState (ShelleyBlock era))

Instance details

Methods

noThunks ∷ Context → LedgerState (ShelleyBlock era) → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → LedgerState (ShelleyBlock era) → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (LedgerState (ShelleyBlock era)) → String

ShelleyBasedEra era ⇒ NoThunks (Ticked (LedgerState (ShelleyBlock era)))

Instance details

Methods

noThunks ∷ Context → Ticked (LedgerState (ShelleyBlock era)) → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → Ticked (LedgerState (ShelleyBlock era)) → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (Ticked (LedgerState (ShelleyBlock era))) → String

ShelleyBasedEra era ⇒ ApplyBlock (LedgerState (ShelleyBlock era)) (ShelleyBlock era)

Instance details

Methods

applyBlockLedgerResult ∷ LedgerCfg (LedgerState (ShelleyBlock era)) → ShelleyBlock era → Ticked (LedgerState (ShelleyBlock era)) → Except (LedgerErr (LedgerState (ShelleyBlock era))) (LedgerResult (LedgerState (ShelleyBlock era)) (LedgerState (ShelleyBlock era)))

reapplyBlockLedgerResult ∷ LedgerCfg (LedgerState (ShelleyBlock era)) → ShelleyBlock era → Ticked (LedgerState (ShelleyBlock era)) → LedgerResult (LedgerState (ShelleyBlock era)) (LedgerState (ShelleyBlock era))