Ouroboros.Consensus.Shelley.Ledger.Ledger

Documentation

data ShelleyLedgerError c

Constructors

TickError !(TickTransitionError c)

BBodyError !(BlockTransitionError c)

ObsoleteNode

Fields

Natural

Current major protocol version

Natural

Maximum major protocol version

HeaderSizeTooLarge

Fields

Natural

Size of the header

Natural

Maximum size of a header

BodySizeTooLarge

Fields

Natural

Size of the block body

Natural

Maximum size of a block body

data family LedgerConfig blk ∷ Type

Static environment required for the ledger

data family LedgerState blk ∷ Type

class (HasHeader blk, HasHeader (Header blk), Show (LedgerState blk), Show (LedgerError blk), Eq (LedgerState blk), Eq (LedgerError blk), NoUnexpectedThunks (LedgerState blk), NoUnexpectedThunks (LedgerError blk), NoUnexpectedThunks (LedgerConfig blk)) ⇒ UpdateLedger blk where

Interaction with the ledger layer

Associated Types

data LedgerState blk ∷ Type

type LedgerError blk ∷ Type

data LedgerConfig blk ∷ Type

Static environment required for the ledger

Methods

applyChainTick ∷ LedgerConfig blk → SlotNo → LedgerState blk → TickedLedgerState blk

Apply "slot based" state transformations

When a block is applied to the ledger state, a number of things happen purely based on the slot number of that block. For example:

In Byron, scheduled updates are applied, and the update system state is updated.

In Shelley, delegation state is updated (on epoch boundaries).

The consensus layer must be able to apply such a "chain tick" function, primarily when validating transactions in the mempool (which, conceptually, live in "some block in the future") or when extracting valid transactions from the mempool to insert into a new block to be produced.

This is not allowed to throw any errors. After all, if this could fail, it would mean a previous block set up the ledger state in such a way that as soon as a certain slot was reached, any block would be invalid.

PRECONDITION: The slot number must be strictly greater than the slot at the tip of the ledger.

applyLedgerBlock ∷ LedgerConfig blk → blk → LedgerState blk → Except (LedgerError blk) (LedgerState blk)

Apply a block to the ledger state.

This should apply the entire block (i.e., including applyChainTick).

reapplyLedgerBlock ∷ LedgerConfig blk → blk → LedgerState blk → LedgerState blk

Re-apply a block to the very same ledger state it was applied in before.

Since a block can only be applied to a single, specific, ledger state, if we apply a previously applied block again it will be applied in the very same ledger state, and therefore can't possibly fail.

It is worth noting that since we already know that the block is valid in the provided ledger state, the ledger layer should not perform any validation checks.

ledgerTipPoint ∷ LedgerState blk → Point blk

Point of the most recently applied block

Should be genesisPoint when no blocks have been applied yet

class (UpdateLedger blk, ShowQuery (Query blk)) ⇒ QueryLedger blk where

Query the ledger state.

Used by the LocalStateQuery protocol to allow clients to query the ledger state.

Associated Types

data Query blk a ∷ Type

Different queries supported by the ledger, indexed by the result type.

Methods

answerQuery ∷ LedgerConfig blk → Query blk result → LedgerState blk → result

Answer the given query about the ledger state.

eqQuery ∷ Query blk result1 → Query blk result2 → Maybe (result1 :~: result2)

Generalisation of value-level equality of two queries.

data family Query blk a ∷ Type

Different queries supported by the ledger, indexed by the result type.

newtype NonMyopicMemberRewards c

Constructors

NonMyopicMemberRewards

Fields

unNonMyopicMemberRewards ∷ Map (Credential c) (Map (KeyHash c) Coin)

Auxiliary

getPParams ∷ ShelleyState c → PParams

Serialisation

decodeShelleyLedgerState ∷ Crypto c ⇒ Decoder r (LedgerState (ShelleyBlock c))

encodeShelleyLedgerState ∷ Crypto c ⇒ LedgerState (ShelleyBlock c) → Encoding

encodeShelleyQuery ∷ Crypto c ⇒ Query (ShelleyBlock c) result → Encoding

decodeShelleyQuery ∷ Crypto c ⇒ Decoder s (Some (Query (ShelleyBlock c)))

encodeShelleyResult ∷ Crypto c ⇒ Query (ShelleyBlock c) result → result → Encoding

decodeShelleyResult ∷ Crypto c ⇒ Query (ShelleyBlock c) result → ∀ s. Decoder s result

Orphan instances

Crypto c ⇒ Eq (LedgerState (ShelleyBlock c))

Instance details

Methods

(==) ∷ LedgerState (ShelleyBlock c) → LedgerState (ShelleyBlock c) → Bool

(/=) ∷ LedgerState (ShelleyBlock c) → LedgerState (ShelleyBlock c) → Bool

Crypto c ⇒ Show (LedgerState (ShelleyBlock c))

Instance details

Methods

showsPrec ∷ Int → LedgerState (ShelleyBlock c) → ShowS

show ∷ LedgerState (ShelleyBlock c) → String

showList ∷ [LedgerState (ShelleyBlock c)] → ShowS

Generic (LedgerConfig (ShelleyBlock c))

Instance details

Associated Types

type Rep (LedgerConfig (ShelleyBlock c)) ∷ Type → Type

Methods

from ∷ LedgerConfig (ShelleyBlock c) → Rep (LedgerConfig (ShelleyBlock c)) x

to ∷ Rep (LedgerConfig (ShelleyBlock c)) x → LedgerConfig (ShelleyBlock c)

Generic (LedgerState (ShelleyBlock c))

Instance details

Associated Types

type Rep (LedgerState (ShelleyBlock c)) ∷ Type → Type

Methods

from ∷ LedgerState (ShelleyBlock c) → Rep (LedgerState (ShelleyBlock c)) x

to ∷ Rep (LedgerState (ShelleyBlock c)) x → LedgerState (ShelleyBlock c)

NoUnexpectedThunks (LedgerConfig (ShelleyBlock c))

Instance details

Methods

noUnexpectedThunks ∷ [String] → LedgerConfig (ShelleyBlock c) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → LedgerConfig (ShelleyBlock c) → IO ThunkInfo

showTypeOf ∷ Proxy (LedgerConfig (ShelleyBlock c)) → String

Crypto c ⇒ NoUnexpectedThunks (LedgerState (ShelleyBlock c))

Instance details

Methods

noUnexpectedThunks ∷ [String] → LedgerState (ShelleyBlock c) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → LedgerState (ShelleyBlock c) → IO ThunkInfo

showTypeOf ∷ Proxy (LedgerState (ShelleyBlock c)) → String

HasHardForkHistory (ShelleyBlock c)

Instance details

Associated Types

type HardForkIndices (ShelleyBlock c) ∷ [Type]

Methods

hardForkShape ∷ BlockConfig (ShelleyBlock c) → Shape (HardForkIndices (ShelleyBlock c))

hardForkTransitions ∷ LedgerConfig (ShelleyBlock c) → LedgerState (ShelleyBlock c) → Transitions (HardForkIndices (ShelleyBlock c))

LedgerDerivedInfo (ShelleyBlock c)

Instance details

Methods

knownSlotLengths ∷ BlockConfig (ShelleyBlock c) → SlotLengths

TPraosCrypto c ⇒ LedgerSupportsProtocol (ShelleyBlock c)

Instance details

Methods

protocolLedgerView ∷ LedgerConfig (ShelleyBlock c) → LedgerState (ShelleyBlock c) → LedgerView (BlockProtocol (ShelleyBlock c))

anachronisticProtocolLedgerView_ ∷ LedgerConfig (ShelleyBlock c) → LedgerState (ShelleyBlock c) → WithOrigin SlotNo → Except AnachronyFailure (LedgerView (BlockProtocol (ShelleyBlock c)))

TPraosCrypto c ⇒ UpdateLedger (ShelleyBlock c)

Instance details

Associated Types

data LedgerState (ShelleyBlock c) ∷ Type

type LedgerError (ShelleyBlock c) ∷ Type

data LedgerConfig (ShelleyBlock c) ∷ Type

Methods

applyChainTick ∷ LedgerConfig (ShelleyBlock c) → SlotNo → LedgerState (ShelleyBlock c) → TickedLedgerState (ShelleyBlock c)

applyLedgerBlock ∷ LedgerConfig (ShelleyBlock c) → ShelleyBlock c → LedgerState (ShelleyBlock c) → Except (LedgerError (ShelleyBlock c)) (LedgerState (ShelleyBlock c))

reapplyLedgerBlock ∷ LedgerConfig (ShelleyBlock c) → ShelleyBlock c → LedgerState (ShelleyBlock c) → LedgerState (ShelleyBlock c)

ledgerTipPoint ∷ LedgerState (ShelleyBlock c) → Point (ShelleyBlock c)

TPraosCrypto c ⇒ QueryLedger (ShelleyBlock c)

Instance details

Associated Types

data Query (ShelleyBlock c) a ∷ Type

Methods

answerQuery ∷ LedgerConfig (ShelleyBlock c) → Query (ShelleyBlock c) result → LedgerState (ShelleyBlock c) → result

eqQuery ∷ Query (ShelleyBlock c) result1 → Query (ShelleyBlock c) result2 → Maybe (result1 :~: result2)

Crypto c ⇒ ShowQuery (Query (ShelleyBlock c))

Instance details

Methods

showResult ∷ Query (ShelleyBlock c) result → result → String

Eq (Query (ShelleyBlock c) result)

Instance details

Methods

(==) ∷ Query (ShelleyBlock c) result → Query (ShelleyBlock c) result → Bool

(/=) ∷ Query (ShelleyBlock c) result → Query (ShelleyBlock c) result → Bool

Show (Query (ShelleyBlock c) result)

Instance details

Methods

showsPrec ∷ Int → Query (ShelleyBlock c) result → ShowS

show ∷ Query (ShelleyBlock c) result → String

showList ∷ [Query (ShelleyBlock c) result] → ShowS