Ouroboros.Consensus.Shelley.Ledger.Ledger

Documentation

data ShelleyLedgerError c

Constructors

TickError !(TickTransitionError c)

BBodyError !(BlockTransitionError c)

data family LedgerState blk ∷ Type

class ApplyBlock (LedgerState blk) blk ⇒ UpdateLedger blk

Interaction with the ledger layer

Associated Types

data LedgerState blk ∷ Type

class (UpdateLedger blk, ShowQuery (Query blk)) ⇒ QueryLedger blk where

Query the ledger state.

Used by the LocalStateQuery protocol to allow clients to query the ledger state.

Associated Types

data Query blk a ∷ Type

Different queries supported by the ledger, indexed by the result type.

Methods

answerQuery ∷ LedgerConfig blk → Query blk result → LedgerState blk → result

Answer the given query about the ledger state.

eqQuery ∷ Query blk result1 → Query blk result2 → Maybe (result1 :~: result2)

Generalisation of value-level equality of two queries.

data family Query blk a ∷ Type

Different queries supported by the ledger, indexed by the result type.

newtype NonMyopicMemberRewards c

Constructors

NonMyopicMemberRewards

Fields

unNonMyopicMemberRewards ∷ Map (Credential c) (Map (KeyHash c) Coin)

Auxiliary

getPParams ∷ ShelleyState c → PParams

Serialisation

decodeShelleyLedgerState ∷ Crypto c ⇒ Decoder r (LedgerState (ShelleyBlock c))

encodeShelleyLedgerState ∷ Crypto c ⇒ LedgerState (ShelleyBlock c) → Encoding

encodeShelleyQuery ∷ Crypto c ⇒ Query (ShelleyBlock c) result → Encoding

decodeShelleyQuery ∷ Crypto c ⇒ Decoder s (Some (Query (ShelleyBlock c)))

encodeShelleyResult ∷ Crypto c ⇒ Query (ShelleyBlock c) result → result → Encoding

decodeShelleyResult ∷ Crypto c ⇒ Query (ShelleyBlock c) result → ∀ s. Decoder s result

encodeShelleyExtLedgerState ∷ Crypto c ⇒ ExtLedgerState (ShelleyBlock c) → Encoding

encodeShelleyHeaderState ∷ Crypto c ⇒ HeaderState (ShelleyBlock c) → Encoding

Orphan instances

Crypto c ⇒ Eq (LedgerState (ShelleyBlock c))

Instance details

Methods

(==) ∷ LedgerState (ShelleyBlock c) → LedgerState (ShelleyBlock c) → Bool

(/=) ∷ LedgerState (ShelleyBlock c) → LedgerState (ShelleyBlock c) → Bool

Crypto c ⇒ Show (LedgerState (ShelleyBlock c))

Instance details

Methods

showsPrec ∷ Int → LedgerState (ShelleyBlock c) → ShowS

show ∷ LedgerState (ShelleyBlock c) → String

showList ∷ [LedgerState (ShelleyBlock c)] → ShowS

Generic (LedgerState (ShelleyBlock c))

Instance details

Associated Types

type Rep (LedgerState (ShelleyBlock c)) ∷ Type → Type

Methods

from ∷ LedgerState (ShelleyBlock c) → Rep (LedgerState (ShelleyBlock c)) x

to ∷ Rep (LedgerState (ShelleyBlock c)) x → LedgerState (ShelleyBlock c)

Crypto c ⇒ NoUnexpectedThunks (LedgerState (ShelleyBlock c))

Instance details

Methods

noUnexpectedThunks ∷ [String] → LedgerState (ShelleyBlock c) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → LedgerState (ShelleyBlock c) → IO ThunkInfo

showTypeOf ∷ Proxy (LedgerState (ShelleyBlock c)) → String

HasHardForkHistory (ShelleyBlock c)

Instance details

Associated Types

type HardForkIndices (ShelleyBlock c) ∷ [Type]

Methods

hardForkShape ∷ BlockConfig (ShelleyBlock c) → Shape (HardForkIndices (ShelleyBlock c))

hardForkTransitions ∷ LedgerConfig (ShelleyBlock c) → LedgerState (ShelleyBlock c) → Transitions (HardForkIndices (ShelleyBlock c))

LedgerDerivedInfo (ShelleyBlock c)

Instance details

Methods

knownSlotLength ∷ BlockConfig (ShelleyBlock c) → SlotLength

TPraosCrypto c ⇒ LedgerSupportsProtocol (ShelleyBlock c)

Instance details

Methods

protocolLedgerView ∷ LedgerConfig (ShelleyBlock c) → LedgerState (ShelleyBlock c) → LedgerView (BlockProtocol (ShelleyBlock c))

ledgerViewForecastAt_ ∷ LedgerConfig (ShelleyBlock c) → LedgerState (ShelleyBlock c) → WithOrigin SlotNo → Maybe (Forecast (LedgerView (BlockProtocol (ShelleyBlock c))))

Crypto c ⇒ ValidateEnvelope (ShelleyBlock c)

Instance details

Associated Types

type OtherHeaderEnvelopeError (ShelleyBlock c) ∷ Type

Methods

validateEnvelope ∷ TopLevelConfig (ShelleyBlock c) → LedgerView (BlockProtocol (ShelleyBlock c)) → WithOrigin (AnnTip (ShelleyBlock c)) → Header (ShelleyBlock c) → Except (HeaderEnvelopeError (ShelleyBlock c)) ()

firstBlockNo ∷ Proxy (ShelleyBlock c) → BlockNo

minimumPossibleSlotNo ∷ Proxy (ShelleyBlock c) → SlotNo

TPraosCrypto c ⇒ IsLedger (LedgerState (ShelleyBlock c))

Instance details

Associated Types

type LedgerCfg (LedgerState (ShelleyBlock c)) ∷ Type

type LedgerErr (LedgerState (ShelleyBlock c)) ∷ Type

Methods

applyChainTick ∷ LedgerCfg (LedgerState (ShelleyBlock c)) → SlotNo → LedgerState (ShelleyBlock c) → TickedLedger (LedgerState (ShelleyBlock c))

TPraosCrypto c ⇒ UpdateLedger (ShelleyBlock c)

Instance details

Associated Types

data LedgerState (ShelleyBlock c) ∷ Type

TPraosCrypto c ⇒ QueryLedger (ShelleyBlock c)

Instance details

Associated Types

data Query (ShelleyBlock c) a ∷ Type

Methods

answerQuery ∷ LedgerConfig (ShelleyBlock c) → Query (ShelleyBlock c) result → LedgerState (ShelleyBlock c) → result

eqQuery ∷ Query (ShelleyBlock c) result1 → Query (ShelleyBlock c) result2 → Maybe (result1 :~: result2)

Crypto c ⇒ ShowQuery (Query (ShelleyBlock c))

Instance details

Methods

showResult ∷ Query (ShelleyBlock c) result → result → String

TPraosCrypto c ⇒ ApplyBlock (LedgerState (ShelleyBlock c)) (ShelleyBlock c)

Instance details

Methods

applyLedgerBlock ∷ LedgerCfg (LedgerState (ShelleyBlock c)) → ShelleyBlock c → TickedLedger (LedgerState (ShelleyBlock c)) → Except (LedgerErr (LedgerState (ShelleyBlock c))) (LedgerState (ShelleyBlock c))

reapplyLedgerBlock ∷ LedgerCfg (LedgerState (ShelleyBlock c)) → ShelleyBlock c → TickedLedger (LedgerState (ShelleyBlock c)) → LedgerState (ShelleyBlock c)

ledgerTipPoint ∷ LedgerState (ShelleyBlock c) → Point (ShelleyBlock c)

Eq (Query (ShelleyBlock c) result)

Instance details

Methods

(==) ∷ Query (ShelleyBlock c) result → Query (ShelleyBlock c) result → Bool

(/=) ∷ Query (ShelleyBlock c) result → Query (ShelleyBlock c) result → Bool

Show (Query (ShelleyBlock c) result)

Instance details

Methods

showsPrec ∷ Int → Query (ShelleyBlock c) result → ShowS

show ∷ Query (ShelleyBlock c) result → String

showList ∷ [Query (ShelleyBlock c) result] → ShowS