Ouroboros.Consensus.Shelley.Ledger.Mempool

Description

Shelley mempool integration

Documentation

data ApplyTxError era

Constructors

ApplyTxError [PredicateFailure (LEDGERS era)]

data family GenTx blk ∷ Type

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data family TxId tx ∷ Type

A generalized transaction, GenTx, identifier.

mkShelleyTx ∷ Era era ⇒ Tx era → GenTx (ShelleyBlock era)

fixedBlockBodyOverhead ∷ Num a ⇒ a

txInBlockSize is used to estimate how many transactions we can grab from the Mempool to put into the block we are going to forge without exceeding the maximum block body size according to the ledger. If we exceed that limit, we will have forged a block that is invalid according to the ledger. We ourselves won't even adopt it, causing us to lose our slot, something we must try to avoid.

For this reason it is better to overestimate the size of a transaction than to underestimate. The only downside is that we maybe could have put one (or more?) transactions extra in that block.

As the sum of the serialised transaction sizes is not equal to the size of the serialised block body (TxSeq) consisting of those transactions (see cardano-node#1545 for an example), we account for some extra overhead per transaction as a safety margin.

Also see perTxOverhead.

perTxOverhead ∷ Num a ⇒ a

See fixedBlockBodyOverhead.

Orphan instances

Typeable era ⇒ ShowProxy (GenTx (ShelleyBlock era) ∷ Type)

Instance details

Methods

showProxy ∷ Proxy (GenTx (ShelleyBlock era)) → String

Typeable era ⇒ ShowProxy (TxId (GenTx (ShelleyBlock era)) ∷ Type)

Instance details

Methods

showProxy ∷ Proxy (TxId (GenTx (ShelleyBlock era))) → String

Era era ⇒ Eq (GenTx (ShelleyBlock era))

Instance details

Methods

(==) ∷ GenTx (ShelleyBlock era) → GenTx (ShelleyBlock era) → Bool

(/=) ∷ GenTx (ShelleyBlock era) → GenTx (ShelleyBlock era) → Bool

Eq (TxId (GenTx (ShelleyBlock era)))

Instance details

Methods

(==) ∷ TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era)) → Bool

(/=) ∷ TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era)) → Bool

Ord (TxId (GenTx (ShelleyBlock era)))

Instance details

Methods

compare ∷ TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era)) → Ordering

(<) ∷ TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era)) → Bool

(<=) ∷ TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era)) → Bool

(>) ∷ TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era)) → Bool

(>=) ∷ TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era)) → Bool

max ∷ TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era))

min ∷ TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era))

Era era ⇒ Show (GenTx (ShelleyBlock era))

Instance details

Methods

showsPrec ∷ Int → GenTx (ShelleyBlock era) → ShowS

show ∷ GenTx (ShelleyBlock era) → String

showList ∷ [GenTx (ShelleyBlock era)] → ShowS

Show (GenTxId (ShelleyBlock era))

Instance details

Methods

showsPrec ∷ Int → GenTxId (ShelleyBlock era) → ShowS

show ∷ GenTxId (ShelleyBlock era) → String

showList ∷ [GenTxId (ShelleyBlock era)] → ShowS

Generic (GenTx (ShelleyBlock era))

Instance details

Associated Types

type Rep (GenTx (ShelleyBlock era)) ∷ Type → Type

Methods

from ∷ GenTx (ShelleyBlock era) → Rep (GenTx (ShelleyBlock era)) x

to ∷ Rep (GenTx (ShelleyBlock era)) x → GenTx (ShelleyBlock era)

Era era ⇒ ToCBOR (GenTx (ShelleyBlock era))

Instance details

Methods

toCBOR ∷ GenTx (ShelleyBlock era) → Encoding

encodedSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy (GenTx (ShelleyBlock era)) → Size

encodedListSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy [GenTx (ShelleyBlock era)] → Size

Era era ⇒ ToCBOR (TxId (GenTx (ShelleyBlock era)))

Instance details

Methods

toCBOR ∷ TxId (GenTx (ShelleyBlock era)) → Encoding

encodedSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy (TxId (GenTx (ShelleyBlock era))) → Size

encodedListSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy [TxId (GenTx (ShelleyBlock era))] → Size

Era era ⇒ FromCBOR (GenTx (ShelleyBlock era))

Instance details

Methods

fromCBOR ∷ Decoder s (GenTx (ShelleyBlock era))

label ∷ Proxy (GenTx (ShelleyBlock era)) → Text

Era era ⇒ FromCBOR (TxId (GenTx (ShelleyBlock era)))

Instance details

Methods

fromCBOR ∷ Decoder s (TxId (GenTx (ShelleyBlock era)))

label ∷ Proxy (TxId (GenTx (ShelleyBlock era))) → Text

Era era ⇒ NoUnexpectedThunks (GenTx (ShelleyBlock era))

Instance details

Methods

noUnexpectedThunks ∷ [String] → GenTx (ShelleyBlock era) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → GenTx (ShelleyBlock era) → IO ThunkInfo

showTypeOf ∷ Proxy (GenTx (ShelleyBlock era)) → String

Typeable era ⇒ NoUnexpectedThunks (TxId (GenTx (ShelleyBlock era)))

Instance details

Methods

noUnexpectedThunks ∷ [String] → TxId (GenTx (ShelleyBlock era)) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → TxId (GenTx (ShelleyBlock era)) → IO ThunkInfo

showTypeOf ∷ Proxy (TxId (GenTx (ShelleyBlock era))) → String

TPraosCrypto era ⇒ LedgerSupportsMempool (ShelleyBlock era)

Instance details

Methods

txInvariant ∷ GenTx (ShelleyBlock era) → Bool

applyTx ∷ LedgerConfig (ShelleyBlock era) → SlotNo → GenTx (ShelleyBlock era) → TickedLedgerState (ShelleyBlock era) → Except (ApplyTxErr (ShelleyBlock era)) (TickedLedgerState (ShelleyBlock era))

reapplyTx ∷ LedgerConfig (ShelleyBlock era) → SlotNo → GenTx (ShelleyBlock era) → TickedLedgerState (ShelleyBlock era) → Except (ApplyTxErr (ShelleyBlock era)) (TickedLedgerState (ShelleyBlock era))

maxTxCapacity ∷ TickedLedgerState (ShelleyBlock era) → Word32

txInBlockSize ∷ GenTx (ShelleyBlock era) → Word32

Era era ⇒ HasTxId (GenTx (ShelleyBlock era))

Instance details

Methods

txId ∷ GenTx (ShelleyBlock era) → TxId (GenTx (ShelleyBlock era))

Era era ⇒ HasTxs (ShelleyBlock era)

Instance details

Methods

extractTxs ∷ ShelleyBlock era → [GenTx (ShelleyBlock era)]

Era era ⇒ Condense (GenTx (ShelleyBlock era))

Instance details

Methods

condense ∷ GenTx (ShelleyBlock era) → String

Condense (GenTxId (ShelleyBlock era))

Instance details

Methods

condense ∷ GenTxId (ShelleyBlock era) → String

Typeable era ⇒ ShowProxy (ApplyTxError era ∷ Type)

Instance details

Methods

showProxy ∷ Proxy (ApplyTxError era) → String