Ouroboros.Consensus.Shelley.Ledger.Mempool

Description

Shelley mempool integration

Documentation

data family GenTx blk

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

newtype ApplyTxError era

Constructors

ApplyTxError [PredicateFailure (EraRule "LEDGER" era)]

data family TxId tx

A generalized transaction, GenTx, identifier.

data family Validated x

" Validated " transaction or block

The ledger defines how to validate transactions and blocks. It's possible the type before and after validation may be distinct (eg Alonzo transactions), which originally motivated this family.

We also gain the related benefit that certain interface functions, such as those that reapply blocks, can have a more precise type now. TODO

Similarly, the Node-to-Client mini protocols can explicitly indicate that the client trusts the blocks from the local server, by having the server send Validated blocks to the client. TODO

Note that validation has different implications for a transaction than for a block. In particular, a validated transaction can be " reapplied " to different ledger states, whereas a validated block must only be " reapplied " to the exact same ledger state (eg as part of rebuilding from an on-disk ledger snapshot).

Since the ledger defines validation, see the ledger details for concrete examples of what determines the validity (wrt to a LedgerState) of a transaction and/or block. Example properties include: a transaction's claimed inputs exist and are still unspent, a block carries a sufficient cryptographic signature, etc.

data WithTop a

Add a unique top element to a lattice.

TODO This should be relocated to `cardano-base:Data.Measure'.

Constructors

NotTop a

Top

fixedBlockBodyOverhead ∷ Num a ⇒ a

txInBlockSize is used to estimate how many transactions we can grab from the Mempool to put into the block we are going to forge without exceeding the maximum block body size according to the ledger. If we exceed that limit, we will have forged a block that is invalid according to the ledger. We ourselves won't even adopt it, causing us to lose our slot, something we must try to avoid.

For this reason it is better to overestimate the size of a transaction than to underestimate. The only downside is that we maybe could have put one (or more?) transactions extra in that block.

As the sum of the serialised transaction sizes is not equal to the size of the serialised block body (TxSeq) consisting of those transactions (see cardano-node#1545 for an example), we account for some extra overhead per transaction as a safety margin.

Also see perTxOverhead.

mkShelleyTx ∷ ∀ era. ShelleyBasedEra era ⇒ Tx era → GenTx (ShelleyBlock era)

mkShelleyValidatedTx ∷ ∀ era. ShelleyBasedEra era ⇒ Validated (Tx era) → Validated (GenTx (ShelleyBlock era))

perTxOverhead ∷ Num a ⇒ a

See fixedBlockBodyOverhead.

Exported for tests

data AlonzoMeasure

Constructors

AlonzoMeasure

Fields

byteSize ∷ !ByteSize

exUnits ∷ !(ExUnits' (WithTop Natural))

fromExUnits ∷ ExUnits → ExUnits' (WithTop Natural)

Orphan instances

Typeable era ⇒ ShowProxy (GenTx (ShelleyBlock era) ∷ Type)

Instance details

Methods

showProxy ∷ Proxy (GenTx (ShelleyBlock era)) → String

Typeable era ⇒ ShowProxy (TxId (GenTx (ShelleyBlock era)) ∷ Type)

Instance details

Methods

showProxy ∷ Proxy (TxId (GenTx (ShelleyBlock era))) → String

Typeable era ⇒ ShowProxy (Validated (GenTx (ShelleyBlock era)) ∷ Type)

Instance details

Methods

showProxy ∷ Proxy (Validated (GenTx (ShelleyBlock era))) → String

ShelleyBasedEra era ⇒ Eq (GenTx (ShelleyBlock era))

Instance details

Methods

(==) ∷ GenTx (ShelleyBlock era) → GenTx (ShelleyBlock era) → Bool

(/=) ∷ GenTx (ShelleyBlock era) → GenTx (ShelleyBlock era) → Bool

Eq (TxId (GenTx (ShelleyBlock era)))

Instance details

Methods

(==) ∷ TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era)) → Bool

(/=) ∷ TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era)) → Bool

ShelleyBasedEra era ⇒ Eq (Validated (GenTx (ShelleyBlock era)))

Instance details

Methods

(==) ∷ Validated (GenTx (ShelleyBlock era)) → Validated (GenTx (ShelleyBlock era)) → Bool

(/=) ∷ Validated (GenTx (ShelleyBlock era)) → Validated (GenTx (ShelleyBlock era)) → Bool

Ord (TxId (GenTx (ShelleyBlock era)))

Instance details

Methods

compare ∷ TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era)) → Ordering

(<) ∷ TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era)) → Bool

(<=) ∷ TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era)) → Bool

(>) ∷ TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era)) → Bool

(>=) ∷ TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era)) → Bool

max ∷ TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era))

min ∷ TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era)) → TxId (GenTx (ShelleyBlock era))

ShelleyBasedEra era ⇒ Show (GenTx (ShelleyBlock era))

Instance details

Methods

showsPrec ∷ Int → GenTx (ShelleyBlock era) → ShowS

show ∷ GenTx (ShelleyBlock era) → String

showList ∷ [GenTx (ShelleyBlock era)] → ShowS

Show (GenTxId (ShelleyBlock era))

Instance details

Methods

showsPrec ∷ Int → GenTxId (ShelleyBlock era) → ShowS

show ∷ GenTxId (ShelleyBlock era) → String

showList ∷ [GenTxId (ShelleyBlock era)] → ShowS

ShelleyBasedEra era ⇒ Show (Validated (GenTx (ShelleyBlock era)))

Instance details

Methods

showsPrec ∷ Int → Validated (GenTx (ShelleyBlock era)) → ShowS

show ∷ Validated (GenTx (ShelleyBlock era)) → String

showList ∷ [Validated (GenTx (ShelleyBlock era))] → ShowS

Generic (GenTx (ShelleyBlock era))

Instance details

Associated Types

type Rep (GenTx (ShelleyBlock era)) ∷ Type → Type

Methods

from ∷ GenTx (ShelleyBlock era) → Rep (GenTx (ShelleyBlock era)) x

to ∷ Rep (GenTx (ShelleyBlock era)) x → GenTx (ShelleyBlock era)

Generic (Validated (GenTx (ShelleyBlock era)))

Instance details

Associated Types

type Rep (Validated (GenTx (ShelleyBlock era))) ∷ Type → Type

Methods

from ∷ Validated (GenTx (ShelleyBlock era)) → Rep (Validated (GenTx (ShelleyBlock era))) x

to ∷ Rep (Validated (GenTx (ShelleyBlock era))) x → Validated (GenTx (ShelleyBlock era))

ShelleyBasedEra (AllegraEra c) ⇒ TxLimits (ShelleyBlock (AllegraEra c))

Instance details

Associated Types

type TxMeasure (ShelleyBlock (AllegraEra c))

Methods

txMeasure ∷ Validated (GenTx (ShelleyBlock (AllegraEra c))) → TxMeasure (ShelleyBlock (AllegraEra c))

txsBlockCapacity ∷ Ticked (LedgerState (ShelleyBlock (AllegraEra c))) → TxMeasure (ShelleyBlock (AllegraEra c))

ShelleyBasedEra (ShelleyEra c) ⇒ TxLimits (ShelleyBlock (ShelleyEra c))

Instance details

Associated Types

type TxMeasure (ShelleyBlock (ShelleyEra c))

Methods

txMeasure ∷ Validated (GenTx (ShelleyBlock (ShelleyEra c))) → TxMeasure (ShelleyBlock (ShelleyEra c))

txsBlockCapacity ∷ Ticked (LedgerState (ShelleyBlock (ShelleyEra c))) → TxMeasure (ShelleyBlock (ShelleyEra c))

ShelleyBasedEra (AlonzoEra c) ⇒ TxLimits (ShelleyBlock (AlonzoEra c))

Instance details

Associated Types

type TxMeasure (ShelleyBlock (AlonzoEra c))

Methods

txMeasure ∷ Validated (GenTx (ShelleyBlock (AlonzoEra c))) → TxMeasure (ShelleyBlock (AlonzoEra c))

txsBlockCapacity ∷ Ticked (LedgerState (ShelleyBlock (AlonzoEra c))) → TxMeasure (ShelleyBlock (AlonzoEra c))

ShelleyBasedEra (MaryEra c) ⇒ TxLimits (ShelleyBlock (MaryEra c))

Instance details

Associated Types

type TxMeasure (ShelleyBlock (MaryEra c))

Methods

txMeasure ∷ Validated (GenTx (ShelleyBlock (MaryEra c))) → TxMeasure (ShelleyBlock (MaryEra c))

txsBlockCapacity ∷ Ticked (LedgerState (ShelleyBlock (MaryEra c))) → TxMeasure (ShelleyBlock (MaryEra c))

ShelleyBasedEra era ⇒ LedgerSupportsMempool (ShelleyBlock era)

Instance details

Methods

txInvariant ∷ GenTx (ShelleyBlock era) → Bool

applyTx ∷ LedgerConfig (ShelleyBlock era) → WhetherToIntervene → SlotNo → GenTx (ShelleyBlock era) → TickedLedgerState (ShelleyBlock era) → Except (ApplyTxErr (ShelleyBlock era)) (TickedLedgerState (ShelleyBlock era), Validated (GenTx (ShelleyBlock era)))

reapplyTx ∷ LedgerConfig (ShelleyBlock era) → SlotNo → Validated (GenTx (ShelleyBlock era)) → TickedLedgerState (ShelleyBlock era) → Except (ApplyTxErr (ShelleyBlock era)) (TickedLedgerState (ShelleyBlock era))

txsMaxBytes ∷ TickedLedgerState (ShelleyBlock era) → Word32

txInBlockSize ∷ GenTx (ShelleyBlock era) → Word32

txForgetValidated ∷ Validated (GenTx (ShelleyBlock era)) → GenTx (ShelleyBlock era)

ShelleyBasedEra era ⇒ HasTxId (GenTx (ShelleyBlock era))

Instance details

Methods

txId ∷ GenTx (ShelleyBlock era) → TxId (GenTx (ShelleyBlock era))

ShelleyBasedEra era ⇒ HasTxs (ShelleyBlock era)

Instance details

Methods

extractTxs ∷ ShelleyBlock era → [GenTx (ShelleyBlock era)]

ShelleyBasedEra era ⇒ Condense (GenTx (ShelleyBlock era))

Instance details

Methods

condense ∷ GenTx (ShelleyBlock era) → String

Condense (GenTxId (ShelleyBlock era))

Instance details

Methods

condense ∷ GenTxId (ShelleyBlock era) → String

ShelleyBasedEra era ⇒ NoThunks (GenTx (ShelleyBlock era))

Instance details

Methods

noThunks ∷ Context → GenTx (ShelleyBlock era) → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → GenTx (ShelleyBlock era) → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (GenTx (ShelleyBlock era)) → String

NoThunks (TxId (GenTx (ShelleyBlock era)))

Instance details

Methods

noThunks ∷ Context → TxId (GenTx (ShelleyBlock era)) → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → TxId (GenTx (ShelleyBlock era)) → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (TxId (GenTx (ShelleyBlock era))) → String

ShelleyBasedEra era ⇒ NoThunks (Validated (GenTx (ShelleyBlock era)))

Instance details

Methods

noThunks ∷ Context → Validated (GenTx (ShelleyBlock era)) → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → Validated (GenTx (ShelleyBlock era)) → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (Validated (GenTx (ShelleyBlock era))) → String

ShelleyBasedEra era ⇒ FromCBOR (GenTx (ShelleyBlock era))

Instance details

Methods

fromCBOR ∷ Decoder s (GenTx (ShelleyBlock era))

label ∷ Proxy (GenTx (ShelleyBlock era)) → Text

(PraosCrypto (EraCrypto era), Typeable era) ⇒ FromCBOR (TxId (GenTx (ShelleyBlock era)))

Instance details

Methods

fromCBOR ∷ Decoder s (TxId (GenTx (ShelleyBlock era)))

label ∷ Proxy (TxId (GenTx (ShelleyBlock era))) → Text

ShelleyBasedEra era ⇒ ToCBOR (GenTx (ShelleyBlock era))

Instance details

Methods

toCBOR ∷ GenTx (ShelleyBlock era) → Encoding

encodedSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy (GenTx (ShelleyBlock era)) → Size

encodedListSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy [GenTx (ShelleyBlock era)] → Size

(PraosCrypto (EraCrypto era), Typeable era) ⇒ ToCBOR (TxId (GenTx (ShelleyBlock era)))

Instance details

Methods

toCBOR ∷ TxId (GenTx (ShelleyBlock era)) → Encoding

encodedSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy (TxId (GenTx (ShelleyBlock era))) → Size

encodedListSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy [TxId (GenTx (ShelleyBlock era))] → Size

Typeable era ⇒ ShowProxy (ApplyTxError era ∷ Type)

Instance details

Methods

showProxy ∷ Proxy (ApplyTxError era) → String