Ouroboros.Consensus.Shelley.Ledger.Mempool

Description

Shelley mempool integration

Documentation

class (UpdateLedger blk, NoUnexpectedThunks (GenTx blk)) ⇒ ApplyTx blk where

Minimal complete definition

applyTx, reapplyTx

Associated Types

data GenTx blk ∷ Type

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

type ApplyTxErr blk ∷ Type

Updating the ledger with a single transaction may result in a different error type as when updating it with a block

Methods

txInvariant ∷ GenTx blk → Bool

Check whether the internal invariants of the transaction hold.

applyTx ∷ LedgerConfig blk → GenTx blk → TickedLedgerState blk → Except (ApplyTxErr blk) (TickedLedgerState blk)

Apply transaction we have not previously seen before

reapplyTx ∷ LedgerConfig blk → GenTx blk → TickedLedgerState blk → Except (ApplyTxErr blk) (TickedLedgerState blk)

Re-apply a transaction

When we re-apply a transaction to a potentially different ledger state expensive checks such as cryptographic hashes can be skipped, but other checks (such as checking for double spending) must still be done.

data ApplyTxError crypto

Constructors

ApplyTxError [PredicateFailure (LEDGERS crypto)]

data family GenTx blk ∷ Type

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data family TxId tx ∷ Type

A generalized transaction, GenTx, identifier.

mkShelleyTx ∷ Crypto c ⇒ Tx c → GenTx (ShelleyBlock c)

Orphan instances

Crypto c ⇒ Eq (GenTx (ShelleyBlock c))

Instance details

Methods

(==) ∷ GenTx (ShelleyBlock c) → GenTx (ShelleyBlock c) → Bool

(/=) ∷ GenTx (ShelleyBlock c) → GenTx (ShelleyBlock c) → Bool

Eq (TxId (GenTx (ShelleyBlock c)))

Instance details

Methods

(==) ∷ TxId (GenTx (ShelleyBlock c)) → TxId (GenTx (ShelleyBlock c)) → Bool

(/=) ∷ TxId (GenTx (ShelleyBlock c)) → TxId (GenTx (ShelleyBlock c)) → Bool

Ord (TxId (GenTx (ShelleyBlock c)))

Instance details

Methods

compare ∷ TxId (GenTx (ShelleyBlock c)) → TxId (GenTx (ShelleyBlock c)) → Ordering

(<) ∷ TxId (GenTx (ShelleyBlock c)) → TxId (GenTx (ShelleyBlock c)) → Bool

(<=) ∷ TxId (GenTx (ShelleyBlock c)) → TxId (GenTx (ShelleyBlock c)) → Bool

(>) ∷ TxId (GenTx (ShelleyBlock c)) → TxId (GenTx (ShelleyBlock c)) → Bool

(>=) ∷ TxId (GenTx (ShelleyBlock c)) → TxId (GenTx (ShelleyBlock c)) → Bool

max ∷ TxId (GenTx (ShelleyBlock c)) → TxId (GenTx (ShelleyBlock c)) → TxId (GenTx (ShelleyBlock c))

min ∷ TxId (GenTx (ShelleyBlock c)) → TxId (GenTx (ShelleyBlock c)) → TxId (GenTx (ShelleyBlock c))

Crypto c ⇒ Show (GenTx (ShelleyBlock c))

Instance details

Methods

showsPrec ∷ Int → GenTx (ShelleyBlock c) → ShowS

show ∷ GenTx (ShelleyBlock c) → String

showList ∷ [GenTx (ShelleyBlock c)] → ShowS

Show (GenTxId (ShelleyBlock c))

Instance details

Methods

showsPrec ∷ Int → GenTxId (ShelleyBlock c) → ShowS

show ∷ GenTxId (ShelleyBlock c) → String

showList ∷ [GenTxId (ShelleyBlock c)] → ShowS

Generic (GenTx (ShelleyBlock c))

Instance details

Associated Types

type Rep (GenTx (ShelleyBlock c)) ∷ Type → Type

Methods

from ∷ GenTx (ShelleyBlock c) → Rep (GenTx (ShelleyBlock c)) x

to ∷ Rep (GenTx (ShelleyBlock c)) x → GenTx (ShelleyBlock c)

Crypto c ⇒ ToCBOR (GenTx (ShelleyBlock c))

Instance details

Methods

toCBOR ∷ GenTx (ShelleyBlock c) → Encoding

encodedSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy (GenTx (ShelleyBlock c)) → Size

encodedListSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy [GenTx (ShelleyBlock c)] → Size

Crypto c ⇒ ToCBOR (TxId (GenTx (ShelleyBlock c)))

Instance details

Methods

toCBOR ∷ TxId (GenTx (ShelleyBlock c)) → Encoding

encodedSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy (TxId (GenTx (ShelleyBlock c))) → Size

encodedListSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy [TxId (GenTx (ShelleyBlock c))] → Size

Crypto c ⇒ FromCBOR (GenTx (ShelleyBlock c))

Instance details

Methods

fromCBOR ∷ Decoder s (GenTx (ShelleyBlock c))

label ∷ Proxy (GenTx (ShelleyBlock c)) → Text

Crypto c ⇒ FromCBOR (TxId (GenTx (ShelleyBlock c)))

Instance details

Methods

fromCBOR ∷ Decoder s (TxId (GenTx (ShelleyBlock c)))

label ∷ Proxy (TxId (GenTx (ShelleyBlock c))) → Text

Crypto c ⇒ NoUnexpectedThunks (GenTx (ShelleyBlock c))

Instance details

Methods

noUnexpectedThunks ∷ [String] → GenTx (ShelleyBlock c) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → GenTx (ShelleyBlock c) → IO ThunkInfo

showTypeOf ∷ Proxy (GenTx (ShelleyBlock c)) → String

Typeable c ⇒ NoUnexpectedThunks (TxId (GenTx (ShelleyBlock c)))

Instance details

Methods

noUnexpectedThunks ∷ [String] → TxId (GenTx (ShelleyBlock c)) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → TxId (GenTx (ShelleyBlock c)) → IO ThunkInfo

showTypeOf ∷ Proxy (TxId (GenTx (ShelleyBlock c))) → String

TPraosCrypto c ⇒ ApplyTx (ShelleyBlock c)

Instance details

Associated Types

data GenTx (ShelleyBlock c) ∷ Type

type ApplyTxErr (ShelleyBlock c) ∷ Type

Methods

txInvariant ∷ GenTx (ShelleyBlock c) → Bool

applyTx ∷ LedgerConfig (ShelleyBlock c) → GenTx (ShelleyBlock c) → TickedLedgerState (ShelleyBlock c) → Except (ApplyTxErr (ShelleyBlock c)) (TickedLedgerState (ShelleyBlock c))

reapplyTx ∷ LedgerConfig (ShelleyBlock c) → GenTx (ShelleyBlock c) → TickedLedgerState (ShelleyBlock c) → Except (ApplyTxErr (ShelleyBlock c)) (TickedLedgerState (ShelleyBlock c))

Crypto c ⇒ HasTxId (GenTx (ShelleyBlock c))

Instance details

Associated Types

data TxId (GenTx (ShelleyBlock c)) ∷ Type

Methods

txId ∷ GenTx (ShelleyBlock c) → TxId (GenTx (ShelleyBlock c))

Crypto c ⇒ HasTxs (ShelleyBlock c)

Instance details

Methods

extractTxs ∷ ShelleyBlock c → [GenTx (ShelleyBlock c)]

Crypto c ⇒ Condense (GenTx (ShelleyBlock c))

Instance details

Methods

condense ∷ GenTx (ShelleyBlock c) → String

Condense (GenTxId (ShelleyBlock c))

Instance details

Methods

condense ∷ GenTxId (ShelleyBlock c) → String