Ouroboros.Consensus.Shelley.Ledger.Query

Documentation

newtype NonMyopicMemberRewards c

Constructors

NonMyopicMemberRewards

Fields

unNonMyopicMemberRewards ∷ Map (Either Coin (Credential 'Staking c)) (Map (KeyHash 'StakePool c) Coin)

data family Query blk ∷ Type → Type

Different queries supported by the ledger, indexed by the result type.

querySupportedVersion ∷ Query (ShelleyBlock era) result → ShelleyNodeToClientVersion → Bool

Is the given query supported by the given ShelleyNodeToClientVersion?

Serialisation

decodeShelleyQuery ∷ ShelleyBasedEra era ⇒ Decoder s (SomeSecond Query (ShelleyBlock era))

decodeShelleyResult ∷ ShelleyBasedEra era ⇒ Query (ShelleyBlock era) result → ∀ s. Decoder s result

encodeShelleyQuery ∷ ShelleyBasedEra era ⇒ Query (ShelleyBlock era) result → Encoding

encodeShelleyResult ∷ ShelleyBasedEra era ⇒ Query (ShelleyBlock era) result → result → Encoding

Orphan instances

ShelleyBasedEra era ⇒ QueryLedger (ShelleyBlock era)

Instance details

Methods

answerQuery ∷ ExtLedgerCfg (ShelleyBlock era) → Query (ShelleyBlock era) result → ExtLedgerState (ShelleyBlock era) → result

SameDepIndex (Query (ShelleyBlock era))

Instance details

Methods

sameDepIndex ∷ Query (ShelleyBlock era) a → Query (ShelleyBlock era) b → Maybe (a :~: b)

ShelleyBasedEra era ⇒ ShowQuery (Query (ShelleyBlock era))

Instance details

Methods

showResult ∷ Query (ShelleyBlock era) result → result → String

Eq (Query (ShelleyBlock era) result)

Instance details

Methods

(==) ∷ Query (ShelleyBlock era) result → Query (ShelleyBlock era) result → Bool

(/=) ∷ Query (ShelleyBlock era) result → Query (ShelleyBlock era) result → Bool

Show (Query (ShelleyBlock era) result)

Instance details

Methods

showsPrec ∷ Int → Query (ShelleyBlock era) result → ShowS

show ∷ Query (ShelleyBlock era) result → String

showList ∷ [Query (ShelleyBlock era) result] → ShowS

Typeable era ⇒ ShowProxy (Query (ShelleyBlock era) ∷ Type → Type)

Instance details

Methods

showProxy ∷ Proxy (Query (ShelleyBlock era)) → String