Ouroboros.Consensus.Shelley.Ledger.Query

Documentation

data family BlockQuery blk ∷ Type → Type

Different queries supported by the ledger, indexed by the result type.

newtype NonMyopicMemberRewards c

Constructors

NonMyopicMemberRewards

Fields

unNonMyopicMemberRewards ∷ Map (Either Coin (Credential 'Staking c)) (Map (KeyHash 'StakePool c) Coin)

querySupportedVersion ∷ BlockQuery (ShelleyBlock era) result → ShelleyNodeToClientVersion → Bool

Is the given query supported by the given ShelleyNodeToClientVersion?

Serialisation

decodeShelleyQuery ∷ ShelleyBasedEra era ⇒ Decoder s (SomeSecond BlockQuery (ShelleyBlock era))

decodeShelleyResult ∷ ShelleyBasedEra era ⇒ BlockQuery (ShelleyBlock era) result → ∀ s. Decoder s result

encodeShelleyQuery ∷ ShelleyBasedEra era ⇒ BlockQuery (ShelleyBlock era) result → Encoding

encodeShelleyResult ∷ ShelleyBasedEra era ⇒ BlockQuery (ShelleyBlock era) result → result → Encoding

Orphan instances

ShelleyBasedEra era ⇒ QueryLedger (ShelleyBlock era)

Instance details

Methods

answerBlockQuery ∷ ExtLedgerCfg (ShelleyBlock era) → BlockQuery (ShelleyBlock era) result → ExtLedgerState (ShelleyBlock era) → result

SameDepIndex (BlockQuery (ShelleyBlock era))

Instance details

Methods

sameDepIndex ∷ BlockQuery (ShelleyBlock era) a → BlockQuery (ShelleyBlock era) b → Maybe (a :~: b)

ShelleyBasedEra era ⇒ ShowQuery (BlockQuery (ShelleyBlock era))

Instance details

Methods

showResult ∷ BlockQuery (ShelleyBlock era) result → result → String

Eq (BlockQuery (ShelleyBlock era) result)

Instance details

Methods

(==) ∷ BlockQuery (ShelleyBlock era) result → BlockQuery (ShelleyBlock era) result → Bool

(/=) ∷ BlockQuery (ShelleyBlock era) result → BlockQuery (ShelleyBlock era) result → Bool

Show (BlockQuery (ShelleyBlock era) result)

Instance details

Methods

showsPrec ∷ Int → BlockQuery (ShelleyBlock era) result → ShowS

show ∷ BlockQuery (ShelleyBlock era) result → String

showList ∷ [BlockQuery (ShelleyBlock era) result] → ShowS

Typeable era ⇒ ShowProxy (BlockQuery (ShelleyBlock era) ∷ Type → Type)

Instance details

Methods

showProxy ∷ Proxy (BlockQuery (ShelleyBlock era)) → String