Ouroboros.Consensus.Shelley.Node

Documentation

newtype MaxMajorProtVer

The maximum major protocol version.

Must be at least the current major protocol version. For Cardano mainnet, the Shelley era has major protocol verison 2.

Constructors

MaxMajorProtVer

Fields

getMaxMajorProtVer ∷ Natural

data ProtocolParamsAllegra c

Parameters needed to run Allegra

Constructors

ProtocolParamsAllegra

Fields

allegraProtVer ∷ ProtVer

allegraMaxTxCapacityOverrides ∷ Overrides (ShelleyBlock (AllegraEra c))

data ProtocolParamsAlonzo c

Parameters needed to run Alonzo

Constructors

ProtocolParamsAlonzo

Fields

alonzoProtVer ∷ ProtVer

alonzoMaxTxCapacityOverrides ∷ Overrides (ShelleyBlock (AlonzoEra c))

data ProtocolParamsMary c

Parameters needed to run Mary

Constructors

ProtocolParamsMary

Fields

maryProtVer ∷ ProtVer

maryMaxTxCapacityOverrides ∷ Overrides (ShelleyBlock (MaryEra c))

data ProtocolParamsShelley c

Parameters needed to run Shelley

Constructors

ProtocolParamsShelley

Fields

shelleyProtVer ∷ ProtVer

shelleyMaxTxCapacityOverrides ∷ Overrides (ShelleyBlock (ShelleyEra c))

data ProtocolParamsShelleyBased era

Parameters common to all Shelley-based ledgers.

When running a chain with multiple Shelley-based eras, in addition to the per-era protocol parameters, one value of ProtocolParamsShelleyBased will be needed, which is shared among all Shelley-based eras.

The era parameter determines from which era the genesis config will be used.

Constructors

ProtocolParamsShelleyBased

Fields

shelleyBasedGenesis ∷ ShelleyGenesis era

shelleyBasedInitialNonce ∷ Nonce

The initial nonce, typically derived from the hash of Genesis config JSON file.

WARNING: chains using different values of this parameter will be mutually incompatible.

shelleyBasedLeaderCredentials ∷ [TPraosLeaderCredentials (EraCrypto era)]

data Nonce

Constructors

Nonce !(Hash Blake2b_256 Nonce)

NeutralNonce

data ProtVer

Constructors

ProtVer

Fields

pvMajor ∷ !Natural

pvMinor ∷ !Natural

data ShelleyGenesis era

Constructors

ShelleyGenesis

Fields

sgSystemStart ∷ !UTCTime

sgNetworkMagic ∷ !Word32

sgNetworkId ∷ !Network

sgActiveSlotsCoeff ∷ !PositiveUnitInterval

sgSecurityParam ∷ !Word64

sgEpochLength ∷ !EpochSize

sgSlotsPerKESPeriod ∷ !Word64

sgMaxKESEvolutions ∷ !Word64

sgSlotLength ∷ !NominalDiffTime

sgUpdateQuorum ∷ !Word64

sgMaxLovelaceSupply ∷ !Word64

sgProtocolParams ∷ !(PParams era)

sgGenDelegs ∷ !(Map (KeyHash 'Genesis (Crypto era)) (GenDelegPair (Crypto era)))

sgInitialFunds ∷ !(Map (Addr (Crypto era)) Coin)

sgStaking ∷ !(ShelleyGenesisStaking (Crypto era))

data ShelleyGenesisStaking crypto

Constructors

ShelleyGenesisStaking

Fields

sgsPools ∷ !(Map (KeyHash 'StakePool crypto) (PoolParams crypto))

sgsStake ∷ !(Map (KeyHash 'Staking crypto) (KeyHash 'StakePool crypto))

emptyGenesisStaking ∷ ShelleyGenesisStaking crypto

data TPraosLeaderCredentials c

Constructors

TPraosLeaderCredentials

Fields

tpraosLeaderCredentialsInitSignKey ∷ SignKeyKES c

The unevolved signing KES key (at evolution 0).

Note that this is not inside TPraosCanBeLeader since it gets evolved automatically, whereas TPraosCanBeLeader does not change.

tpraosLeaderCredentialsCanBeLeader ∷ TPraosCanBeLeader c

tpraosLeaderCredentialsLabel ∷ Text

Identifier for this set of credentials.

Useful when the node is running with multiple sets of credentials.

protocolClientInfoShelley ∷ ProtocolClientInfo (ShelleyBlock era)

protocolInfoShelley ∷ ∀ m c. (IOLike m, ShelleyBasedEra (ShelleyEra c), TxLimits (ShelleyBlock (ShelleyEra c))) ⇒ ProtocolParamsShelleyBased (ShelleyEra c) → ProtocolParamsShelley c → ProtocolInfo m (ShelleyBlock (ShelleyEra c))

protocolInfoShelleyBased ∷ ∀ m era. (IOLike m, ShelleyBasedEra era, TxLimits (ShelleyBlock era)) ⇒ ProtocolParamsShelleyBased era → TranslationContext era → ProtVer → Overrides (ShelleyBlock era) → ProtocolInfo m (ShelleyBlock era)

registerGenesisStaking ∷ ∀ era. ShelleyBasedEra era ⇒ ShelleyGenesisStaking (EraCrypto era) → NewEpochState era → NewEpochState era

Register the initial staking information in the NewEpochState.

HERE BE DRAGONS! This function is intended to help in testing.

In production, the genesis should not contain any initial staking.

Any existing staking information is overridden, but the UTxO is left untouched.

TODO adapt and reuse registerGenesisStaking from cardano-ledger-specs.

registerInitialFunds ∷ ∀ era. (ShelleyBasedEra era, HasCallStack) ⇒ Map (Addr (EraCrypto era)) Coin → NewEpochState era → NewEpochState era

Register the initial funds in the NewEpochState.

HERE BE DRAGONS! This function is intended to help in testing.

In production, the genesis should not contain any initial funds.

The given funds are added to the existing UTxO.

PRECONDITION: the given funds must not be part of the existing UTxO. > forall (addr, _) in initialFunds. > Map.notElem (SL.initialFundsPseudoTxIn addr) existingUTxO

PROPERTY: > genesisUTxO genesis > == genesisUTxO' (sgInitialFunds genesis) > == extractUTxO (registerInitialFunds (sgInitialFunds genesis) > NewEpochState)

TODO move to cardano-ledger-specs.

shelleyBlockForging ∷ ∀ m era. (ShelleyBasedEra era, TxLimits (ShelleyBlock era), IOLike m) ⇒ TPraosParams → Overrides (ShelleyBlock era) → TPraosLeaderCredentials (EraCrypto era) → m (BlockForging m (ShelleyBlock era))

Create a BlockForging record for a single era.

In case the same credentials should be shared across multiple Shelley-based eras, use shelleySharedBlockForging.

shelleySharedBlockForging ∷ ∀ m c eras. (PraosCrypto c, All (ShelleyEraWithCrypto c) eras, IOLike m) ⇒ Proxy eras → TPraosParams → TPraosLeaderCredentials c → NP (Overrides :.: ShelleyBlock) eras → m (NP (BlockForging m :.: ShelleyBlock) eras)

Create a BlockForging record for each of the given Shelley-based eras, safely sharing the same set of credentials for all of them.

The name of the era (separated by a _) will be appended to each forgeLabel.

tpraosBlockIssuerVKey ∷ TPraosLeaderCredentials c → VKey 'BlockIssuer c

validateGenesis ∷ ShelleyBasedEra era ⇒ ShelleyGenesis era → Either String ()

Check the validity of the genesis config. To be used in conjunction with assertWithMsg.

Orphan instances

ShelleyBasedEra era ⇒ RunNode (ShelleyBlock era)

Instance details

ShelleyBasedEra era ⇒ NodeInitStorage (ShelleyBlock era)

Instance details

Methods

nodeImmutableDbChunkInfo ∷ StorageConfig (ShelleyBlock era) → ChunkInfo

nodeCheckIntegrity ∷ StorageConfig (ShelleyBlock era) → ShelleyBlock era → Bool

nodeInitChainDB ∷ IOLike m ⇒ StorageConfig (ShelleyBlock era) → InitChainDB m (ShelleyBlock era) → m ()

ConfigSupportsNode (ShelleyBlock era)

Instance details

Methods

getSystemStart ∷ BlockConfig (ShelleyBlock era) → SystemStart

getNetworkMagic ∷ BlockConfig (ShelleyBlock era) → NetworkMagic

ShelleyBasedEra era ⇒ BlockSupportsMetrics (ShelleyBlock era)

Instance details

Methods

isSelfIssued ∷ BlockConfig (ShelleyBlock era) → Header (ShelleyBlock era) → WhetherSelfIssued