Ouroboros.Consensus.Shelley.Node

Documentation

protocolInfoShelley ∷ ∀ c. ShelleyGenesis c → ProtVer → Maybe (TPraosLeaderCredentials c) → ProtocolInfo (ShelleyBlock c)

protocolClientInfoShelley ∷ ProtocolClientInfo (ShelleyBlock c)

data ShelleyGenesis c

Shelley genesis information

Note that this is needed only for a pure Shelley network, hence it being defined here rather than in its own module. In mainnet, Shelley will transition naturally from Byron, and thus will never have its own genesis information.

Constructors

ShelleyGenesis

Fields

sgStartTime ∷ !SystemStart

sgNetworkMagic ∷ !NetworkMagic

sgProtocolMagicId ∷ !ProtocolMagicId

sgActiveSlotsCoeff ∷ !Double

sgSecurityParam ∷ !SecurityParam

sgEpochLength ∷ !EpochSize

sgSlotsPerKESPeriod ∷ !Word64

sgMaxKESEvolutions ∷ !Word64

sgSlotLength ∷ !SlotLength

sgUpdateQuorum ∷ !Word64

sgMaxMajorPV ∷ !Natural

sgMaxLovelaceSupply ∷ !Word64

sgProtocolParams ∷ !PParams

sgGenDelegs ∷ !(Map (KeyHash Genesis c) (KeyHash GenesisDelegate c))

sgInitialFunds ∷ !(Map (Addr c) Coin)

sgStaking ∷ !(ShelleyGenesisStaking c)

initialFundsPseudoTxIn ∷ ∀ c. Addr c → TxIn c

Compute the TxIn of the initial UTxO pseudo-transaction corresponding to the given address in the genesis initial funds.

The Shelley initial UTxO is constructed from the sgInitialFunds which is not a full UTxO but just a map from addresses to coin values.

This gets turned into a UTxO by making a pseudo-transaction for each address, with the 0th output being the coin value. So to spend from the initial UTxO we need this same TxIn to use as an input to the spending transaction.

data ShelleyGenesisStaking c

Genesis Shelley staking configuration.

This allows us to configure some initial stake pools and delegation to them, in order to test Praos in a static configuration, without requiring on-chain registration and delegation.

For simplicity, pools defined in the genesis staking do not pay deposits for their registration.

Constructors

ShelleyGenesisStaking

Fields

sgsPools ∷ !(Map (KeyHash StakePool c) (PoolParams c))

Pools to register

The key in this map is the hash of the public key of the _pool_. This need not correspond to any payment or staking key, but must correspond to the cold key held by TPraosIsCoreNode.

sgsStake ∷ !(Map (KeyHash Staking c) (KeyHash StakePool c))

Stake-holding key hash credentials and the pools to delegate that stake to. We require the raw staking key hash in order to:

Avoid pointer addresses, which would be tricky when there's no slot or transaction to point to.

Avoid script credentials.

data TPraosLeaderCredentials c

Constructors

TPraosLeaderCredentials

Fields

tpraosLeaderCredentialsSignKey ∷ HotKey c

Signing KES key. Note that this is not inside TPraosIsCoreNode since it gets evolved automatically, whereas TPraosIsCoreNode does not change.

tpraosLeaderCredentialsIsCoreNode ∷ TPraosIsCoreNode c

data ProtVer

emptyGenesisStaking ∷ ShelleyGenesisStaking c

Empty genesis staking

Orphan instances

TPraosCrypto c ⇒ RunNode (ShelleyBlock c)

Instance details

Methods

nodeForgeBlock ∷ MonadRandom m ⇒ TopLevelConfig (ShelleyBlock c) → Update m (NodeState (ShelleyBlock c)) → BlockNo → TickedLedgerState (ShelleyBlock c) → [GenTx (ShelleyBlock c)] → IsLeader (BlockProtocol (ShelleyBlock c)) → m (ShelleyBlock c)

nodeBlockMatchesHeader ∷ Header (ShelleyBlock c) → ShelleyBlock c → Bool

nodeBlockFetchSize ∷ Header (ShelleyBlock c) → SizeInBytes

nodeIsEBB ∷ Header (ShelleyBlock c) → Maybe EpochNo

nodeImmDbChunkInfo ∷ TopLevelConfig (ShelleyBlock c) → ChunkInfo

nodeStartTime ∷ TopLevelConfig (ShelleyBlock c) → SystemStart

nodeNetworkMagic ∷ TopLevelConfig (ShelleyBlock c) → NetworkMagic

nodeProtocolMagicId ∷ TopLevelConfig (ShelleyBlock c) → ProtocolMagicId

nodeHashInfo ∷ Proxy (ShelleyBlock c) → HashInfo (HeaderHash (ShelleyBlock c))

nodeMaxBlockSize ∷ LedgerState (ShelleyBlock c) → Word32

nodeBlockEncodingOverhead ∷ LedgerState (ShelleyBlock c) → Word32

nodeCheckIntegrity ∷ TopLevelConfig (ShelleyBlock c) → ShelleyBlock c → Bool

nodeAddHeaderEnvelope ∷ Proxy (ShelleyBlock c) → IsEBB → SizeInBytes → ByteString → ByteString

nodeTxSize ∷ GenTx (ShelleyBlock c) → TxSizeInBytes

nodeInitChainDB ∷ IOLike m ⇒ TopLevelConfig (ShelleyBlock c) → InitChainDB m (ShelleyBlock c) → m ()

nodeExceptionIsFatal ∷ Proxy (ShelleyBlock c) → SomeException → Maybe ExitReason

nodeEncodeBlockWithInfo ∷ CodecConfig (ShelleyBlock c) → ShelleyBlock c → BinaryInfo Encoding

nodeEncodeBlock ∷ CodecConfig (ShelleyBlock c) → ShelleyBlock c → Encoding

nodeEncodeHeader ∷ CodecConfig (ShelleyBlock c) → SerialisationVersion (ShelleyBlock c) → Header (ShelleyBlock c) → Encoding

nodeEncodeWrappedHeader ∷ CodecConfig (ShelleyBlock c) → SerialisationAcrossNetwork (ShelleyBlock c) → Serialised (Header (ShelleyBlock c)) → Encoding

nodeEncodeGenTx ∷ GenTx (ShelleyBlock c) → Encoding

nodeEncodeGenTxId ∷ GenTxId (ShelleyBlock c) → Encoding

nodeEncodeHeaderHash ∷ Proxy (ShelleyBlock c) → HeaderHash (ShelleyBlock c) → Encoding

nodeEncodeLedgerState ∷ TopLevelConfig (ShelleyBlock c) → LedgerState (ShelleyBlock c) → Encoding

nodeEncodeConsensusState ∷ TopLevelConfig (ShelleyBlock c) → ConsensusState (BlockProtocol (ShelleyBlock c)) → Encoding

nodeEncodeApplyTxError ∷ Proxy (ShelleyBlock c) → ApplyTxErr (ShelleyBlock c) → Encoding

nodeEncodeAnnTip ∷ Proxy (ShelleyBlock c) → AnnTip (ShelleyBlock c) → Encoding

nodeEncodeQuery ∷ Query (ShelleyBlock c) result → Encoding

nodeEncodeResult ∷ Query (ShelleyBlock c) result → result → Encoding

nodeDecodeHeader ∷ CodecConfig (ShelleyBlock c) → SerialisationVersion (ShelleyBlock c) → Decoder s (ByteString → Header (ShelleyBlock c))

nodeDecodeWrappedHeader ∷ CodecConfig (ShelleyBlock c) → SerialisationAcrossNetwork (ShelleyBlock c) → Decoder s (Serialised (Header (ShelleyBlock c)))

nodeDecodeBlock ∷ CodecConfig (ShelleyBlock c) → Decoder s (ByteString → ShelleyBlock c)

nodeDecodeGenTx ∷ Decoder s (GenTx (ShelleyBlock c))

nodeDecodeGenTxId ∷ Decoder s (GenTxId (ShelleyBlock c))

nodeDecodeHeaderHash ∷ Proxy (ShelleyBlock c) → Decoder s (HeaderHash (ShelleyBlock c))

nodeDecodeLedgerState ∷ TopLevelConfig (ShelleyBlock c) → Decoder s (LedgerState (ShelleyBlock c))

nodeDecodeConsensusState ∷ TopLevelConfig (ShelleyBlock c) → Decoder s (ConsensusState (BlockProtocol (ShelleyBlock c)))

nodeDecodeApplyTxError ∷ Proxy (ShelleyBlock c) → Decoder s (ApplyTxErr (ShelleyBlock c))

nodeDecodeAnnTip ∷ Proxy (ShelleyBlock c) → Decoder s (AnnTip (ShelleyBlock c))

nodeDecodeQuery ∷ Decoder s (Some (Query (ShelleyBlock c)))

nodeDecodeResult ∷ Query (ShelleyBlock c) result → ∀ s. Decoder s result