Ouroboros.Consensus.Shelley.Node.Serialisation

Orphan instances

Era era ⇒ ReconstructNestedCtxt Header (ShelleyBlock era)

Instance details

Methods

reconstructPrefixLen ∷ proxy (Header (ShelleyBlock era)) → PrefixLen

reconstructNestedCtxt ∷ proxy (Header (ShelleyBlock era)) → ShortByteString → SizeInBytes → SomeBlock (NestedCtxt Header) (ShelleyBlock era)

Era era ⇒ SerialiseNodeToNodeConstraints (ShelleyBlock era)

Instance details

Era era ⇒ SerialiseNodeToClientConstraints (ShelleyBlock era)

Instance details

Era era ⇒ SerialiseDiskConstraints (ShelleyBlock era)

Instance details

Era era ⇒ HasBinaryBlockInfo (ShelleyBlock era)

Instance details

Methods

getBinaryBlockInfo ∷ ShelleyBlock era → BinaryBlockInfo

Era era ⇒ SerialiseNodeToNode (ShelleyBlock era) (GenTxId (ShelleyBlock era))

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → GenTxId (ShelleyBlock era) → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → ∀ s. Decoder s (GenTxId (ShelleyBlock era))

Era era ⇒ SerialiseNodeToNode (ShelleyBlock era) (GenTx (ShelleyBlock era))
The To/FromCBOR instances defined in cardano-ledger-specs use CBOR-in-CBOR to get the annotation.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → GenTx (ShelleyBlock era) → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → ∀ s. Decoder s (GenTx (ShelleyBlock era))

SerialiseNodeToNode (ShelleyBlock era) (SerialisedHeader (ShelleyBlock era))
We use CBOR-in-CBOR

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → SerialisedHeader (ShelleyBlock era) → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → ∀ s. Decoder s (SerialisedHeader (ShelleyBlock era))

Era era ⇒ SerialiseNodeToNode (ShelleyBlock era) (Header (ShelleyBlock era))
CBOR-in-CBOR to be compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → Header (ShelleyBlock era) → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → ∀ s. Decoder s (Header (ShelleyBlock era))

SerialiseNodeToNode (ShelleyBlock era) (Serialised (ShelleyBlock era))
Serialised uses CBOR-in-CBOR by default.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → Serialised (ShelleyBlock era) → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → ∀ s. Decoder s (Serialised (ShelleyBlock era))

Era era ⇒ SerialiseNodeToNode (ShelleyBlock era) (ShelleyBlock era)
CBOR-in-CBOR for the annotation. This also makes it compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → ShelleyBlock era → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → ∀ s. Decoder s (ShelleyBlock era)

Era era ⇒ SerialiseNodeToClient (ShelleyBlock era) (ApplyTxError era)
ApplyTxErr '(ShelleyBlock era)'

Instance details

Methods

encodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → ApplyTxError era → Encoding

decodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → ∀ s. Decoder s (ApplyTxError era)

Era era ⇒ SerialiseNodeToClient (ShelleyBlock era) (GenTx (ShelleyBlock era))
Uses CBOR-in-CBOR in the To/FromCBOR instances to get the annotation.

Instance details

Methods

encodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → GenTx (ShelleyBlock era) → Encoding

decodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → ∀ s. Decoder s (GenTx (ShelleyBlock era))

SerialiseNodeToClient (ShelleyBlock era) (Serialised (ShelleyBlock era))
Serialised uses CBOR-in-CBOR by default.

Instance details

Methods

encodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → Serialised (ShelleyBlock era) → Encoding

decodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → ∀ s. Decoder s (Serialised (ShelleyBlock era))

Era era ⇒ SerialiseNodeToClient (ShelleyBlock era) (ShelleyBlock era)
CBOR-in-CBOR for the annotation. This also makes it compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → ShelleyBlock era → Encoding

decodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → ∀ s. Decoder s (ShelleyBlock era)

Era era ⇒ SerialiseResult (ShelleyBlock era) (Query (ShelleyBlock era))

Instance details

Methods

encodeResult ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → Query (ShelleyBlock era) result → result → Encoding

decodeResult ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → Query (ShelleyBlock era) result → ∀ s. Decoder s result

Era era ⇒ EncodeDisk (ShelleyBlock era) (AnnTip (ShelleyBlock era))

Instance details

Methods

encodeDisk ∷ CodecConfig (ShelleyBlock era) → AnnTip (ShelleyBlock era) → Encoding

Era era ⇒ EncodeDisk (ShelleyBlock era) (TPraosState era)
ChainDepState (BlockProtocol (ShelleyBlock era))

Instance details

Methods

encodeDisk ∷ CodecConfig (ShelleyBlock era) → TPraosState era → Encoding

Era era ⇒ EncodeDisk (ShelleyBlock era) (LedgerState (ShelleyBlock era))

Instance details

Methods

encodeDisk ∷ CodecConfig (ShelleyBlock era) → LedgerState (ShelleyBlock era) → Encoding

Era era ⇒ EncodeDisk (ShelleyBlock era) (Header (ShelleyBlock era))

Instance details

Methods

encodeDisk ∷ CodecConfig (ShelleyBlock era) → Header (ShelleyBlock era) → Encoding

Era era ⇒ EncodeDisk (ShelleyBlock era) (ShelleyBlock era)

Instance details

Methods

encodeDisk ∷ CodecConfig (ShelleyBlock era) → ShelleyBlock era → Encoding

Era era ⇒ DecodeDisk (ShelleyBlock era) (AnnTip (ShelleyBlock era))

Instance details

Methods

decodeDisk ∷ CodecConfig (ShelleyBlock era) → ∀ s. Decoder s (AnnTip (ShelleyBlock era))

Era era ⇒ DecodeDisk (ShelleyBlock era) (TPraosState era)
ChainDepState (BlockProtocol (ShelleyBlock era))

Instance details

Methods

decodeDisk ∷ CodecConfig (ShelleyBlock era) → ∀ s. Decoder s (TPraosState era)

Era era ⇒ DecodeDisk (ShelleyBlock era) (LedgerState (ShelleyBlock era))

Instance details

Methods

decodeDisk ∷ CodecConfig (ShelleyBlock era) → ∀ s. Decoder s (LedgerState (ShelleyBlock era))

Era era ⇒ EncodeDiskDepIx (NestedCtxt Header) (ShelleyBlock era)

Instance details

Methods

encodeDiskDepIx ∷ CodecConfig (ShelleyBlock era) → SomeBlock (NestedCtxt Header) (ShelleyBlock era) → Encoding

Era era ⇒ EncodeDiskDep (NestedCtxt Header) (ShelleyBlock era)

Instance details

Methods

encodeDiskDep ∷ CodecConfig (ShelleyBlock era) → NestedCtxt Header (ShelleyBlock era) a → a → Encoding

Era era ⇒ DecodeDiskDepIx (NestedCtxt Header) (ShelleyBlock era)

Instance details

Methods

decodeDiskDepIx ∷ CodecConfig (ShelleyBlock era) → Decoder s (SomeBlock (NestedCtxt Header) (ShelleyBlock era))

Era era ⇒ DecodeDiskDep (NestedCtxt Header) (ShelleyBlock era)

Instance details

Methods

decodeDiskDep ∷ CodecConfig (ShelleyBlock era) → NestedCtxt Header (ShelleyBlock era) a → ∀ s. Decoder s (ByteString → a)

Era era ⇒ SerialiseNodeToClient (ShelleyBlock era) (SomeBlock Query (ShelleyBlock era))

Instance details

Methods

encodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → SomeBlock Query (ShelleyBlock era) → Encoding

decodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → ∀ s. Decoder s (SomeBlock Query (ShelleyBlock era))

Era era ⇒ DecodeDisk (ShelleyBlock era) (ByteString → Header (ShelleyBlock era))

Instance details

Methods

decodeDisk ∷ CodecConfig (ShelleyBlock era) → ∀ s. Decoder s (ByteString → Header (ShelleyBlock era))

Era era ⇒ DecodeDisk (ShelleyBlock era) (ByteString → ShelleyBlock era)

Instance details

Methods

decodeDisk ∷ CodecConfig (ShelleyBlock era) → ∀ s. Decoder s (ByteString → ShelleyBlock era)