Ouroboros.Consensus.Shelley.Node.Serialisation

Orphan instances

ShelleyBasedEra era ⇒ ReconstructNestedCtxt Header (ShelleyBlock era)

Instance details

Methods

reconstructPrefixLen ∷ proxy (Header (ShelleyBlock era)) → PrefixLen

reconstructNestedCtxt ∷ proxy (Header (ShelleyBlock era)) → ShortByteString → SizeInBytes → SomeSecond (NestedCtxt Header) (ShelleyBlock era)

ShelleyBasedEra era ⇒ SerialiseNodeToNodeConstraints (ShelleyBlock era)

Instance details

Methods

estimateBlockSize ∷ Header (ShelleyBlock era) → SizeInBytes

ShelleyBasedEra era ⇒ SerialiseNodeToClientConstraints (ShelleyBlock era)

Instance details

ShelleyBasedEra era ⇒ SerialiseDiskConstraints (ShelleyBlock era)

Instance details

ShelleyBasedEra era ⇒ HasBinaryBlockInfo (ShelleyBlock era)

Instance details

Methods

getBinaryBlockInfo ∷ ShelleyBlock era → BinaryBlockInfo

ShelleyBasedEra era ⇒ SerialiseNodeToClient (ShelleyBlock era) SlotNo

Instance details

Methods

encodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → SlotNo → Encoding

decodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → ∀ s. Decoder s SlotNo

ShelleyBasedEra era ⇒ SerialiseNodeToNode (ShelleyBlock era) (GenTxId (ShelleyBlock era))

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → GenTxId (ShelleyBlock era) → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → ∀ s. Decoder s (GenTxId (ShelleyBlock era))

ShelleyBasedEra era ⇒ SerialiseNodeToNode (ShelleyBlock era) (GenTx (ShelleyBlock era))
The To/FromCBOR instances defined in cardano-ledger-specs use CBOR-in-CBOR to get the annotation.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → GenTx (ShelleyBlock era) → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → ∀ s. Decoder s (GenTx (ShelleyBlock era))

SerialiseNodeToNode (ShelleyBlock era) (SerialisedHeader (ShelleyBlock era))
We use CBOR-in-CBOR

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → SerialisedHeader (ShelleyBlock era) → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → ∀ s. Decoder s (SerialisedHeader (ShelleyBlock era))

ShelleyBasedEra era ⇒ SerialiseNodeToNode (ShelleyBlock era) (Header (ShelleyBlock era))
CBOR-in-CBOR to be compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → Header (ShelleyBlock era) → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → ∀ s. Decoder s (Header (ShelleyBlock era))

SerialiseNodeToNode (ShelleyBlock era) (Serialised (ShelleyBlock era))
Serialised uses CBOR-in-CBOR by default.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → Serialised (ShelleyBlock era) → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → ∀ s. Decoder s (Serialised (ShelleyBlock era))

ShelleyBasedEra era ⇒ SerialiseNodeToNode (ShelleyBlock era) (ShelleyBlock era)
CBOR-in-CBOR for the annotation. This also makes it compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → ShelleyBlock era → Encoding

decodeNodeToNode ∷ CodecConfig (ShelleyBlock era) → BlockNodeToNodeVersion (ShelleyBlock era) → ∀ s. Decoder s (ShelleyBlock era)

ShelleyBasedEra era ⇒ SerialiseNodeToClient (ShelleyBlock era) (ApplyTxError era)
ApplyTxErr '(ShelleyBlock era)'

Instance details

Methods

encodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → ApplyTxError era → Encoding

decodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → ∀ s. Decoder s (ApplyTxError era)

ShelleyBasedEra era ⇒ SerialiseNodeToClient (ShelleyBlock era) (GenTxId (ShelleyBlock era))

Instance details

Methods

encodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → GenTxId (ShelleyBlock era) → Encoding

decodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → ∀ s. Decoder s (GenTxId (ShelleyBlock era))

ShelleyBasedEra era ⇒ SerialiseNodeToClient (ShelleyBlock era) (GenTx (ShelleyBlock era))
Uses CBOR-in-CBOR in the To/FromCBOR instances to get the annotation.

Instance details

Methods

encodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → GenTx (ShelleyBlock era) → Encoding

decodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → ∀ s. Decoder s (GenTx (ShelleyBlock era))

SerialiseNodeToClient (ShelleyBlock era) (Serialised (ShelleyBlock era))
Serialised uses CBOR-in-CBOR by default.

Instance details

Methods

encodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → Serialised (ShelleyBlock era) → Encoding

decodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → ∀ s. Decoder s (Serialised (ShelleyBlock era))

ShelleyBasedEra era ⇒ SerialiseNodeToClient (ShelleyBlock era) (ShelleyBlock era)
CBOR-in-CBOR for the annotation. This also makes it compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → ShelleyBlock era → Encoding

decodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → ∀ s. Decoder s (ShelleyBlock era)

ShelleyBasedEra era ⇒ SerialiseResult (ShelleyBlock era) (BlockQuery (ShelleyBlock era))

Instance details

Methods

encodeResult ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → BlockQuery (ShelleyBlock era) result → result → Encoding

decodeResult ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → BlockQuery (ShelleyBlock era) result → ∀ s. Decoder s result

ShelleyBasedEra era ⇒ EncodeDisk (ShelleyBlock era) (AnnTip (ShelleyBlock era))

Instance details

Methods

encodeDisk ∷ CodecConfig (ShelleyBlock era) → AnnTip (ShelleyBlock era) → Encoding

(ShelleyBasedEra era, EraCrypto era ~ c) ⇒ EncodeDisk (ShelleyBlock era) (TPraosState c)
ChainDepState (BlockProtocol (ShelleyBlock era))

Instance details

Methods

encodeDisk ∷ CodecConfig (ShelleyBlock era) → TPraosState c → Encoding

ShelleyBasedEra era ⇒ EncodeDisk (ShelleyBlock era) (LedgerState (ShelleyBlock era))

Instance details

Methods

encodeDisk ∷ CodecConfig (ShelleyBlock era) → LedgerState (ShelleyBlock era) → Encoding

ShelleyBasedEra era ⇒ EncodeDisk (ShelleyBlock era) (Header (ShelleyBlock era))

Instance details

Methods

encodeDisk ∷ CodecConfig (ShelleyBlock era) → Header (ShelleyBlock era) → Encoding

ShelleyBasedEra era ⇒ EncodeDisk (ShelleyBlock era) (ShelleyBlock era)

Instance details

Methods

encodeDisk ∷ CodecConfig (ShelleyBlock era) → ShelleyBlock era → Encoding

ShelleyBasedEra era ⇒ DecodeDisk (ShelleyBlock era) (AnnTip (ShelleyBlock era))

Instance details

Methods

decodeDisk ∷ CodecConfig (ShelleyBlock era) → ∀ s. Decoder s (AnnTip (ShelleyBlock era))

(ShelleyBasedEra era, EraCrypto era ~ c) ⇒ DecodeDisk (ShelleyBlock era) (TPraosState c)
ChainDepState (BlockProtocol (ShelleyBlock era))

Instance details

Methods

decodeDisk ∷ CodecConfig (ShelleyBlock era) → ∀ s. Decoder s (TPraosState c)

ShelleyBasedEra era ⇒ DecodeDisk (ShelleyBlock era) (LedgerState (ShelleyBlock era))

Instance details

Methods

decodeDisk ∷ CodecConfig (ShelleyBlock era) → ∀ s. Decoder s (LedgerState (ShelleyBlock era))

ShelleyBasedEra era ⇒ EncodeDiskDepIx (NestedCtxt Header) (ShelleyBlock era)

Instance details

Methods

encodeDiskDepIx ∷ CodecConfig (ShelleyBlock era) → SomeSecond (NestedCtxt Header) (ShelleyBlock era) → Encoding

ShelleyBasedEra era ⇒ EncodeDiskDep (NestedCtxt Header) (ShelleyBlock era)

Instance details

Methods

encodeDiskDep ∷ CodecConfig (ShelleyBlock era) → NestedCtxt Header (ShelleyBlock era) a → a → Encoding

ShelleyBasedEra era ⇒ DecodeDiskDepIx (NestedCtxt Header) (ShelleyBlock era)

Instance details

Methods

decodeDiskDepIx ∷ CodecConfig (ShelleyBlock era) → Decoder s (SomeSecond (NestedCtxt Header) (ShelleyBlock era))

ShelleyBasedEra era ⇒ DecodeDiskDep (NestedCtxt Header) (ShelleyBlock era)

Instance details

Methods

decodeDiskDep ∷ CodecConfig (ShelleyBlock era) → NestedCtxt Header (ShelleyBlock era) a → ∀ s. Decoder s (ByteString → a)

ShelleyBasedEra era ⇒ SerialiseNodeToClient (ShelleyBlock era) (SomeSecond BlockQuery (ShelleyBlock era))

Instance details

Methods

encodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → SomeSecond BlockQuery (ShelleyBlock era) → Encoding

decodeNodeToClient ∷ CodecConfig (ShelleyBlock era) → BlockNodeToClientVersion (ShelleyBlock era) → ∀ s. Decoder s (SomeSecond BlockQuery (ShelleyBlock era))

ShelleyBasedEra era ⇒ DecodeDisk (ShelleyBlock era) (ByteString → Header (ShelleyBlock era))

Instance details

Methods

decodeDisk ∷ CodecConfig (ShelleyBlock era) → ∀ s. Decoder s (ByteString → Header (ShelleyBlock era))

ShelleyBasedEra era ⇒ DecodeDisk (ShelleyBlock era) (ByteString → ShelleyBlock era)

Instance details

Methods

decodeDisk ∷ CodecConfig (ShelleyBlock era) → ∀ s. Decoder s (ByteString → ShelleyBlock era)