Ouroboros.Consensus.Shelley.Protocol

Description

Transitional Praos.

Transitional praos allows for the overlaying of Praos with an overlay schedule determining slots to be produced by BFT

Documentation

data TPraos c

data TPraosChainSelectView c

View of the ledger tip for chain selection.

We order between chains as follows: - By chain length, with longer chains always preferred; _else_ - If the tip of each chain was issued by the same agent, then we prefer the chain whose tip has the highest ocert issue number, if one exists; _else_ - All chains are considered equally preferable

Constructors

ChainSelectView

Fields

csvChainLength ∷ BlockNo

csvIssuer ∷ VKey BlockIssuer c

csvIssueNo ∷ Natural

data TPraosFields c toSign

Constructors

TPraosFields

Fields

tpraosSignature ∷ SignedKES (KES c) toSign

tpraosToSign ∷ toSign

forgeTPraosFields ∷ (TPraosCrypto c, Signable (KES c) toSign) ⇒ HotKey c → IsLeader (TPraos c) → (TPraosToSign c → toSign) → TPraosFields c toSign

data TPraosToSign c

Fields arising from transitional praos execution which must be included in the block signature.

Constructors

TPraosToSign

Fields

tpraosToSignIssuerVK ∷ VerKeyDSIGN (DSIGN c)

Verification key for the issuer of this block.

Note that unlike in Classic/BFT where we have a key for the genesis delegate on whose behalf we are issuing this block, this key corresponds to the stake pool/core node actually forging the block.

tpraosToSignVrfVK ∷ VerKeyVRF (VRF c)

tpraosToSignEta ∷ CertifiedVRF (VRF c) Nonce

Verifiable result containing the updated nonce value.

tpraosToSignLeader ∷ CertifiedVRF (VRF c) UnitInterval

Verifiable proof of the leader value, used to determine whether the node has the right to issue a block in this slot.

We include a value here even for blocks forged under the BFT schedule. It is not required that such a value be verifiable (though by default it will be verifiably correct, but unused.)

tpraosToSignOCert ∷ OCert c

Lightweight delegation certificate mapping the cold (DSIGN) key to the online KES key.

type TPraosValidateView c = BHeader c

Because we are using the executable spec, rather than implementing the protocol directly here, we have a fixed header type rather than an abstraction. So our validate view is fixed to this.

data TPraosParams

TPraos parameters that are node independent

Constructors

TPraosParams

Fields

tpraosEpochInfo ∷ !(EpochInfo Identity)

tpraosSlotsPerKESPeriod ∷ !Word64

See slotsPerKESPeriod.

tpraosLeaderF ∷ !ActiveSlotCoeff

Active slots coefficient. This parameter represents the proportion of slots in which blocks should be issued. This can be interpreted as the probability that a party holding all the stake will be elected as leader for a given slot.

tpraosSecurityParam ∷ !SecurityParam

See securityParameter.

tpraosMaxKESEvo ∷ !Word64

Maximum number of KES iterations, see maxKESEvo.

tpraosQuorum ∷ !Word64

Quorum for update system votes and MIR certificates, see quorum.

tpraosMaxMajorPV ∷ !Natural

All blocks invalid after this protocol version, see maxMajorPV.

tpraosMaxLovelaceSupply ∷ !Word64

Maximum number of lovelace in the system, see maxLovelaceSupply.

data TPraosProof c

Assembled proof that the issuer has the right to issue a block in the selected slot.

Constructors

TPraosProof

Fields

tpraosEta ∷ CertifiedVRF (VRF c) Nonce

tpraosLeader ∷ CertifiedVRF (VRF c) UnitInterval

tpraosIsCoreNode ∷ TPraosIsCoreNode c

data TPraosIsCoreNode c

Constructors

TPraosIsCoreNode

Fields

tpraosIsCoreNodeOpCert ∷ !(OCert c)

Certificate delegating rights from the stake pool cold key (or genesis stakeholder delegate cold key) to the online KES key.

tpraosIsCoreNodeColdVerKey ∷ !(VKey BlockIssuer c)

Stake pool cold key or genesis stakeholder delegate cold key.

tpraosIsCoreNodeSignKeyVRF ∷ !(SignKeyVRF (VRF c))

data TPraosIsCoreNodeOrNot c

Constructors

TPraosIsACoreNode !(TPraosIsCoreNode c)

TPraosIsNotACoreNode

mkShelleyGlobals ∷ TPraosParams → Globals

Crypto

class (HashAlgorithm (HASH c), DSIGNAlgorithm (DSIGN c), KESAlgorithm (KES c), VRFAlgorithm (VRF c), ContextDSIGN (DSIGN c) ~ (), ContextKES (KES c) ~ (), ContextVRF (VRF c) ~ (), Typeable c) ⇒ Crypto c

Minimal complete definition

networkMagicId

class (Crypto c, Typeable (VRF c), Condense (SigKES (KES c)), Signable (DSIGN c) (VerKeyKES (KES c), Natural, KESPeriod), Signable (DSIGN c) (Hash (HASH c) (TxBody c)), Signable (KES c) (BHBody c), Signable (VRF c) Seed, ToCBOR (VerKeyDSIGN (DSIGN c))) ⇒ TPraosCrypto c

data TPraosStandardCrypto

Type instances

data family ConsensusConfig p ∷ Type

Static configuration required to run the consensus protocol

Every method in the ConsensusProtocol class takes the consensus configuration as a parameter, so having this as a data family rather than a type family resolves most ambiguity.

Defined out of the class so that protocols can define this type without having to define the entire protocol at the same time (or indeed in the same module).