Ouroboros.Consensus.Shelley.Protocol.Abstract

Description

Commonality between multiple protocols.

Everything in this module is indexed on the protocol (or the crypto), rather than on the block type. This allows it to be imported in Ouroboros.Consensus.Shelley.Ledger.Block.

Documentation

type family ProtoCrypto proto ∷ Type

class (Eq (EnvelopeCheckError proto), NoThunks (EnvelopeCheckError proto), Show (EnvelopeCheckError proto)) ⇒ ProtocolHeaderSupportsEnvelope proto where

Indicates that the header (determined by the protocol) supports " Envelope " functionality. Envelope functionality refers to the minimal functionality required to construct a chain.

Associated Types

type EnvelopeCheckError proto ∷ Type

Methods

pHeaderHash ∷ ShelleyProtocolHeader proto → ShelleyHash (ProtoCrypto proto)

pHeaderPrevHash ∷ ShelleyProtocolHeader proto → PrevHash (ProtoCrypto proto)

pHeaderBodyHash ∷ ShelleyProtocolHeader proto → Hash (ProtoCrypto proto) EraIndependentBlockBody

pHeaderSlot ∷ ShelleyProtocolHeader proto → SlotNo

pHeaderBlock ∷ ShelleyProtocolHeader proto → BlockNo

pHeaderSize ∷ ShelleyProtocolHeader proto → Natural

pHeaderBlockSize ∷ ShelleyProtocolHeader proto → Natural

envelopeChecks ∷ ConsensusConfig proto → Ticked (LedgerView proto) → ShelleyProtocolHeader proto → Except (EnvelopeCheckError proto) ()

Carry out any protocol-specific envelope checks. For example, this might check things like maximum header size.

class ProtocolHeaderSupportsKES proto where

ProtocolHeaderSupportsKES describes functionality common to protocols using key evolving signature schemes. This includes verifying the header integrity (e.g. validating the KES signature), as well as constructing the header (made specific to KES-using protocols through the need to handle the hot key).

Methods

configSlotsPerKESPeriod ∷ ConsensusConfig proto → Word64

Extract the "slots per KES period" value from the protocol config.

Note that we do not require ConsensusConfig in verifyHeaderIntegrity since that function is also invoked with StorageConfig.

verifyHeaderIntegrity

Arguments

∷ Word64
Slots per KES period

→ ShelleyProtocolHeader proto

→ Bool

Verify that the signature on a header is correct and valid.

mkHeader

Arguments

∷ ∀ crypto m. (Crypto crypto, Monad m, crypto ~ ProtoCrypto proto)

⇒ HotKey crypto m

→ CanBeLeader proto

→ IsLeader proto

→ SlotNo
Slot no

→ BlockNo
Block no

→ PrevHash crypto
Hash of the previous block

→ Hash crypto EraIndependentBlockBody
Hash of the block body to include in the header

→ Int
Size of the block body

→ ProtVer
Protocol version

→ m (ShelleyProtocolHeader proto)

class ProtocolHeaderSupportsLedger proto where

Indicates that the protocol header supports the Shelley ledger. We may need to generalise this if, in the future, the ledger requires different things from the protocol.

Methods

mkHeaderView ∷ ShelleyProtocolHeader proto → BHeaderView (ProtoCrypto proto)

class ProtocolHeaderSupportsProtocol proto where

ProtocolHeaderSupportsProtocol` provides support for the concrete block header to support the ConsensusProtocol itself.

Associated Types

type CannotForgeError proto ∷ Type

Methods

protocolHeaderView ∷ ShelleyProtocolHeader proto → ValidateView proto

pHeaderIssuer ∷ ShelleyProtocolHeader proto → VKey 'BlockIssuer (ProtoCrypto proto)

pHeaderIssueNo ∷ ShelleyProtocolHeader proto → Word64

pTieBreakVRFValue ∷ ShelleyProtocolHeader proto → OutputVRF (VRF (ProtoCrypto proto))

A VRF value in the header, used to choose between otherwise equally preferable chains.

newtype ShelleyHash crypto

Constructors

ShelleyHash

Fields

unShelleyHash ∷ Hash crypto EraIndependentBlockHeader

class (ConsensusProtocol proto, Typeable (ShelleyProtocolHeader proto), ProtocolHeaderSupportsEnvelope proto, ProtocolHeaderSupportsKES proto, ProtocolHeaderSupportsProtocol proto, ProtocolHeaderSupportsLedger proto, Serialise (ChainDepState proto), SignedHeader (ShelleyProtocolHeader proto)) ⇒ ShelleyProtocol proto

type family ShelleyProtocolHeader proto = (sh ∷ Type) | sh → proto

Shelley header, determined by the associated protocol.