Ouroboros.Consensus.Shelley.Protocol.State

Description

Consensus state for Transitional Praos

Documentation

data TPraosState c

Praos consensus state.

The transitional praos implementation itself has a concept of state, given by the STS.State (PRTCL c) type. This state, however, doesn't support rewinding.

currentPRTCLState ∷ HasCallStack ⇒ TPraosState c → State (PRTCL c)

Extract the current state

empty ∷ State (PRTCL c) → TPraosState c

lastSlot ∷ HasCallStack ⇒ TPraosState c → WithOrigin SlotNo

Find the slot for the last state snapshot.

append ∷ State (PRTCL c) → TPraosState c → TPraosState c

Append a new state to the history.

This does not prune anything from the old history - so it's possible after calling this to have a state containing more history than needed.

rewind

Arguments

∷ WithOrigin SlotNo
Slot to rewind to

→ TPraosState c

→ Maybe (TPraosState c)

Rewind the state to the specified slot

The state is rewound to its position at the _end_ of the specified slot (i.e. after any blocks in that slot have been applied)

Callers of this function should ensure that the slot we are requesting to rewind to contains a block which was previously applies. However, we do not technically require this - if a block has not been applied in this slot, then we simply return the state as it was following the last applied block.

prune

Arguments

∷ Int
Size (in terms of number of blocks) to prune the state to.

→ TPraosState c

→ TPraosState c

Prune the state to a given maximum size

size ∷ TPraosState c → Int