Ouroboros.Consensus.Shelley.ShelleyHFC

Description

This module is the Shelley Hard Fork Combinator

Documentation

type ProtocolShelley = HardForkProtocol '[ShelleyBlock StandardShelley]

type ShelleyBlockHFC era = HardForkBlock '[ShelleyBlock era]

Shelley as the single era in the hard fork combinator

data ShelleyPartialLedgerConfig era

Constructors

ShelleyPartialLedgerConfig

Fields

shelleyLedgerConfig ∷ !(ShelleyLedgerConfig era)

We cache the non-partial ledger config containing a dummy EpochInfo that needs to be replaced with the correct one.

We do this to avoid recomputing the ledger config each time completeLedgerConfig is called, as mkShelleyLedgerConfig does some rather expensive computations that shouldn't be repeated too often (e.g., sgActiveSlotCoeff).

shelleyTriggerHardFork ∷ !TriggerHardFork

forecastAcrossShelley

Arguments

∷ ∀ eraFrom eraTo. (EraCrypto eraFrom ~ EraCrypto eraTo, ShelleyBasedEra eraFrom)

⇒ ShelleyLedgerConfig eraFrom

→ ShelleyLedgerConfig eraTo

→ Bound
Transition between the two eras

→ SlotNo
Forecast for this slot

→ LedgerState (ShelleyBlock eraFrom)

→ Except OutsideForecastRange (Ticked (WrapLedgerView (ShelleyBlock eraTo)))

Forecast from a Shelley-based era to the next Shelley-based era.

translateChainDepStateAcrossShelley ∷ ∀ eraFrom eraTo. EraCrypto eraFrom ~ EraCrypto eraTo ⇒ RequiringBoth WrapConsensusConfig (Translate WrapChainDepState) (ShelleyBlock eraFrom) (ShelleyBlock eraTo)

translateLedgerViewAcrossShelley ∷ ∀ eraFrom eraTo. (EraCrypto eraFrom ~ EraCrypto eraTo, ShelleyBasedEra eraFrom) ⇒ RequiringBoth WrapLedgerConfig (TranslateForecast LedgerState WrapLedgerView) (ShelleyBlock eraFrom) (ShelleyBlock eraTo)

Orphan instances

(ShelleyBasedEra era, ShelleyBasedEra (PreviousEra era), EraCrypto (PreviousEra era) ~ EraCrypto era) ⇒ TranslateEra era ShelleyTip

Instance details

Associated Types

type TranslationError era ShelleyTip

Methods

translateEra ∷ TranslationContext era → ShelleyTip (PreviousEra era) → Except (TranslationError era ShelleyTip) (ShelleyTip era)

(ShelleyBasedEra era, TranslateEra era WrapTx) ⇒ TranslateEra era (WrapValidatedGenTx :.: ShelleyBlock)

Instance details

Associated Types

type TranslationError era (WrapValidatedGenTx :.: ShelleyBlock)

Methods

translateEra ∷ TranslationContext era → (WrapValidatedGenTx :.: ShelleyBlock) (PreviousEra era) → Except (TranslationError era (WrapValidatedGenTx :.: ShelleyBlock)) ((WrapValidatedGenTx :.: ShelleyBlock) era)

(ShelleyBasedEra era, TranslateEra era WrapTx) ⇒ TranslateEra era (GenTx :.: ShelleyBlock)

Instance details

Associated Types

type TranslationError era (GenTx :.: ShelleyBlock)

Methods

translateEra ∷ TranslationContext era → (GenTx :.: ShelleyBlock) (PreviousEra era) → Except (TranslationError era (GenTx :.: ShelleyBlock)) ((GenTx :.: ShelleyBlock) era)

(ShelleyBasedEra era, TranslateEra era ShelleyTip, TranslateEra era NewEpochState, TranslationError era NewEpochState ~ Void) ⇒ TranslateEra era (LedgerState :.: ShelleyBlock)

Instance details

Associated Types

type TranslationError era (LedgerState :.: ShelleyBlock)

Methods

translateEra ∷ TranslationContext era → (LedgerState :.: ShelleyBlock) (PreviousEra era) → Except (TranslationError era (LedgerState :.: ShelleyBlock)) ((LedgerState :.: ShelleyBlock) era)

ShelleyBasedEra era ⇒ SerialiseConstraintsHFC (ShelleyBlock era)

Instance details

ShelleyBasedEra era ⇒ NoHardForks (ShelleyBlock era)

Instance details

Methods

getEraParams ∷ TopLevelConfig (ShelleyBlock era) → EraParams

toPartialConsensusConfig ∷ proxy (ShelleyBlock era) → ConsensusConfig (BlockProtocol (ShelleyBlock era)) → PartialConsensusConfig (BlockProtocol (ShelleyBlock era))

toPartialLedgerConfig ∷ proxy (ShelleyBlock era) → LedgerConfig (ShelleyBlock era) → PartialLedgerConfig (ShelleyBlock era)

ShelleyBasedEra era ⇒ SingleEraBlock (ShelleyBlock era)

Instance details

Methods

singleEraTransition ∷ PartialLedgerConfig (ShelleyBlock era) → EraParams → Bound → LedgerState (ShelleyBlock era) → Maybe EpochNo

singleEraInfo ∷ proxy (ShelleyBlock era) → SingleEraInfo (ShelleyBlock era)

PraosCrypto c ⇒ HasPartialConsensusConfig (TPraos c)

Instance details

Associated Types

type PartialConsensusConfig (TPraos c)

Methods

completeConsensusConfig ∷ proxy (TPraos c) → EpochInfo (Except PastHorizonException) → PartialConsensusConfig (TPraos c) → ConsensusConfig (TPraos c)

ShelleyBasedEra era ⇒ HasPartialLedgerConfig (ShelleyBlock era)

Instance details

Associated Types

type PartialLedgerConfig (ShelleyBlock era)

Methods

completeLedgerConfig ∷ proxy (ShelleyBlock era) → EpochInfo (Except PastHorizonException) → PartialLedgerConfig (ShelleyBlock era) → LedgerConfig (ShelleyBlock era)

ShelleyBasedEra era ⇒ SupportedNetworkProtocolVersion (ShelleyBlockHFC era)
Forward to the ShelleyBlock instance. Only supports HardForkNodeToNodeDisabled, which is compatible with nodes running with ShelleyBlock.

Instance details

Methods

supportedNodeToNodeVersions ∷ Proxy (ShelleyBlockHFC era) → Map NodeToNodeVersion (BlockNodeToNodeVersion (ShelleyBlockHFC era))

supportedNodeToClientVersions ∷ Proxy (ShelleyBlockHFC era) → Map NodeToClientVersion (BlockNodeToClientVersion (ShelleyBlockHFC era))

latestReleasedNodeVersion ∷ Proxy (ShelleyBlockHFC era) → (Maybe NodeToNodeVersion, Maybe NodeToClientVersion)

ShelleyBasedEra era ⇒ SerialiseHFC '[ShelleyBlock era]
Use the default implementations. This means the serialisation of blocks includes an era wrapper. Each block should do this from the start to be prepared for future hard forks without having to do any bit twiddling.

Instance details

Methods

encodeDiskHfcBlock ∷ CodecConfig (HardForkBlock '[ShelleyBlock era]) → HardForkBlock '[ShelleyBlock era] → Encoding

decodeDiskHfcBlock ∷ CodecConfig (HardForkBlock '[ShelleyBlock era]) → ∀ s. Decoder s (ByteString → HardForkBlock '[ShelleyBlock era])

reconstructHfcPrefixLen ∷ proxy (Header (HardForkBlock '[ShelleyBlock era])) → PrefixLen

reconstructHfcNestedCtxt ∷ proxy (Header (HardForkBlock '[ShelleyBlock era])) → ShortByteString → SizeInBytes → SomeSecond (NestedCtxt Header) (HardForkBlock '[ShelleyBlock era])

getHfcBinaryBlockInfo ∷ HardForkBlock '[ShelleyBlock era] → BinaryBlockInfo

estimateHfcBlockSize ∷ Header (HardForkBlock '[ShelleyBlock era]) → SizeInBytes